Struct aws_sdk_s3::client::fluent_builders::CreateMultipartUpload

pub struct CreateMultipartUpload { /* private fields */ }

Fluent builder constructing a request to CreateMultipartUpload.

This action initiates a multipart upload and returns an upload ID. This upload ID is used to associate all of the parts in the specific multipart upload. You specify this upload ID in each of your subsequent upload part requests (see UploadPart). You also include this upload ID in the final request to either complete or abort the multipart upload request.

For more information about multipart uploads, see Multipart Upload Overview.

If you have configured a lifecycle rule to abort incomplete multipart uploads, the upload must complete within the number of days specified in the bucket lifecycle configuration. Otherwise, the incomplete multipart upload becomes eligible for an abort action and Amazon S3 aborts the multipart upload. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

For information about the permissions required to use the multipart upload API, see Multipart Upload and Permissions.

For request signing, multipart upload is just a series of regular requests. You initiate a multipart upload, send one or more requests to upload parts, and then complete the multipart upload process. You sign each request individually. There is nothing special about signing multipart upload requests. For more information about signing, see Authenticating Requests (Amazon Web Services Signature Version 4).

After you initiate a multipart upload and upload one or more parts, to stop being charged for storing the uploaded parts, you must either complete or abort the multipart upload. Amazon S3 frees up the space used to store the parts and stop charging you for storing them only after you either complete or abort a multipart upload.

You can optionally request server-side encryption. For server-side encryption, Amazon S3 encrypts your data as it writes it to disks in its data centers and decrypts it when you access it. You can provide your own encryption key, or use Amazon Web Services KMS keys or Amazon S3-managed encryption keys. If you choose to provide your own encryption key, the request headers you provide in UploadPart and UploadPartCopy requests must match the headers you used in the request to initiate the upload by using CreateMultipartUpload.

To perform a multipart upload with encryption using an Amazon Web Services KMS key, the requester must have permission to the kms:Decrypt and kms:GenerateDataKey* actions on the key. These permissions are required because Amazon S3 must decrypt and read data from the encrypted file parts before it completes the multipart upload. For more information, see Multipart upload API and permissions in the Amazon S3 User Guide.

If your Identity and Access Management (IAM) user or role is in the same Amazon Web Services account as the KMS key, then you must have these permissions on the key policy. If your IAM user or role belongs to a different account than the key, then you must have the permissions on both the key policy and your IAM user or role.

For more information, see Protecting Data Using Server-Side Encryption.

Access Permissions

When copying an object, you can optionally specify the accounts or groups that should be granted specific permissions on the new object. There are two ways to grant the permissions using the request headers:

Specify a canned ACL with the x-amz-acl request header. For more information, see Canned ACL.

Specify access permissions explicitly with the x-amz-grant-read, x-amz-grant-read-acp, x-amz-grant-write-acp, and x-amz-grant-full-control headers. These parameters map to the set of permissions that Amazon S3 supports in an ACL. For more information, see Access Control List (ACL) Overview.

You can use either a canned ACL or specify access permissions explicitly. You cannot do both.

Server-Side- Encryption-Specific Request Headers

You can optionally tell Amazon S3 to encrypt data at rest using server-side encryption. Server-side encryption is for data encryption at rest. Amazon S3 encrypts your data as it writes it to disks in its data centers and decrypts it when you access it. The option you use depends on whether you want to use Amazon Web Services managed encryption keys or provide your own encryption key.

Use encryption keys managed by Amazon S3 or customer managed key stored in Amazon Web Services Key Management Service (Amazon Web Services KMS) – If you want Amazon Web Services to manage the keys used to encrypt data, specify the following headers in the request.

x-amz-server-side-encryption

x-amz-server-side-encryption-aws-kms-key-id

x-amz-server-side-encryption-context

If you specify x-amz-server-side-encryption:aws:kms, but don't provide x-amz-server-side-encryption-aws-kms-key-id, Amazon S3 uses the Amazon Web Services managed key in Amazon Web Services KMS to protect the data.

All GET and PUT requests for an object protected by Amazon Web Services KMS fail if you don't make them with SSL or by using SigV4.

For more information about server-side encryption with KMS key (SSE-KMS), see Protecting Data Using Server-Side Encryption with KMS keys.

Use customer-provided encryption keys – If you want to manage your own encryption keys, provide all the following headers in the request.

x-amz-server-side-encryption-customer-algorithm

x-amz-server-side-encryption-customer-key

x-amz-server-side-encryption-customer-key-MD5

For more information about server-side encryption with KMS keys (SSE-KMS), see Protecting Data Using Server-Side Encryption with KMS keys.

Access-Control-List (ACL)-Specific Request Headers

You also can use the following access control–related headers with this operation. By default, all objects are private. Only the owner has full access control. When adding a new object, you can grant permissions to individual Amazon Web Services accounts or to predefined groups defined by Amazon S3. These permissions are then added to the access control list (ACL) on the object. For more information, see Using ACLs. With this operation, you can grant access permissions using one of the following two methods:

Specify a canned ACL (x-amz-acl) — Amazon S3 supports a set of predefined ACLs, known as canned ACLs. Each canned ACL has a predefined set of grantees and permissions. For more information, see Canned ACL.

Specify access permissions explicitly — To explicitly grant access permissions to specific Amazon Web Services accounts or groups, use the following headers. Each header maps to specific permissions that Amazon S3 supports in an ACL. For more information, see Access Control List (ACL) Overview. In the header, you specify a list of grantees who get the specific permission. To grant permissions explicitly, use:

x-amz-grant-read

x-amz-grant-write

x-amz-grant-read-acp

x-amz-grant-write-acp

x-amz-grant-full-control

You specify each grantee as a type=value pair, where the type is one of the following:

id – if the value specified is the canonical user ID of an Amazon Web Services account

uri – if you are granting permissions to a predefined group

emailAddress – if the value specified is the email address of an Amazon Web Services account

Using email addresses to specify a grantee is only supported in the following Amazon Web Services Regions:

US East (N. Virginia)

US West (N. California)

US West (Oregon)

Asia Pacific (Singapore)

Asia Pacific (Sydney)

Asia Pacific (Tokyo)

Europe (Ireland)

South America (São Paulo)

For a list of all the Amazon S3 supported Regions and endpoints, see Regions and Endpoints in the Amazon Web Services General Reference.

For example, the following x-amz-grant-read header grants the Amazon Web Services accounts identified by account IDs permissions to read object data and its metadata:

x-amz-grant-read: id="11112222333", id="444455556666"

The following operations are related to CreateMultipartUpload:

UploadPart

CompleteMultipartUpload

AbortMultipartUpload

ListParts

ListMultipartUploads

Implementations

impl CreateMultipartUpload

pub async fn send( self ) -> Result<CreateMultipartUploadOutput, SdkError<CreateMultipartUploadError>>

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn acl(self, input: ObjectCannedAcl) -> Self

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

pub fn set_acl(self, input: Option<ObjectCannedAcl>) -> Self

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

pub fn bucket(self, input: impl Into<String>) -> Self

The name of the bucket to which to initiate the upload

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_bucket(self, input: Option<String>) -> Self

The name of the bucket to which to initiate the upload

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn cache_control(self, input: impl Into<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn set_cache_control(self, input: Option<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(self, input: impl Into<String>) -> Self

Specifies presentational information for the object.

pub fn set_content_disposition(self, input: Option<String>) -> Self

Specifies presentational information for the object.

pub fn content_encoding(self, input: impl Into<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn set_content_encoding(self, input: Option<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(self, input: impl Into<String>) -> Self

The language the content is in.

pub fn set_content_language(self, input: Option<String>) -> Self

The language the content is in.

pub fn content_type(self, input: impl Into<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn set_content_type(self, input: Option<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn expires(self, input: DateTime) -> Self

The date and time at which the object is no longer cacheable.

pub fn set_expires(self, input: Option<DateTime>) -> Self

The date and time at which the object is no longer cacheable.

pub fn grant_full_control(self, input: impl Into<String>) -> Self

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_full_control(self, input: Option<String>) -> Self

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

pub fn grant_read(self, input: impl Into<String>) -> Self

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_read(self, input: Option<String>) -> Self

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

pub fn grant_read_acp(self, input: impl Into<String>) -> Self

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_read_acp(self, input: Option<String>) -> Self

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

pub fn grant_write_acp(self, input: impl Into<String>) -> Self

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_write_acp(self, input: Option<String>) -> Self

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

pub fn key(self, input: impl Into<String>) -> Self

Object key for which the multipart upload is to be initiated.

pub fn set_key(self, input: Option<String>) -> Self

Object key for which the multipart upload is to be initiated.

pub fn metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Metadata.

To override the contents of this collection use set_metadata.

A map of metadata to store with the object in S3.

pub fn set_metadata(self, input: Option<HashMap<String, String>>) -> Self

A map of metadata to store with the object in S3.

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn set_server_side_encryption( self, input: Option<ServerSideEncryption> ) -> Self

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn storage_class(self, input: StorageClass) -> Self

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Self

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

pub fn website_redirect_location(self, input: impl Into<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

pub fn set_website_redirect_location(self, input: Option<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

Specifies the algorithm to use to when encrypting the object (for example, AES256).

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

Specifies the algorithm to use to when encrypting the object (for example, AES256).

pub fn sse_customer_key(self, input: impl Into<String>) -> Self

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

pub fn set_sse_customer_key(self, input: Option<String>) -> Self

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

Specifies the ID of the symmetric customer managed key to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

Specifies the ID of the symmetric customer managed key to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

pub fn ssekms_encryption_context(self, input: impl Into<String>) -> Self

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

pub fn set_ssekms_encryption_context(self, input: Option<String>) -> Self

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

pub fn bucket_key_enabled(self, input: bool) -> Self

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with an object action doesn’t affect bucket-level settings for S3 Bucket Key.

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with an object action doesn’t affect bucket-level settings for S3 Bucket Key.

pub fn request_payer(self, input: RequestPayer) -> Self

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn set_request_payer(self, input: Option<RequestPayer>) -> Self

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn tagging(self, input: impl Into<String>) -> Self

The tag-set for the object. The tag-set must be encoded as URL Query parameters.

pub fn set_tagging(self, input: Option<String>) -> Self

The tag-set for the object. The tag-set must be encoded as URL Query parameters.

pub fn object_lock_mode(self, input: ObjectLockMode) -> Self

Specifies the Object Lock mode that you want to apply to the uploaded object.

pub fn set_object_lock_mode(self, input: Option<ObjectLockMode>) -> Self

Specifies the Object Lock mode that you want to apply to the uploaded object.

pub fn object_lock_retain_until_date(self, input: DateTime) -> Self

Specifies the date and time when you want the Object Lock to expire.

pub fn set_object_lock_retain_until_date(self, input: Option<DateTime>) -> Self

Specifies the date and time when you want the Object Lock to expire.

pub fn object_lock_legal_hold_status( self, input: ObjectLockLegalHoldStatus ) -> Self

Specifies whether you want to apply a Legal Hold to the uploaded object.

pub fn set_object_lock_legal_hold_status( self, input: Option<ObjectLockLegalHoldStatus> ) -> Self

Specifies whether you want to apply a Legal Hold to the uploaded object.

pub fn expected_bucket_owner(self, input: impl Into<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, input: Option<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.