Struct aws_sdk_s3::client::fluent_builders::ListObjects

pub struct ListObjects<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> ListObjects<C>

pub async fn send(self) -> Result<ListObjectsOutput, SdkError<ListObjectsError>> where
C: SmithyConnector,

pub fn bucket(self, input: impl Into<String>) -> Self

The name of the bucket containing the objects.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the AWS SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the AWS SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_bucket(self, input: Option<String>) -> Self

pub fn delimiter(self, input: impl Into<String>) -> Self

A delimiter is a character you use to group keys.

pub fn set_delimiter(self, input: Option<String>) -> Self

pub fn encoding_type(self, input: EncodingType) -> Self

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

pub fn set_encoding_type(self, input: Option<EncodingType>) -> Self

pub fn marker(self, input: impl Into<String>) -> Self

Specifies the key to start with when listing objects in a bucket.

pub fn set_marker(self, input: Option<String>) -> Self

pub fn max_keys(self, input: i32) -> Self

Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

pub fn set_max_keys(self, input: Option<i32>) -> Self

pub fn prefix(self, input: impl Into<String>) -> Self

Limits the response to keys that begin with the specified prefix.

pub fn set_prefix(self, input: Option<String>) -> Self

pub fn request_payer(self, input: RequestPayer) -> Self

Confirms that the requester knows that she or he will be charged for the list objects request. Bucket owners need not specify this parameter in their requests.

pub fn set_request_payer(self, input: Option<RequestPayer>) -> Self

pub fn expected_bucket_owner(self, input: impl Into<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, input: Option<String>) -> Self