Struct aws_sdk_s3::client::fluent_builders::ListObjectsV2

pub struct ListObjectsV2 { /* fields omitted */ }

Implementations

impl ListObjectsV2

pub async fn send(
self
) -> Result<ListObjectsV2Output, SdkError<ListObjectsV2Error>>

pub fn bucket(self, inp: impl Into<String>) -> Self

Bucket name to list.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the AWS SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the AWS SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_bucket(self, inp: String) -> Self

pub fn delimiter(self, inp: impl Into<String>) -> Self

A delimiter is a character you use to group keys.

pub fn set_delimiter(self, inp: Option<String>) -> Self

pub fn encoding_type(self, inp: EncodingType) -> Self

Encoding type used by Amazon S3 to encode object keys in the response.

pub fn set_encoding_type(self, inp: Option<EncodingType>) -> Self

pub fn max_keys(self, inp: i32) -> Self

Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

pub fn set_max_keys(self, inp: i32) -> Self

pub fn prefix(self, inp: impl Into<String>) -> Self

Limits the response to keys that begin with the specified prefix.

pub fn set_prefix(self, inp: Option<String>) -> Self

pub fn continuation_token(self, inp: impl Into<String>) -> Self

ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key.

pub fn set_continuation_token(self, inp: Option<String>) -> Self

pub fn fetch_owner(self, inp: bool) -> Self

The owner field is not present in listV2 by default, if you want to return owner field with each key in the result then set the fetch owner field to true.

pub fn set_fetch_owner(self, inp: bool) -> Self

pub fn start_after(self, inp: impl Into<String>) -> Self

StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket.

pub fn set_start_after(self, inp: Option<String>) -> Self

pub fn request_payer(self, inp: RequestPayer) -> Self

Confirms that the requester knows that she or he will be charged for the list objects request in V2 style. Bucket owners need not specify this parameter in their requests.

pub fn set_request_payer(self, inp: Option<RequestPayer>) -> Self

pub fn expected_bucket_owner(self, inp: impl Into<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, inp: Option<String>) -> Self