Struct aws_sdk_s3::client::fluent_builders::PutObjectRetention

pub struct PutObjectRetention { /* private fields */ }

Fluent builder constructing a request to PutObjectRetention.

Places an Object Retention configuration on an object. For more information, see Locking Objects. Users or accounts require the s3:PutObjectRetention permission in order to place an Object Retention configuration on objects. Bypassing a Governance Retention configuration requires the s3:BypassGovernanceRetention permission.

This action is not supported by Amazon S3 on Outposts.

Permissions

When the Object Lock retention mode is set to compliance, you need s3:PutObjectRetention and s3:BypassGovernanceRetention permissions. For other requests to PutObjectRetention, only s3:PutObjectRetention permissions are required.

Implementations

impl PutObjectRetention

pub async fn send( self ) -> Result<PutObjectRetentionOutput, SdkError<PutObjectRetentionError>>

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn bucket(self, input: impl Into<String>) -> Self

The bucket name that contains the object you want to apply this Object Retention configuration to.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

pub fn set_bucket(self, input: Option<String>) -> Self

The bucket name that contains the object you want to apply this Object Retention configuration to.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

pub fn key(self, input: impl Into<String>) -> Self

The key name for the object that you want to apply this Object Retention configuration to.

pub fn set_key(self, input: Option<String>) -> Self

The key name for the object that you want to apply this Object Retention configuration to.

pub fn retention(self, input: ObjectLockRetention) -> Self

The container element for the Object Retention configuration.

pub fn set_retention(self, input: Option<ObjectLockRetention>) -> Self

The container element for the Object Retention configuration.

pub fn request_payer(self, input: RequestPayer) -> Self

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn set_request_payer(self, input: Option<RequestPayer>) -> Self

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn version_id(self, input: impl Into<String>) -> Self

The version ID for the object that you want to apply this Object Retention configuration to.

pub fn set_version_id(self, input: Option<String>) -> Self

The version ID for the object that you want to apply this Object Retention configuration to.

pub fn bypass_governance_retention(self, input: bool) -> Self

Indicates whether this action should bypass Governance-mode restrictions.

pub fn set_bypass_governance_retention(self, input: Option<bool>) -> Self

Indicates whether this action should bypass Governance-mode restrictions.

pub fn content_md5(self, input: impl Into<String>) -> Self

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

pub fn set_content_md5(self, input: Option<String>) -> Self

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

pub fn expected_bucket_owner(self, input: impl Into<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, input: Option<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.