Struct aws_sdk_s3::client::fluent_builders::UploadPartCopy

pub struct UploadPartCopy<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> UploadPartCopy<C>

pub async fn send(
self
) -> Result<UploadPartCopyOutput, SdkError<UploadPartCopyError>> where
C: SmithyConnector,

pub fn bucket(self, input: impl Into<String>) -> Self

The bucket name.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the AWS SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the AWS SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_bucket(self, input: Option<String>) -> Self

pub fn copy_source(self, input: impl Into<String>) -> Self

Specifies the source object for the copy operation. You specify the value in one of two formats, depending on whether you want to access the source object through an access point:

For objects not accessed through an access point, specify the name of the source bucket and key of the source object, separated by a slash (/). For example, to copy the object reports/january.pdf from the bucket awsexamplebucket, use awsexamplebucket/reports/january.pdf. The value must be URL encoded.

For objects accessed through access points, specify the Amazon Resource Name (ARN) of the object as accessed through the access point, in the format arn:aws:s3:::accesspoint//object/. For example, to copy the object reports/january.pdf through access point my-access-point owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3:us-west-2:123456789012:accesspoint/my-access-point/object/reports/january.pdf. The value must be URL encoded.

Amazon S3 supports copy operations using access points only when the source and destination buckets are in the same AWS Region.

Alternatively, for objects accessed through Amazon S3 on Outposts, specify the ARN of the object as accessed in the format arn:aws:s3-outposts:::outpost//object/. For example, to copy the object reports/january.pdf through outpost my-outpost owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/object/reports/january.pdf. The value must be URL encoded.

To copy a specific version of an object, append ?versionId= to the value (for example, awsexamplebucket/reports/january.pdf?versionId=QUpfdndhfd8438MNFDN93jdnJFkdmqnh893). If you don't specify a version ID, Amazon S3 copies the latest version of the source object.

pub fn set_copy_source(self, input: Option<String>) -> Self

pub fn copy_source_if_match(self, input: impl Into<String>) -> Self

Copies the object if its entity tag (ETag) matches the specified tag.

pub fn set_copy_source_if_match(self, input: Option<String>) -> Self

pub fn copy_source_if_modified_since(self, input: Instant) -> Self

Copies the object if it has been modified since the specified time.

pub fn set_copy_source_if_modified_since(self, input: Option<Instant>) -> Self

pub fn copy_source_if_none_match(self, input: impl Into<String>) -> Self

Copies the object if its entity tag (ETag) is different than the specified ETag.

pub fn set_copy_source_if_none_match(self, input: Option<String>) -> Self

pub fn copy_source_if_unmodified_since(self, input: Instant) -> Self

Copies the object if it hasn't been modified since the specified time.

pub fn set_copy_source_if_unmodified_since(self, input: Option<Instant>) -> Self

pub fn copy_source_range(self, input: impl Into<String>) -> Self

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first 10 bytes of the source. You can copy a range only if the source object is greater than 5 MB.

pub fn set_copy_source_range(self, input: Option<String>) -> Self

pub fn key(self, input: impl Into<String>) -> Self

Object key for which the multipart upload was initiated.

pub fn set_key(self, input: Option<String>) -> Self

pub fn part_number(self, input: i32) -> Self

Part number of part being copied. This is a positive integer between 1 and 10,000.

pub fn set_part_number(self, input: Option<i32>) -> Self

pub fn upload_id(self, input: impl Into<String>) -> Self

Upload ID identifying the multipart upload whose part is being copied.

pub fn set_upload_id(self, input: Option<String>) -> Self

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

Specifies the algorithm to use to when encrypting the object (for example, AES256).

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

pub fn sse_customer_key(self, input: impl Into<String>) -> Self

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

pub fn set_sse_customer_key(self, input: Option<String>) -> Self

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

pub fn copy_source_sse_customer_algorithm(
self,
input: impl Into<String>
) -> Self

Specifies the algorithm to use when decrypting the source object (for example, AES256).

pub fn set_copy_source_sse_customer_algorithm(
self,
input: Option<String>
) -> Self

pub fn copy_source_sse_customer_key(self, input: impl Into<String>) -> Self

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

pub fn set_copy_source_sse_customer_key(self, input: Option<String>) -> Self

pub fn copy_source_sse_customer_key_md5(self, input: impl Into<String>) -> Self

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn set_copy_source_sse_customer_key_md5(self, input: Option<String>) -> Self

pub fn request_payer(self, input: RequestPayer) -> Self

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn set_request_payer(self, input: Option<RequestPayer>) -> Self

pub fn expected_bucket_owner(self, input: impl Into<String>) -> Self

The account ID of the expected destination bucket owner. If the destination bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, input: Option<String>) -> Self

pub fn expected_source_bucket_owner(self, input: impl Into<String>) -> Self

The account ID of the expected source bucket owner. If the source bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_source_bucket_owner(self, input: Option<String>) -> Self