Struct aws_sdk_s3::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonS3.

This client allows ergonomic access to a AmazonS3-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn abort_multipart_upload(&self) -> AbortMultipartUpload<C, M, R>

pub fn complete_multipart_upload(&self) -> CompleteMultipartUpload<C, M, R>

pub fn copy_object(&self) -> CopyObject<C, M, R>

pub fn create_bucket(&self) -> CreateBucket<C, M, R>

pub fn create_multipart_upload(&self) -> CreateMultipartUpload<C, M, R>

pub fn delete_bucket(&self) -> DeleteBucket<C, M, R>

pub fn delete_bucket_analytics_configuration(
&self
) -> DeleteBucketAnalyticsConfiguration<C, M, R>

pub fn delete_bucket_cors(&self) -> DeleteBucketCors<C, M, R>

pub fn delete_bucket_encryption(&self) -> DeleteBucketEncryption<C, M, R>

pub fn delete_bucket_intelligent_tiering_configuration(
&self
) -> DeleteBucketIntelligentTieringConfiguration<C, M, R>

pub fn delete_bucket_inventory_configuration(
&self
) -> DeleteBucketInventoryConfiguration<C, M, R>

pub fn delete_bucket_lifecycle(&self) -> DeleteBucketLifecycle<C, M, R>

pub fn delete_bucket_metrics_configuration(
&self
) -> DeleteBucketMetricsConfiguration<C, M, R>

pub fn delete_bucket_ownership_controls(
&self
) -> DeleteBucketOwnershipControls<C, M, R>

pub fn delete_bucket_policy(&self) -> DeleteBucketPolicy<C, M, R>

pub fn delete_bucket_replication(&self) -> DeleteBucketReplication<C, M, R>

pub fn delete_bucket_tagging(&self) -> DeleteBucketTagging<C, M, R>

pub fn delete_bucket_website(&self) -> DeleteBucketWebsite<C, M, R>

pub fn delete_object(&self) -> DeleteObject<C, M, R>

pub fn delete_objects(&self) -> DeleteObjects<C, M, R>

pub fn delete_object_tagging(&self) -> DeleteObjectTagging<C, M, R>

pub fn delete_public_access_block(&self) -> DeletePublicAccessBlock<C, M, R>

pub fn get_bucket_accelerate_configuration(
&self
) -> GetBucketAccelerateConfiguration<C, M, R>

pub fn get_bucket_acl(&self) -> GetBucketAcl<C, M, R>

pub fn get_bucket_analytics_configuration(
&self
) -> GetBucketAnalyticsConfiguration<C, M, R>

pub fn get_bucket_cors(&self) -> GetBucketCors<C, M, R>

pub fn get_bucket_encryption(&self) -> GetBucketEncryption<C, M, R>

pub fn get_bucket_intelligent_tiering_configuration(
&self
) -> GetBucketIntelligentTieringConfiguration<C, M, R>

pub fn get_bucket_inventory_configuration(
&self
) -> GetBucketInventoryConfiguration<C, M, R>

pub fn get_bucket_lifecycle_configuration(
&self
) -> GetBucketLifecycleConfiguration<C, M, R>

pub fn get_bucket_location(&self) -> GetBucketLocation<C, M, R>

pub fn get_bucket_logging(&self) -> GetBucketLogging<C, M, R>

pub fn get_bucket_metrics_configuration(
&self
) -> GetBucketMetricsConfiguration<C, M, R>

pub fn get_bucket_notification_configuration(
&self
) -> GetBucketNotificationConfiguration<C, M, R>

pub fn get_bucket_ownership_controls(
&self
) -> GetBucketOwnershipControls<C, M, R>

pub fn get_bucket_policy(&self) -> GetBucketPolicy<C, M, R>

pub fn get_bucket_policy_status(&self) -> GetBucketPolicyStatus<C, M, R>

pub fn get_bucket_replication(&self) -> GetBucketReplication<C, M, R>

pub fn get_bucket_request_payment(&self) -> GetBucketRequestPayment<C, M, R>

pub fn get_bucket_tagging(&self) -> GetBucketTagging<C, M, R>

pub fn get_bucket_versioning(&self) -> GetBucketVersioning<C, M, R>

pub fn get_bucket_website(&self) -> GetBucketWebsite<C, M, R>

pub fn get_object(&self) -> GetObject<C, M, R>

pub fn get_object_acl(&self) -> GetObjectAcl<C, M, R>

pub fn get_object_legal_hold(&self) -> GetObjectLegalHold<C, M, R>

pub fn get_object_lock_configuration(
&self
) -> GetObjectLockConfiguration<C, M, R>

pub fn get_object_retention(&self) -> GetObjectRetention<C, M, R>

pub fn get_object_tagging(&self) -> GetObjectTagging<C, M, R>

pub fn get_object_torrent(&self) -> GetObjectTorrent<C, M, R>

pub fn get_public_access_block(&self) -> GetPublicAccessBlock<C, M, R>

pub fn head_bucket(&self) -> HeadBucket<C, M, R>

pub fn head_object(&self) -> HeadObject<C, M, R>

pub fn list_bucket_analytics_configurations(
&self
) -> ListBucketAnalyticsConfigurations<C, M, R>

pub fn list_bucket_intelligent_tiering_configurations(
&self
) -> ListBucketIntelligentTieringConfigurations<C, M, R>

pub fn list_bucket_inventory_configurations(
&self
) -> ListBucketInventoryConfigurations<C, M, R>

pub fn list_bucket_metrics_configurations(
&self
) -> ListBucketMetricsConfigurations<C, M, R>

pub fn list_buckets(&self) -> ListBuckets<C, M, R>

pub fn list_multipart_uploads(&self) -> ListMultipartUploads<C, M, R>

pub fn list_objects(&self) -> ListObjects<C, M, R>

pub fn list_objects_v2(&self) -> ListObjectsV2<C, M, R>

pub fn list_object_versions(&self) -> ListObjectVersions<C, M, R>

pub fn list_parts(&self) -> ListParts<C, M, R>

pub fn put_bucket_accelerate_configuration(
&self
) -> PutBucketAccelerateConfiguration<C, M, R>

pub fn put_bucket_acl(&self) -> PutBucketAcl<C, M, R>

pub fn put_bucket_analytics_configuration(
&self
) -> PutBucketAnalyticsConfiguration<C, M, R>

pub fn put_bucket_cors(&self) -> PutBucketCors<C, M, R>

pub fn put_bucket_encryption(&self) -> PutBucketEncryption<C, M, R>

pub fn put_bucket_intelligent_tiering_configuration(
&self
) -> PutBucketIntelligentTieringConfiguration<C, M, R>

pub fn put_bucket_inventory_configuration(
&self
) -> PutBucketInventoryConfiguration<C, M, R>

pub fn put_bucket_lifecycle_configuration(
&self
) -> PutBucketLifecycleConfiguration<C, M, R>

pub fn put_bucket_logging(&self) -> PutBucketLogging<C, M, R>

pub fn put_bucket_metrics_configuration(
&self
) -> PutBucketMetricsConfiguration<C, M, R>

pub fn put_bucket_notification_configuration(
&self
) -> PutBucketNotificationConfiguration<C, M, R>

pub fn put_bucket_ownership_controls(
&self
) -> PutBucketOwnershipControls<C, M, R>

pub fn put_bucket_policy(&self) -> PutBucketPolicy<C, M, R>

pub fn put_bucket_replication(&self) -> PutBucketReplication<C, M, R>

pub fn put_bucket_request_payment(&self) -> PutBucketRequestPayment<C, M, R>

pub fn put_bucket_tagging(&self) -> PutBucketTagging<C, M, R>

pub fn put_bucket_versioning(&self) -> PutBucketVersioning<C, M, R>

pub fn put_bucket_website(&self) -> PutBucketWebsite<C, M, R>

pub fn put_object(&self) -> PutObject<C, M, R>

pub fn put_object_acl(&self) -> PutObjectAcl<C, M, R>

pub fn put_object_legal_hold(&self) -> PutObjectLegalHold<C, M, R>

pub fn put_object_lock_configuration(
&self
) -> PutObjectLockConfiguration<C, M, R>

pub fn put_object_retention(&self) -> PutObjectRetention<C, M, R>

pub fn put_object_tagging(&self) -> PutObjectTagging<C, M, R>

pub fn put_public_access_block(&self) -> PutPublicAccessBlock<C, M, R>

pub fn restore_object(&self) -> RestoreObject<C, M, R>

pub fn select_object_content(&self) -> SelectObjectContent<C, M, R>

pub fn upload_part(&self) -> UploadPart<C, M, R>

pub fn upload_part_copy(&self) -> UploadPartCopy<C, M, R>

pub fn write_get_object_response(&self) -> WriteGetObjectResponse<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self