aws_sdk_s3

Enum Error

#[non_exhaustive]
pub enum Error {

    AccessDenied(AccessDenied),
    BucketAlreadyExists(BucketAlreadyExists),
    BucketAlreadyOwnedByYou(BucketAlreadyOwnedByYou),
    EncryptionTypeMismatch(EncryptionTypeMismatch),
    IdempotencyParameterMismatch(IdempotencyParameterMismatch),
    InvalidObjectState(InvalidObjectState),
    InvalidRequest(InvalidRequest),
    InvalidWriteOffset(InvalidWriteOffset),
    NoSuchBucket(NoSuchBucket),
    NoSuchKey(NoSuchKey),
    NoSuchUpload(NoSuchUpload),
    NotFound(NotFound),
    ObjectAlreadyInActiveTierError(ObjectAlreadyInActiveTierError),
    ObjectNotInActiveTierError(ObjectNotInActiveTierError),
    TooManyParts(TooManyParts),
    Unhandled(Unhandled),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDenied(AccessDenied)

You might receive this error for several reasons. For details, see the description of this API operation.

BucketAlreadyExists(BucketAlreadyExists)

The requested bucket name is not available. The bucket namespace is shared by all users of the system. Select a different name and try again.

BucketAlreadyOwnedByYou(BucketAlreadyOwnedByYou)

The bucket you tried to create already exists, and you own it. Amazon S3 returns this error in all Amazon Web Services Regions except in the North Virginia Region. For legacy compatibility, if you re-create an existing bucket that you already own in the North Virginia Region, Amazon S3 returns 200 OK and resets the bucket access control lists (ACLs).

EncryptionTypeMismatch(EncryptionTypeMismatch)

The existing object was created with a different encryption type. Subsequent write requests must include the appropriate encryption parameters in the request or while creating the session.

IdempotencyParameterMismatch(IdempotencyParameterMismatch)

Parameters on this idempotent request are inconsistent with parameters used in previous request(s).

For a list of error codes and more information on Amazon S3 errors, see Error codes.

Idempotency ensures that an API request completes no more than one time. With an idempotent request, if the original request completes successfully, any subsequent retries complete successfully without performing any further actions.

InvalidObjectState(InvalidObjectState)

Object is archived and inaccessible until restored.

If the object you are retrieving is stored in the S3 Glacier Flexible Retrieval storage class, the S3 Glacier Deep Archive storage class, the S3 Intelligent-Tiering Archive Access tier, or the S3 Intelligent-Tiering Deep Archive Access tier, before you can retrieve the object you must first restore a copy using RestoreObject. Otherwise, this operation returns an InvalidObjectState error. For information about restoring archived objects, see Restoring Archived Objects in the Amazon S3 User Guide.

InvalidRequest(InvalidRequest)

A parameter or header in your request isn't valid. For details, see the description of this API operation.

InvalidWriteOffset(InvalidWriteOffset)

The write offset value that you specified does not match the current object size.

NoSuchBucket(NoSuchBucket)

The specified bucket does not exist.

NoSuchKey(NoSuchKey)

The specified key does not exist.

NoSuchUpload(NoSuchUpload)

The specified multipart upload does not exist.

NotFound(NotFound)

The specified content does not exist.

ObjectAlreadyInActiveTierError(ObjectAlreadyInActiveTierError)

This action is not allowed against this storage tier.

ObjectNotInActiveTierError(ObjectNotInActiveTierError)

The source object of the COPY action is not in the active tier and is only stored in Amazon S3 Glacier.

TooManyParts(TooManyParts)

You have attempted to add more parts than the maximum of 10000 that are allowed for this object. You can use the CopyObject operation to copy this object to another and then add more data to the newly copied object.

Unhandled(Unhandled)

Deprecated: Matching Unhandled directly is not forwards compatible. Instead, match using a variable wildcard pattern and check .code(): err if err.code() == Some("SpecificExceptionCode") => { /* handle the error */ } See ProvideErrorMetadata for what information is available for the error.

An unexpected error occurred (e.g., invalid JSON returned by the service or an unknown error code).