Struct aws_sdk_s3::input::copy_object_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CopyObjectInput

Implementations

impl Builder

pub fn acl(self, input: ObjectCannedAcl) -> Self

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

pub fn set_acl(self, input: Option<ObjectCannedAcl>) -> Self

pub fn bucket(self, input: impl Into<String>) -> Self

The name of the destination bucket.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the AWS SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the AWS SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_bucket(self, input: Option<String>) -> Self

pub fn cache_control(self, input: impl Into<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn set_cache_control(self, input: Option<String>) -> Self

pub fn content_disposition(self, input: impl Into<String>) -> Self

Specifies presentational information for the object.

pub fn set_content_disposition(self, input: Option<String>) -> Self

pub fn content_encoding(self, input: impl Into<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn set_content_encoding(self, input: Option<String>) -> Self

pub fn content_language(self, input: impl Into<String>) -> Self

The language the content is in.

pub fn set_content_language(self, input: Option<String>) -> Self

pub fn content_type(self, input: impl Into<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn set_content_type(self, input: Option<String>) -> Self

pub fn copy_source(self, input: impl Into<String>) -> Self

Specifies the source object for the copy operation. You specify the value in one of two formats, depending on whether you want to access the source object through an access point:

For objects not accessed through an access point, specify the name of the source bucket and the key of the source object, separated by a slash (/). For example, to copy the object reports/january.pdf from the bucket awsexamplebucket, use awsexamplebucket/reports/january.pdf. The value must be URL encoded.

For objects accessed through access points, specify the Amazon Resource Name (ARN) of the object as accessed through the access point, in the format arn:aws:s3:::accesspoint//object/. For example, to copy the object reports/january.pdf through access point my-access-point owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3:us-west-2:123456789012:accesspoint/my-access-point/object/reports/january.pdf. The value must be URL encoded.

Amazon S3 supports copy operations using access points only when the source and destination buckets are in the same AWS Region.

Alternatively, for objects accessed through Amazon S3 on Outposts, specify the ARN of the object as accessed in the format arn:aws:s3-outposts:::outpost//object/. For example, to copy the object reports/january.pdf through outpost my-outpost owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/object/reports/january.pdf. The value must be URL encoded.

To copy a specific version of an object, append ?versionId= to the value (for example, awsexamplebucket/reports/january.pdf?versionId=QUpfdndhfd8438MNFDN93jdnJFkdmqnh893). If you don't specify a version ID, Amazon S3 copies the latest version of the source object.

pub fn set_copy_source(self, input: Option<String>) -> Self

pub fn copy_source_if_match(self, input: impl Into<String>) -> Self

Copies the object if its entity tag (ETag) matches the specified tag.

pub fn set_copy_source_if_match(self, input: Option<String>) -> Self

pub fn copy_source_if_modified_since(self, input: Instant) -> Self

Copies the object if it has been modified since the specified time.

pub fn set_copy_source_if_modified_since(self, input: Option<Instant>) -> Self

pub fn copy_source_if_none_match(self, input: impl Into<String>) -> Self

Copies the object if its entity tag (ETag) is different than the specified ETag.

pub fn set_copy_source_if_none_match(self, input: Option<String>) -> Self

pub fn copy_source_if_unmodified_since(self, input: Instant) -> Self

Copies the object if it hasn't been modified since the specified time.

pub fn set_copy_source_if_unmodified_since(self, input: Option<Instant>) -> Self

pub fn expires(self, input: Instant) -> Self

The date and time at which the object is no longer cacheable.

pub fn set_expires(self, input: Option<Instant>) -> Self

pub fn grant_full_control(self, input: impl Into<String>) -> Self

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_full_control(self, input: Option<String>) -> Self

pub fn grant_read(self, input: impl Into<String>) -> Self

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_read(self, input: Option<String>) -> Self

pub fn grant_read_acp(self, input: impl Into<String>) -> Self

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_read_acp(self, input: Option<String>) -> Self

pub fn grant_write_acp(self, input: impl Into<String>) -> Self

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_write_acp(self, input: Option<String>) -> Self

pub fn key(self, input: impl Into<String>) -> Self

The key of the destination object.

pub fn set_key(self, input: Option<String>) -> Self

pub fn metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_metadata(self, input: Option<HashMap<String, String>>) -> Self

pub fn metadata_directive(self, input: MetadataDirective) -> Self

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

pub fn set_metadata_directive(self, input: Option<MetadataDirective>) -> Self

pub fn tagging_directive(self, input: TaggingDirective) -> Self

Specifies whether the object tag-set are copied from the source object or replaced with tag-set provided in the request.

pub fn set_tagging_directive(self, input: Option<TaggingDirective>) -> Self

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn set_server_side_encryption(
self,
input: Option<ServerSideEncryption>
) -> Self

pub fn storage_class(self, input: StorageClass) -> Self

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Self

pub fn website_redirect_location(self, input: impl Into<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

pub fn set_website_redirect_location(self, input: Option<String>) -> Self

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

Specifies the algorithm to use to when encrypting the object (for example, AES256).

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

pub fn sse_customer_key(self, input: impl Into<String>) -> Self

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

pub fn set_sse_customer_key(self, input: Option<String>) -> Self

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported AWS SDKs and AWS CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

pub fn ssekms_encryption_context(self, input: impl Into<String>) -> Self

Specifies the AWS KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

pub fn set_ssekms_encryption_context(self, input: Option<String>) -> Self

pub fn bucket_key_enabled(self, input: bool) -> Self

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a COPY action doesn’t affect bucket-level settings for S3 Bucket Key.

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

pub fn copy_source_sse_customer_algorithm(
self,
input: impl Into<String>
) -> Self

Specifies the algorithm to use when decrypting the source object (for example, AES256).

pub fn set_copy_source_sse_customer_algorithm(
self,
input: Option<String>
) -> Self

pub fn copy_source_sse_customer_key(self, input: impl Into<String>) -> Self

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

pub fn set_copy_source_sse_customer_key(self, input: Option<String>) -> Self

pub fn copy_source_sse_customer_key_md5(self, input: impl Into<String>) -> Self

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn set_copy_source_sse_customer_key_md5(self, input: Option<String>) -> Self

pub fn request_payer(self, input: RequestPayer) -> Self

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn set_request_payer(self, input: Option<RequestPayer>) -> Self

pub fn tagging(self, input: impl Into<String>) -> Self

The tag-set for the object destination object this value must be used in conjunction with the TaggingDirective. The tag-set must be encoded as URL Query parameters.

pub fn set_tagging(self, input: Option<String>) -> Self

pub fn object_lock_mode(self, input: ObjectLockMode) -> Self

The Object Lock mode that you want to apply to the copied object.

pub fn set_object_lock_mode(self, input: Option<ObjectLockMode>) -> Self

pub fn object_lock_retain_until_date(self, input: Instant) -> Self

The date and time when you want the copied object's Object Lock to expire.

pub fn set_object_lock_retain_until_date(self, input: Option<Instant>) -> Self

pub fn object_lock_legal_hold_status(
self,
input: ObjectLockLegalHoldStatus
) -> Self

Specifies whether you want to apply a Legal Hold to the copied object.

pub fn set_object_lock_legal_hold_status(
self,
input: Option<ObjectLockLegalHoldStatus>
) -> Self

pub fn expected_bucket_owner(self, input: impl Into<String>) -> Self

The account ID of the expected destination bucket owner. If the destination bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, input: Option<String>) -> Self

pub fn expected_source_bucket_owner(self, input: impl Into<String>) -> Self

The account ID of the expected source bucket owner. If the source bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_source_bucket_owner(self, input: Option<String>) -> Self

pub fn build(self) -> Result<CopyObjectInput, BuildError>

Consumes the builder and constructs a CopyObjectInput