Struct aws_sdk_s3::input::put_object_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for PutObjectInput

Implementations

impl Builder

pub fn acl(self, input: ObjectCannedAcl) -> Self

The canned ACL to apply to the object. For more information, see Canned ACL.

This action is not supported by Amazon S3 on Outposts.

pub fn set_acl(self, input: Option<ObjectCannedAcl>) -> Self

pub fn body(self, input: ByteStream) -> Self

Object data.

pub fn set_body(self, input: Option<ByteStream>) -> Self

pub fn bucket(self, input: impl Into<String>) -> Self

The bucket name to which the PUT action was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the AWS SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the AWS SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_bucket(self, input: Option<String>) -> Self

pub fn cache_control(self, input: impl Into<String>) -> Self

Can be used to specify caching behavior along the request/reply chain. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

pub fn set_cache_control(self, input: Option<String>) -> Self

pub fn content_disposition(self, input: impl Into<String>) -> Self

Specifies presentational information for the object. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec19.html#sec19.5.1.

pub fn set_content_disposition(self, input: Option<String>) -> Self

pub fn content_encoding(self, input: impl Into<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.11.

pub fn set_content_encoding(self, input: Option<String>) -> Self

pub fn content_language(self, input: impl Into<String>) -> Self

The language the content is in.

pub fn set_content_language(self, input: Option<String>) -> Self

pub fn content_length(self, input: i64) -> Self

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13.

pub fn set_content_length(self, input: Option<i64>) -> Self

pub fn content_md5(self, input: impl Into<String>) -> Self

The base64-encoded 128-bit MD5 digest of the message (without the headers) according to RFC 1864. This header can be used as a message integrity check to verify that the data is the same data that was originally sent. Although it is optional, we recommend using the Content-MD5 mechanism as an end-to-end integrity check. For more information about REST request authentication, see REST Authentication.

pub fn set_content_md5(self, input: Option<String>) -> Self

pub fn content_type(self, input: impl Into<String>) -> Self

A standard MIME type describing the format of the contents. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.17.

pub fn set_content_type(self, input: Option<String>) -> Self

pub fn expires(self, input: Instant) -> Self

The date and time at which the object is no longer cacheable. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.21.

pub fn set_expires(self, input: Option<Instant>) -> Self

pub fn grant_full_control(self, input: impl Into<String>) -> Self

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_full_control(self, input: Option<String>) -> Self

pub fn grant_read(self, input: impl Into<String>) -> Self

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_read(self, input: Option<String>) -> Self

pub fn grant_read_acp(self, input: impl Into<String>) -> Self

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_read_acp(self, input: Option<String>) -> Self

pub fn grant_write_acp(self, input: impl Into<String>) -> Self

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

pub fn set_grant_write_acp(self, input: Option<String>) -> Self

pub fn key(self, input: impl Into<String>) -> Self

Object key for which the PUT action was initiated.

pub fn set_key(self, input: Option<String>) -> Self

pub fn metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_metadata(self, input: Option<HashMap<String, String>>) -> Self

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn set_server_side_encryption(
self,
input: Option<ServerSideEncryption>
) -> Self

pub fn storage_class(self, input: StorageClass) -> Self

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Self

pub fn website_redirect_location(self, input: impl Into<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata. For information about object metadata, see Object Key and Metadata.

In the following example, the request header sets the redirect to an object (anotherPage.html) in the same bucket:

x-amz-website-redirect-location: /anotherPage.html

In the following example, the request header sets the object redirect to another website:

x-amz-website-redirect-location: http://www.example.com/

For more information about website hosting in Amazon S3, see Hosting Websites on Amazon S3 and How to Configure Website Page Redirects.

pub fn set_website_redirect_location(self, input: Option<String>) -> Self

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

Specifies the algorithm to use to when encrypting the object (for example, AES256).

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

pub fn sse_customer_key(self, input: impl Into<String>) -> Self

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

pub fn set_sse_customer_key(self, input: Option<String>) -> Self

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the AWS Key Management Service (AWS KMS) symmetrical customer managed customer master key (CMK) that was used for the object. If you specify x-amz-server-side-encryption:aws:kms, but do not provide x-amz-server-side-encryption-aws-kms-key-id, Amazon S3 uses the AWS managed CMK in AWS to protect the data. If the KMS key does not exist in the same account issuing the command, you must use the full ARN and not just the ID.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

pub fn ssekms_encryption_context(self, input: impl Into<String>) -> Self

Specifies the AWS KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

pub fn set_ssekms_encryption_context(self, input: Option<String>) -> Self

pub fn bucket_key_enabled(self, input: bool) -> Self

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a PUT action doesn’t affect bucket-level settings for S3 Bucket Key.

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

pub fn request_payer(self, input: RequestPayer) -> Self

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn set_request_payer(self, input: Option<RequestPayer>) -> Self

pub fn tagging(self, input: impl Into<String>) -> Self

The tag-set for the object. The tag-set must be encoded as URL Query parameters. (For example, "Key1=Value1")

pub fn set_tagging(self, input: Option<String>) -> Self

pub fn object_lock_mode(self, input: ObjectLockMode) -> Self

The Object Lock mode that you want to apply to this object.

pub fn set_object_lock_mode(self, input: Option<ObjectLockMode>) -> Self

pub fn object_lock_retain_until_date(self, input: Instant) -> Self

The date and time when you want this object's Object Lock to expire. Must be formatted as a timestamp parameter.

pub fn set_object_lock_retain_until_date(self, input: Option<Instant>) -> Self

pub fn object_lock_legal_hold_status(
self,
input: ObjectLockLegalHoldStatus
) -> Self

Specifies whether a legal hold will be applied to this object. For more information about S3 Object Lock, see Object Lock.

pub fn set_object_lock_legal_hold_status(
self,
input: Option<ObjectLockLegalHoldStatus>
) -> Self

pub fn expected_bucket_owner(self, input: impl Into<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, input: Option<String>) -> Self

pub fn build(self) -> Result<PutObjectInput, BuildError>

Consumes the builder and constructs a PutObjectInput