Struct aws_sdk_s3::input::PutObjectInput

#[non_exhaustive]
pub struct PutObjectInput {

    pub acl: Option<ObjectCannedAcl>,
    pub body: ByteStream,
    pub bucket: Option<String>,
    pub cache_control: Option<String>,
    pub content_disposition: Option<String>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_length: i64,
    pub content_md5: Option<String>,
    pub content_type: Option<String>,
    pub expires: Option<Instant>,
    pub grant_full_control: Option<String>,
    pub grant_read: Option<String>,
    pub grant_read_acp: Option<String>,
    pub grant_write_acp: Option<String>,
    pub key: Option<String>,
    pub metadata: Option<HashMap<String, String>>,
    pub server_side_encryption: Option<ServerSideEncryption>,
    pub storage_class: Option<StorageClass>,
    pub website_redirect_location: Option<String>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub ssekms_key_id: Option<String>,
    pub ssekms_encryption_context: Option<String>,
    pub bucket_key_enabled: bool,
    pub request_payer: Option<RequestPayer>,
    pub tagging: Option<String>,
    pub object_lock_mode: Option<ObjectLockMode>,
    pub object_lock_retain_until_date: Option<Instant>,
    pub object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>,
    pub expected_bucket_owner: Option<String>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

acl: Option<ObjectCannedAcl>

The canned ACL to apply to the object. For more information, see Canned ACL.

This action is not supported by Amazon S3 on Outposts.

body: ByteStream

Object data.

bucket: Option<String>

The bucket name to which the PUT action was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the AWS SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the AWS SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

cache_control: Option<String>

Can be used to specify caching behavior along the request/reply chain. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

content_disposition: Option<String>

Specifies presentational information for the object. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec19.html#sec19.5.1.

content_encoding: Option<String>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.11.

content_language: Option<String>

The language the content is in.

content_length: i64

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13.

content_md5: Option<String>

The base64-encoded 128-bit MD5 digest of the message (without the headers) according to RFC 1864. This header can be used as a message integrity check to verify that the data is the same data that was originally sent. Although it is optional, we recommend using the Content-MD5 mechanism as an end-to-end integrity check. For more information about REST request authentication, see REST Authentication.

content_type: Option<String>

A standard MIME type describing the format of the contents. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.17.

expires: Option<Instant>

The date and time at which the object is no longer cacheable. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.21.

grant_full_control: Option<String>

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

grant_read: Option<String>

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

grant_read_acp: Option<String>

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

grant_write_acp: Option<String>

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

key: Option<String>

Object key for which the PUT action was initiated.

metadata: Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

server_side_encryption: Option<ServerSideEncryption>

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

storage_class: Option<StorageClass>

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

website_redirect_location: Option<String>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata. For information about object metadata, see Object Key and Metadata.

In the following example, the request header sets the redirect to an object (anotherPage.html) in the same bucket:

x-amz-website-redirect-location: /anotherPage.html

In the following example, the request header sets the object redirect to another website:

x-amz-website-redirect-location: http://www.example.com/

For more information about website hosting in Amazon S3, see Hosting Websites on Amazon S3 and How to Configure Website Page Redirects.

sse_customer_algorithm: Option<String>

Specifies the algorithm to use to when encrypting the object (for example, AES256).

sse_customer_key: Option<String>

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

sse_customer_key_md5: Option<String>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

ssekms_key_id: Option<String>

If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the AWS Key Management Service (AWS KMS) symmetrical customer managed customer master key (CMK) that was used for the object. If you specify x-amz-server-side-encryption:aws:kms, but do not provide x-amz-server-side-encryption-aws-kms-key-id, Amazon S3 uses the AWS managed CMK in AWS to protect the data. If the KMS key does not exist in the same account issuing the command, you must use the full ARN and not just the ID.

ssekms_encryption_context: Option<String>

Specifies the AWS KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

bucket_key_enabled: bool

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a PUT action doesn’t affect bucket-level settings for S3 Bucket Key.

request_payer: Option<RequestPayer>

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

tagging: Option<String>

The tag-set for the object. The tag-set must be encoded as URL Query parameters. (For example, "Key1=Value1")

object_lock_mode: Option<ObjectLockMode>

The Object Lock mode that you want to apply to this object.

object_lock_retain_until_date: Option<Instant>

The date and time when you want this object's Object Lock to expire. Must be formatted as a timestamp parameter.

object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>

Specifies whether a legal hold will be applied to this object. For more information about S3 Object Lock, see Object Lock.

expected_bucket_owner: Option<String>

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

Implementations

impl PutObjectInput

pub fn make_operation(
self,
_config: &Config
) -> Result<Operation<PutObject, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<PutObject>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture PutObjectInput