Struct aws_sdk_s3::input::PutObjectRetentionInput

#[non_exhaustive]
pub struct PutObjectRetentionInput {
    pub bucket: Option<String>,
    pub key: Option<String>,
    pub retention: Option<ObjectLockRetention>,
    pub request_payer: Option<RequestPayer>,
    pub version_id: Option<String>,
    pub bypass_governance_retention: bool,
    pub content_md5: Option<String>,
    pub expected_bucket_owner: Option<String>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

bucket: Option<String>

The bucket name that contains the object you want to apply this Object Retention configuration to.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

key: Option<String>

The key name for the object that you want to apply this Object Retention configuration to.

retention: Option<ObjectLockRetention>

The container element for the Object Retention configuration.

request_payer: Option<RequestPayer>

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

version_id: Option<String>

The version ID for the object that you want to apply this Object Retention configuration to.

bypass_governance_retention: bool

Indicates whether this action should bypass Governance-mode restrictions.

content_md5: Option<String>

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

expected_bucket_owner: Option<String>

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

Implementations

impl PutObjectRetentionInput

pub async fn make_operation( &self, _config: &Config ) -> Result<Operation<PutObjectRetention, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<PutObjectRetention>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture PutObjectRetentionInput

impl PutObjectRetentionInput

pub fn bucket(&self) -> Option<&str>

The bucket name that contains the object you want to apply this Object Retention configuration to.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

pub fn key(&self) -> Option<&str>

The key name for the object that you want to apply this Object Retention configuration to.

pub fn retention(&self) -> Option<&ObjectLockRetention>

The container element for the Object Retention configuration.

pub fn request_payer(&self) -> Option<&RequestPayer>

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn version_id(&self) -> Option<&str>

The version ID for the object that you want to apply this Object Retention configuration to.

pub fn bypass_governance_retention(&self) -> bool

Indicates whether this action should bypass Governance-mode restrictions.

pub fn content_md5(&self) -> Option<&str>

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

pub fn expected_bucket_owner(&self) -> Option<&str>

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.