Struct aws_sdk_s3::input::WriteGetObjectResponseInput

#[non_exhaustive]
pub struct WriteGetObjectResponseInput {

    pub request_route: Option<String>,
    pub request_token: Option<String>,
    pub body: ByteStream,
    pub status_code: i32,
    pub error_code: Option<String>,
    pub error_message: Option<String>,
    pub accept_ranges: Option<String>,
    pub cache_control: Option<String>,
    pub content_disposition: Option<String>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_length: i64,
    pub content_range: Option<String>,
    pub content_type: Option<String>,
    pub delete_marker: bool,
    pub e_tag: Option<String>,
    pub expires: Option<Instant>,
    pub expiration: Option<String>,
    pub last_modified: Option<Instant>,
    pub missing_meta: i32,
    pub metadata: Option<HashMap<String, String>>,
    pub object_lock_mode: Option<ObjectLockMode>,
    pub object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>,
    pub object_lock_retain_until_date: Option<Instant>,
    pub parts_count: i32,
    pub replication_status: Option<ReplicationStatus>,
    pub request_charged: Option<RequestCharged>,
    pub restore: Option<String>,
    pub server_side_encryption: Option<ServerSideEncryption>,
    pub sse_customer_algorithm: Option<String>,
    pub ssekms_key_id: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub storage_class: Option<StorageClass>,
    pub tag_count: i32,
    pub version_id: Option<String>,
    pub bucket_key_enabled: bool,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

request_route: Option<String>

Route prefix to the HTTP URL generated.

request_token: Option<String>

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

body: ByteStream

The object data.

status_code: i32

The integer status code for an HTTP response of a corresponding GetObject request.

Status Codes

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

error_code: Option<String>

A string that uniquely identifies an error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. Regex value is "^[A-Z][a-zA-Z]+$".

error_message: Option<String>

Contains a generic description of the error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

accept_ranges: Option<String>

Indicates that a range of bytes was specified.

cache_control: Option<String>

Specifies caching behavior along the request/reply chain.

content_disposition: Option<String>

Specifies presentational information for the object.

content_encoding: Option<String>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

content_language: Option<String>

The language the content is in.

content_length: i64

The size of the content body in bytes.

content_range: Option<String>

The portion of the object returned in the response.

content_type: Option<String>

A standard MIME type describing the format of the object data.

delete_marker: bool

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker.

e_tag: Option<String>

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

expires: Option<Instant>

The date and time at which the object is no longer cacheable.

expiration: Option<String>

If object stored in Amazon S3 expiration is configured (see PUT Bucket lifecycle) it includes expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

last_modified: Option<Instant>

The date and time that the object was last modified.

missing_meta: i32

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

metadata: Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

object_lock_mode: Option<ObjectLockMode>

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>

Indicates whether an object stored in Amazon S3 has an active legal hold.

object_lock_retain_until_date: Option<Instant>

The date and time when Object Lock is configured to expire.

parts_count: i32

The count of parts this object has.

replication_status: Option<ReplicationStatus>

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

request_charged: Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request.

restore: Option<String>

Provides information about object restoration operation and expiration time of the restored object copy.

server_side_encryption: Option<ServerSideEncryption>

The server-side encryption algorithm used when storing requested object in Amazon S3 (for example, AES256, aws:kms).

sse_customer_algorithm: Option<String>

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

ssekms_key_id: Option<String>

If present, specifies the ID of the AWS Key Management Service (AWS KMS) symmetric customer managed customer master key (CMK) that was used for stored in Amazon S3 object.

sse_customer_key_md5: Option<String>

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

storage_class: Option<StorageClass>

The class of storage used to store object in Amazon S3.

tag_count: i32

The number of tags, if any, on the object.

version_id: Option<String>

An ID used to reference a specific version of the object.

bucket_key_enabled: bool

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with AWS KMS (SSE-KMS).

Implementations

impl WriteGetObjectResponseInput

pub fn make_operation(
self,
_config: &Config
) -> Result<Operation<WriteGetObjectResponse, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<WriteGetObjectResponse>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture WriteGetObjectResponseInput