Struct aws_sdk_s3::input::upload_part_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UploadPartInput

Implementations

impl Builder

pub fn body(self, inp: ByteStream) -> Self

Object data.

pub fn set_body(self, inp: ByteStream) -> Self

pub fn bucket(self, inp: impl Into<String>) -> Self

The name of the bucket to which the multipart upload was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the AWS SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the AWS SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_bucket(self, inp: String) -> Self

pub fn content_length(self, inp: i64) -> Self

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

pub fn set_content_length(self, inp: i64) -> Self

pub fn content_md5(self, inp: impl Into<String>) -> Self

The base64-encoded 128-bit MD5 digest of the part data. This parameter is auto-populated when using the command from the CLI. This parameter is required if object lock parameters are specified.

pub fn set_content_md5(self, inp: Option<String>) -> Self

pub fn key(self, inp: impl Into<String>) -> Self

Object key for which the multipart upload was initiated.

pub fn set_key(self, inp: String) -> Self

pub fn part_number(self, inp: i32) -> Self

Part number of part being uploaded. This is a positive integer between 1 and 10,000.

pub fn set_part_number(self, inp: i32) -> Self

pub fn upload_id(self, inp: impl Into<String>) -> Self

Upload ID identifying the multipart upload whose part is being uploaded.

pub fn set_upload_id(self, inp: Option<String>) -> Self

pub fn sse_customer_algorithm(self, inp: impl Into<String>) -> Self

Specifies the algorithm to use to when encrypting the object (for example, AES256).

pub fn set_sse_customer_algorithm(self, inp: Option<String>) -> Self

pub fn sse_customer_key(self, inp: impl Into<String>) -> Self

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

pub fn set_sse_customer_key(self, inp: Option<String>) -> Self

pub fn sse_customer_key_md5(self, inp: impl Into<String>) -> Self

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn set_sse_customer_key_md5(self, inp: Option<String>) -> Self

pub fn request_payer(self, inp: RequestPayer) -> Self

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn set_request_payer(self, inp: Option<RequestPayer>) -> Self

pub fn expected_bucket_owner(self, inp: impl Into<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, inp: Option<String>) -> Self

pub fn build(self) -> Result<UploadPartInput, BuildError>

Consumes the builder and constructs a UploadPartInput