Module aws_sdk_s3::model

Modules

abort_incomplete_multipart_upload

See AbortIncompleteMultipartUpload

accelerate_configuration

See AccelerateConfiguration

access_control_policy

See AccessControlPolicy

access_control_translation

See AccessControlTranslation

analytics_and_operator

See AnalyticsAndOperator

analytics_configuration

See AnalyticsConfiguration

analytics_export_destination

See AnalyticsExportDestination

analytics_s3_bucket_destination

See AnalyticsS3BucketDestination

bucket

See Bucket

bucket_lifecycle_configuration

See BucketLifecycleConfiguration

bucket_logging_status

See BucketLoggingStatus

common_prefix

See CommonPrefix

completed_multipart_upload

See CompletedMultipartUpload

completed_part

See CompletedPart

condition

See Condition

copy_object_result

See CopyObjectResult

copy_part_result

See CopyPartResult

cors_configuration

See CorsConfiguration

cors_rule

See CorsRule

create_bucket_configuration

See CreateBucketConfiguration

csv_input

See CsvInput

csv_output

See CsvOutput

default_retention

See DefaultRetention

delete

See Delete

delete_marker_entry

See DeleteMarkerEntry

delete_marker_replication

See DeleteMarkerReplication

deleted_object

See DeletedObject

destination

See Destination

encryption

See Encryption

encryption_configuration

See EncryptionConfiguration

error

See Error

error_document

See ErrorDocument

existing_object_replication

See ExistingObjectReplication

filter_rule

See FilterRule

glacier_job_parameters

See GlacierJobParameters

grant

See Grant

grantee

See Grantee

index_document

See IndexDocument

initiator

See Initiator

input_serialization

See InputSerialization

intelligent_tiering_and_operator

See IntelligentTieringAndOperator

intelligent_tiering_configuration

See IntelligentTieringConfiguration

intelligent_tiering_filter

See IntelligentTieringFilter

inventory_configuration

See InventoryConfiguration

inventory_destination

See InventoryDestination

inventory_encryption

See InventoryEncryption

inventory_filter

See InventoryFilter

inventory_s3_bucket_destination

See InventoryS3BucketDestination

inventory_schedule

See InventorySchedule

json_input

See JsonInput

json_output

See JsonOutput

lambda_function_configuration

See LambdaFunctionConfiguration

lifecycle_expiration

See LifecycleExpiration

lifecycle_rule

See LifecycleRule

lifecycle_rule_and_operator

See LifecycleRuleAndOperator

logging_enabled

See LoggingEnabled

metadata_entry

See MetadataEntry

metrics

See Metrics

metrics_and_operator

See MetricsAndOperator

metrics_configuration

See MetricsConfiguration

multipart_upload

See MultipartUpload

noncurrent_version_expiration

See NoncurrentVersionExpiration

noncurrent_version_transition

See NoncurrentVersionTransition

notification_configuration

See NotificationConfiguration

notification_configuration_filter

See NotificationConfigurationFilter

object

See Object

object_identifier

See ObjectIdentifier

object_lock_configuration

See ObjectLockConfiguration

object_lock_legal_hold

See ObjectLockLegalHold

object_lock_retention

See ObjectLockRetention

object_lock_rule

See ObjectLockRule

object_version

See ObjectVersion

output_location

See OutputLocation

output_serialization

See OutputSerialization

owner

See Owner

ownership_controls

See OwnershipControls

ownership_controls_rule

See OwnershipControlsRule

parquet_input

See ParquetInput

part

See Part

policy_status

See PolicyStatus

public_access_block_configuration

See PublicAccessBlockConfiguration

queue_configuration

See QueueConfiguration

redirect

See Redirect

redirect_all_requests_to

See RedirectAllRequestsTo

replica_modifications

See ReplicaModifications

replication_configuration

See ReplicationConfiguration

replication_rule

See ReplicationRule

replication_rule_and_operator

See ReplicationRuleAndOperator

replication_time

See ReplicationTime

replication_time_value

See ReplicationTimeValue

request_payment_configuration

See RequestPaymentConfiguration

restore_request

See RestoreRequest

routing_rule

See RoutingRule

s3_key_filter

See S3KeyFilter

s3_location

See S3Location

select_parameters

See SelectParameters

server_side_encryption_by_default

See ServerSideEncryptionByDefault

server_side_encryption_configuration

See ServerSideEncryptionConfiguration

server_side_encryption_rule

See ServerSideEncryptionRule

source_selection_criteria

See SourceSelectionCriteria

sse_kms_encrypted_objects

See SseKmsEncryptedObjects

ssekms

See Ssekms

sses3

See Sses3

storage_class_analysis

See StorageClassAnalysis

storage_class_analysis_data_export

See StorageClassAnalysisDataExport

tag

See Tag

tagging

See Tagging

target_grant

See TargetGrant

tiering

See Tiering

topic_configuration

See TopicConfiguration

transition

See Transition

versioning_configuration

See VersioningConfiguration

website_configuration

See WebsiteConfiguration

Structs

AbortIncompleteMultipartUpload

Specifies the days since the initiation of an incomplete multipart upload that Amazon S3 will

AccelerateConfiguration

Configures the transfer acceleration state for an Amazon S3 bucket. For more information, see

AccessControlPolicy

Contains the elements that set the ACL permissions for an object per grantee.

AccessControlTranslation

A container for information about access control for replicas.

AnalyticsAndOperator

A conjunction (logical AND) of predicates, which is used in evaluating a metrics filter.

AnalyticsConfiguration

Specifies the configuration and any analyses for the analytics filter of an Amazon S3 bucket.

AnalyticsExportDestination

Where to publish the analytics results.

AnalyticsS3BucketDestination

Contains information about where to publish the analytics results.

Bucket

In terms of implementation, a Bucket is a resource. An Amazon S3 bucket name is globally

BucketLifecycleConfiguration

Specifies the lifecycle configuration for objects in an Amazon S3 bucket. For more

BucketLoggingStatus

Container for logging status information.

CommonPrefix

Container for all (if there are any) keys between Prefix and the next occurrence of the

CompletedMultipartUpload

The container for the completed multipart upload details.

CompletedPart

Details of the parts that were uploaded.

Condition

A container for describing a condition that must be met for the specified redirect to

CopyObjectResult

Container for all response elements.

CopyPartResult

Container for all response elements.

CorsConfiguration

Describes the cross-origin access configuration for objects in an Amazon S3 bucket. For more

CorsRule

Specifies a cross-origin access rule for an Amazon S3 bucket.

CreateBucketConfiguration

The configuration information for the bucket.

CsvInput

Describes how an uncompressed comma-separated values (CSV)-formatted input object is

CsvOutput

Describes how uncompressed comma-separated values (CSV)-formatted results are

DefaultRetention

The container element for specifying the default Object Lock retention settings for new

Delete

Container for the objects to delete.

DeleteMarkerEntry

Information about the delete marker.

DeleteMarkerReplication

Specifies whether Amazon S3 replicates delete markers. If you specify a Filter

DeletedObject

Information about the deleted object.

Destination

Specifies information about where to publish analysis or configuration results for an

Encryption

Contains the type of server-side encryption used.

EncryptionConfiguration

Specifies encryption-related information for an Amazon S3 bucket that is a destination for

Error

Container for all error elements.

ErrorDocument

The error information.

Event

The bucket event for which to send notifications.

ExistingObjectReplication

Optional configuration to replicate existing source bucket objects. For more

FilterRule

Specifies the Amazon S3 object key name to filter on and whether to filter on the suffix or

GlacierJobParameters

Container for S3 Glacier job parameters.

Grant

Container for grant information.

Grantee

Container for the person being granted permissions.

IndexDocument

Container for the Suffix element.

Initiator

Container element that identifies who initiated the multipart upload.

InputSerialization

Describes the serialization format of the object.

IntelligentTieringAndOperator

A container for specifying S3 Intelligent-Tiering filters. The filters determine the

IntelligentTieringConfiguration

Specifies the S3 Intelligent-Tiering configuration for an Amazon S3 bucket.

IntelligentTieringFilter

The Filter is used to identify objects that the S3 Intelligent-Tiering

InventoryConfiguration

Specifies the inventory configuration for an Amazon S3 bucket. For more information, see

InventoryDestination

Specifies the inventory configuration for an Amazon S3 bucket.

InventoryEncryption

Contains the type of server-side encryption used to encrypt the inventory

InventoryFilter

Specifies an inventory filter. The inventory only includes objects that meet the

InventoryS3BucketDestination

Contains the bucket name, file format, bucket owner (optional), and prefix (optional)

InventorySchedule

Specifies the schedule for generating inventory results.

JsonInput

Specifies JSON as object's input serialization format.

JsonOutput

Specifies JSON as request's output serialization format.

LambdaFunctionConfiguration

A container for specifying the configuration for AWS Lambda notifications.

LifecycleExpiration

Container for the expiration for the lifecycle of the object.

LifecycleRule

A lifecycle rule for individual objects in an Amazon S3 bucket.

LifecycleRuleAndOperator

This is used in a Lifecycle Rule Filter to apply a logical AND to two or more

LoggingEnabled

Describes where logs are stored and the prefix that Amazon S3 assigns to all log object keys

MetadataEntry

A metadata key-value pair to store with an object.

Metrics

A container specifying replication metrics-related settings enabling replication

MetricsAndOperator

A conjunction (logical AND) of predicates, which is used in evaluating a metrics filter.

MetricsConfiguration

Specifies a metrics configuration for the CloudWatch request metrics (specified by the

MultipartUpload

Container for the MultipartUpload for the Amazon S3 object.

NoncurrentVersionExpiration

Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently

NoncurrentVersionTransition

Container for the transition rule that describes when noncurrent objects transition to

NotificationConfiguration

A container for specifying the notification configuration of the bucket. If this element

NotificationConfigurationFilter

Specifies object key name filtering rules. For information about key name filtering, see

Object

An object consists of data and its descriptive metadata.

ObjectIdentifier

Object Identifier is unique value to identify objects.

ObjectLockConfiguration

The container element for Object Lock configuration parameters.

ObjectLockLegalHold

A Legal Hold configuration for an object.

ObjectLockRetention

A Retention configuration for an object.

ObjectLockRule

The container element for an Object Lock rule.

ObjectVersion

The version of an object.

OutputLocation

Describes the location where the restore job's output is stored.

OutputSerialization

Describes how results of the Select job are serialized.

Owner

Container for the owner's display name and ID.

OwnershipControls

The container element for a bucket's ownership controls.

OwnershipControlsRule

The container element for an ownership control rule.

ParquetInput

Container for Parquet.

Part

Container for elements related to a part.

PolicyStatus

The container element for a bucket's policy status.

PublicAccessBlockConfiguration

The PublicAccessBlock configuration that you want to apply to this Amazon S3 bucket. You can

QueueConfiguration

Specifies the configuration for publishing messages to an Amazon Simple Queue Service

Redirect

Specifies how requests are redirected. In the event of an error, you can specify a

RedirectAllRequestsTo

Specifies the redirect behavior of all requests to a website endpoint of an Amazon S3

ReplicaModifications

A filter that you can specify for selection for modifications on replicas. Amazon S3 doesn't

ReplicationConfiguration

A container for replication rules. You can add up to 1,000 rules. The maximum size of a

ReplicationRule

Specifies which Amazon S3 objects to replicate and where to store the replicas.

ReplicationRuleAndOperator

A container for specifying rule filters. The filters determine the subset of objects to

ReplicationTime

A container specifying S3 Replication Time Control (S3 RTC) related information, including whether S3 RTC is

ReplicationTimeValue

A container specifying the time value for S3 Replication Time Control (S3 RTC) and replication metrics

RequestPaymentConfiguration

Container for Payer.

RestoreRequest

Container for restore job parameters.

RoutingRule

Specifies the redirect behavior and when a redirect is applied. For more information

S3KeyFilter

A container for object key name prefix and suffix filtering rules.

S3Location

Describes an Amazon S3 location that will receive the results of the restore request.

SelectParameters

Describes the parameters for Select job types.

ServerSideEncryptionByDefault

Describes the default server-side encryption to apply to new objects in the bucket. If a

ServerSideEncryptionConfiguration

Specifies the default server-side-encryption configuration.

ServerSideEncryptionRule

Specifies the default server-side encryption configuration.

SourceSelectionCriteria

A container that describes additional filters for identifying the source objects that

SseKmsEncryptedObjects

A container for filter information for the selection of S3 objects encrypted with AWS

Ssekms

Specifies the use of SSE-KMS to encrypt delivered inventory reports.

Sses3

Specifies the use of SSE-S3 to encrypt delivered inventory reports.

StorageClassAnalysis

Specifies data related to access patterns to be collected and made available to analyze

StorageClassAnalysisDataExport

Container for data related to the storage class analysis for an Amazon S3 bucket for

Tag

A container of a key value name pair.

Tagging

Container for TagSet elements.

TargetGrant

Container for granting information.

Tiering

The S3 Intelligent-Tiering storage class is designed to optimize storage costs by

TopicConfiguration

A container for specifying the configuration for publication of messages to an Amazon

Transition

Specifies when an object transitions to a specified storage class. For more information

VersioningConfiguration

Describes the versioning state of an Amazon S3 bucket. For more information, see PUT

WebsiteConfiguration

Specifies website configuration parameters for an Amazon S3 bucket.

Enums

AnalyticsFilter

AnalyticsS3ExportFileFormat

ArchiveStatus

BucketAccelerateStatus

BucketCannedAcl

BucketLocationConstraint

BucketLogsPermission

BucketVersioningStatus

CompressionType

DeleteMarkerReplicationStatus

EncodingType

Requests Amazon S3 to encode the object keys in the response and specifies the encoding

ExistingObjectReplicationStatus

ExpirationStatus

ExpressionType

FileHeaderInfo

FilterRuleName

IntelligentTieringAccessTier

IntelligentTieringStatus

InventoryFormat

InventoryFrequency

InventoryIncludedObjectVersions

InventoryOptionalField

JsonType

LifecycleRuleFilter

MetadataDirective

MetricsFilter

MetricsStatus

MfaDelete

MfaDeleteStatus

ObjectCannedAcl

ObjectLockEnabled

ObjectLockLegalHoldStatus

ObjectLockMode

ObjectLockRetentionMode

ObjectOwnership

The container element for object ownership for a bucket's ownership controls.

ObjectStorageClass

ObjectVersionStorageClass

OwnerOverride

Payer

Permission

Protocol

QuoteFields

ReplicaModificationsStatus

ReplicationRuleFilter

ReplicationRuleStatus

ReplicationStatus

ReplicationTimeStatus

RequestCharged

If present, indicates that the requester was successfully charged for the

RequestPayer

Confirms that the requester knows that they will be charged for the request. Bucket

RestoreRequestType

ServerSideEncryption

SseKmsEncryptedObjectsStatus

StorageClass

StorageClassAnalysisSchemaVersion

TaggingDirective

Tier

TransitionStorageClass

Type