Struct aws_sdk_s3::model::ReplicationRule

#[non_exhaustive]
pub struct ReplicationRule { /* private fields */ }

Specifies which Amazon S3 objects to replicate and where to store the replicas.

Implementations

impl ReplicationRule

pub fn id(&self) -> Option<&str>

A unique identifier for the rule. The maximum value is 255 characters.

pub fn priority(&self) -> i32

The priority indicates which rule has precedence whenever two or more replication rules conflict. Amazon S3 will attempt to replicate objects according to all replication rules. However, if there are two or more rules with the same destination bucket, then objects will be replicated according to the rule with the highest priority. The higher the number, the higher the priority.

For more information, see Replication in the Amazon S3 User Guide.

pub fn prefix(&self) -> Option<&str>

๐Deprecated

An object key name prefix that identifies the object or objects to which the rule applies. The maximum prefix length is 1,024 characters. To include all objects in a bucket, specify an empty string.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

pub fn filter(&self) -> Option<&ReplicationRuleFilter>

A filter that identifies the subset of objects to which the replication rule applies. A Filter must specify exactly one Prefix, Tag, or an And child element.

pub fn status(&self) -> Option<&ReplicationRuleStatus>

Specifies whether the rule is enabled.

pub fn source_selection_criteria(&self) -> Option<&SourceSelectionCriteria>

A container that describes additional filters for identifying the source objects that you want to replicate. You can choose to enable or disable the replication of these objects. Currently, Amazon S3 supports only the filter that you can specify for objects created with server-side encryption using a customer managed key stored in Amazon Web Services Key Management Service (SSE-KMS).

pub fn existing_object_replication(&self) -> Option<&ExistingObjectReplication>

pub fn destination(&self) -> Option<&Destination>

A container for information about the replication destination and its configurations including enabling the S3 Replication Time Control (S3 RTC).

pub fn delete_marker_replication(&self) -> Option<&DeleteMarkerReplication>

Specifies whether Amazon S3 replicates delete markers. If you specify a Filter in your replication configuration, you must also include a DeleteMarkerReplication element. If your Filter includes a Tag element, the DeleteMarkerReplication Status must be set to Disabled, because Amazon S3 does not support replicating delete markers for tag-based rules. For an example configuration, see Basic Rule Configuration.

For more information about delete marker replication, see Basic Rule Configuration.

If you are using an earlier version of the replication configuration, Amazon S3 handles replication of delete markers differently. For more information, see Backward Compatibility.

impl ReplicationRule

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ReplicationRule.