Module aws_sdk_s3::operation

Structs

AbortMultipartUpload

This action aborts a multipart upload. After a multipart upload is aborted, no

CompleteMultipartUpload

Completes a multipart upload by assembling previously uploaded parts.

CopyObject

Creates a copy of an object that is already stored in Amazon S3.

CreateBucket

Creates a new S3 bucket. To create a bucket, you must register with Amazon S3 and have a

CreateMultipartUpload

This action initiates a multipart upload and returns an upload ID. This upload ID is

DeleteBucket

Deletes the S3 bucket. All objects (including all object versions and delete markers) in

DeleteBucketAnalyticsConfiguration

Deletes an analytics configuration for the bucket (specified by the analytics

DeleteBucketCors

Deletes the cors configuration information set for the bucket.

DeleteBucketEncryption

This implementation of the DELETE action removes default encryption from the bucket.

DeleteBucketIntelligentTieringConfiguration

Deletes the S3 Intelligent-Tiering configuration from the specified bucket.

DeleteBucketInventoryConfiguration

Deletes an inventory configuration (identified by the inventory ID) from the

DeleteBucketLifecycle

Deletes the lifecycle configuration from the specified bucket. Amazon S3 removes all the

DeleteBucketMetricsConfiguration

Deletes a metrics configuration for the Amazon CloudWatch request metrics (specified by the

DeleteBucketOwnershipControls

Removes OwnershipControls for an Amazon S3 bucket. To use this operation, you

DeleteBucketPolicy

This implementation of the DELETE action uses the policy subresource to delete the

DeleteBucketReplication

Deletes the replication configuration from the bucket.

DeleteBucketTagging

Deletes the tags from the bucket.

DeleteBucketWebsite

This action removes the website configuration for a bucket. Amazon S3 returns a 200

DeleteObject

Removes the null version (if there is one) of an object and inserts a delete marker,

DeleteObjectTagging

Removes the entire tag set from the specified object. For more information about

DeleteObjects

This action enables you to delete multiple objects from a bucket using a single HTTP

DeletePublicAccessBlock

Removes the PublicAccessBlock configuration for an Amazon S3 bucket. To use this

GetBucketAccelerateConfiguration

This implementation of the GET action uses the accelerate subresource to

GetBucketAcl

This implementation of the GET action uses the acl

GetBucketAnalyticsConfiguration

This implementation of the GET action returns an analytics configuration (identified

GetBucketCors

Returns the cors configuration information set for the bucket.

GetBucketEncryption

Returns the default encryption configuration for an Amazon S3 bucket. If the bucket does not

GetBucketIntelligentTieringConfiguration

Gets the S3 Intelligent-Tiering configuration from the specified bucket.

GetBucketInventoryConfiguration

Returns an inventory configuration (identified by the inventory configuration ID) from

GetBucketLifecycleConfiguration

GetBucketLocation

Returns the Region the bucket resides in. You set the bucket's Region using the

GetBucketLogging

Returns the logging status of a bucket and the permissions users have to view and modify

GetBucketMetricsConfiguration

Gets a metrics configuration (specified by the metrics configuration ID) from the

GetBucketNotificationConfiguration

Returns the notification configuration of a bucket.

GetBucketOwnershipControls

Retrieves OwnershipControls for an Amazon S3 bucket. To use this operation, you

GetBucketPolicy

Returns the policy of a specified bucket. If you are using an identity other than the

GetBucketPolicyStatus

Retrieves the policy status for an Amazon S3 bucket, indicating whether the bucket is public.

GetBucketReplication

Returns the replication configuration of a bucket.

GetBucketRequestPayment

Returns the request payment configuration of a bucket. To use this version of the

GetBucketTagging

Returns the tag set associated with the bucket.

GetBucketVersioning

Returns the versioning state of a bucket.

GetBucketWebsite

Returns the website configuration for a bucket. To host website on Amazon S3, you can

GetObject

Retrieves objects from Amazon S3. To use GET, you must have READ

GetObjectAcl

Returns the access control list (ACL) of an object. To use this operation, you must have

GetObjectLegalHold

Gets an object's current Legal Hold status. For more information, see Locking Objects.

GetObjectLockConfiguration

Gets the Object Lock configuration for a bucket. The rule specified in the Object Lock

GetObjectRetention

Retrieves an object's retention settings. For more information, see Locking Objects.

GetObjectTagging

Returns the tag-set of an object. You send the GET request against the tagging

GetObjectTorrent

Returns torrent files from a bucket. BitTorrent can save you bandwidth when you're

GetPublicAccessBlock

Retrieves the PublicAccessBlock configuration for an Amazon S3 bucket. To use

HeadBucket

This action is useful to determine if a bucket exists and you have permission to

HeadObject

The HEAD action retrieves metadata from an object without returning the object

ListBucketAnalyticsConfigurations

Lists the analytics configurations for the bucket. You can have up to 1,000 analytics

ListBucketIntelligentTieringConfigurations

Lists the S3 Intelligent-Tiering configuration from the specified bucket.

ListBucketInventoryConfigurations

Returns a list of inventory configurations for the bucket. You can have up to 1,000

ListBucketMetricsConfigurations

Lists the metrics configurations for the bucket. The metrics configurations are only for

ListBuckets

Returns a list of all buckets owned by the authenticated sender of the request.

ListMultipartUploads

This action lists in-progress multipart uploads. An in-progress multipart upload is a

ListObjectVersions

Returns metadata about all versions of the objects in a bucket. You can also use request

ListObjects

Returns some or all (up to 1,000) of the objects in a bucket. You can use the request

ListObjectsV2

Returns some or all (up to 1,000) of the objects in a bucket with each request. You can use

ListParts

Lists the parts that have been uploaded for a specific multipart upload. This operation

PutBucketAccelerateConfiguration

Sets the accelerate configuration of an existing bucket. Amazon S3 Transfer Acceleration is a

PutBucketAcl

Sets the permissions on an existing bucket using access control lists (ACL). For more

PutBucketAnalyticsConfiguration

Sets an analytics configuration for the bucket (specified by the analytics configuration

PutBucketCors

Sets the cors configuration for your bucket. If the configuration exists,

PutBucketEncryption

This action uses the encryption subresource to configure default

PutBucketIntelligentTieringConfiguration

Puts a S3 Intelligent-Tiering configuration to the specified bucket.

PutBucketInventoryConfiguration

This implementation of the PUT action adds an inventory configuration

PutBucketLifecycleConfiguration

Creates a new lifecycle configuration for the bucket or replaces an existing lifecycle

PutBucketLogging

Set the logging parameters for a bucket and to specify permissions for who can view and

PutBucketMetricsConfiguration

Sets a metrics configuration (specified by the metrics configuration ID) for the bucket.

PutBucketNotificationConfiguration

Enables notifications of specified events for a bucket. For more information about event

PutBucketOwnershipControls

Creates or modifies OwnershipControls for an Amazon S3 bucket. To use this

PutBucketPolicy

Applies an Amazon S3 bucket policy to an Amazon S3 bucket. If you are using an identity other than

PutBucketReplication

Creates a replication configuration or replaces an existing one. For more information,

PutBucketRequestPayment

Sets the request payment configuration for a bucket. By default, the bucket owner pays

PutBucketTagging

Sets the tags for a bucket.

PutBucketVersioning

Sets the versioning state of an existing bucket. To set the versioning state, you must

PutBucketWebsite

Sets the configuration of the website that is specified in the website

PutObject

Adds an object to a bucket. You must have WRITE permissions on a bucket to add an object

PutObjectAcl

Uses the acl subresource to set the access control list (ACL) permissions

PutObjectLegalHold

Applies a Legal Hold configuration to the specified object. For more information, see

PutObjectLockConfiguration

Places an Object Lock configuration on the specified bucket. The rule specified in the

PutObjectRetention

Places an Object Retention configuration on an object. For more information, see Locking Objects.

PutObjectTagging

Sets the supplied tag-set to an object that already exists in a bucket.

PutPublicAccessBlock

Creates or modifies the PublicAccessBlock configuration for an Amazon S3 bucket.

RestoreObject

Restores an archived copy of an object back into Amazon S3

UploadPart

Uploads a part in a multipart upload.

UploadPartCopy

Uploads a part by copying data from an existing object as data source. You specify the

WriteGetObjectResponse

Passes transformed