aws_sdk_s3::operation

Module get_bucket_notification_configuration

Types for the GetBucketNotificationConfiguration operation.

Modules

builders

Builders

Structs

GetBucketNotificationConfiguration

Orchestration and serialization glue logic for GetBucketNotificationConfiguration.

GetBucketNotificationConfigurationInput

GetBucketNotificationConfigurationOutput

A container for specifying the notification configuration of the bucket. If this element is empty, notifications are turned off for the bucket.

Enums

GetBucketNotificationConfigurationError

Error type for the GetBucketNotificationConfigurationError operation.