aws_sdk_s3::operation::get_object::builders

Struct GetObjectOutputBuilder

#[non_exhaustive]
pub struct GetObjectOutputBuilder { /* private fields */ }

A builder for GetObjectOutput.

Implementations

impl GetObjectOutputBuilder

pub fn body(self, input: ByteStream) -> Self

Object data.

pub fn set_body(self, input: Option<ByteStream>) -> Self

Object data.

pub fn get_body(&self) -> &Option<ByteStream>

Object data.

pub fn delete_marker(self, input: bool) -> Self

Indicates whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

If the current version of the object is a delete marker, Amazon S3 behaves as if the object was deleted and includes x-amz-delete-marker: true in the response.

If the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

pub fn set_delete_marker(self, input: Option<bool>) -> Self

Indicates whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

If the current version of the object is a delete marker, Amazon S3 behaves as if the object was deleted and includes x-amz-delete-marker: true in the response.

If the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

pub fn get_delete_marker(&self) -> &Option<bool>

Indicates whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

If the current version of the object is a delete marker, Amazon S3 behaves as if the object was deleted and includes x-amz-delete-marker: true in the response.

If the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

pub fn accept_ranges(self, input: impl Into<String>) -> Self

Indicates that a range of bytes was specified in the request.

pub fn set_accept_ranges(self, input: Option<String>) -> Self

Indicates that a range of bytes was specified in the request.

pub fn get_accept_ranges(&self) -> &Option<String>

Indicates that a range of bytes was specified in the request.

pub fn expiration(self, input: impl Into<String>) -> Self

If the object expiration is configured (see PutBucketLifecycleConfiguration ), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL-encoded.

Object expiration information is not returned in directory buckets and this header returns the value "NotImplemented" in all responses for directory buckets.

pub fn set_expiration(self, input: Option<String>) -> Self

If the object expiration is configured (see PutBucketLifecycleConfiguration ), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL-encoded.

Object expiration information is not returned in directory buckets and this header returns the value "NotImplemented" in all responses for directory buckets.

pub fn get_expiration(&self) -> &Option<String>

If the object expiration is configured (see PutBucketLifecycleConfiguration ), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL-encoded.

Object expiration information is not returned in directory buckets and this header returns the value "NotImplemented" in all responses for directory buckets.

pub fn restore(self, input: impl Into<String>) -> Self

Provides information about object restoration action and expiration time of the restored object copy.

This functionality is not supported for directory buckets. Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn set_restore(self, input: Option<String>) -> Self

Provides information about object restoration action and expiration time of the restored object copy.

This functionality is not supported for directory buckets. Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn get_restore(&self) -> &Option<String>

Provides information about object restoration action and expiration time of the restored object copy.

This functionality is not supported for directory buckets. Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn last_modified(self, input: DateTime) -> Self

Date and time when the object was last modified.

General purpose buckets - When you specify a versionId of the object in your request, if the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

pub fn set_last_modified(self, input: Option<DateTime>) -> Self

Date and time when the object was last modified.

General purpose buckets - When you specify a versionId of the object in your request, if the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

pub fn get_last_modified(&self) -> &Option<DateTime>

Date and time when the object was last modified.

General purpose buckets - When you specify a versionId of the object in your request, if the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

pub fn content_length(self, input: i64) -> Self

Size of the body in bytes.

pub fn set_content_length(self, input: Option<i64>) -> Self

Size of the body in bytes.

pub fn get_content_length(&self) -> &Option<i64>

Size of the body in bytes.

pub fn e_tag(self, input: impl Into<String>) -> Self

An entity tag (ETag) is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn set_e_tag(self, input: Option<String>) -> Self

An entity tag (ETag) is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn get_e_tag(&self) -> &Option<String>

An entity tag (ETag) is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn checksum_crc32(self, input: impl Into<String>) -> Self

The Base64 encoded, 32-bit CRC32 checksum of the object. This checksum is only present if the object was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn set_checksum_crc32(self, input: Option<String>) -> Self

The Base64 encoded, 32-bit CRC32 checksum of the object. This checksum is only present if the object was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn get_checksum_crc32(&self) -> &Option<String>

The Base64 encoded, 32-bit CRC32 checksum of the object. This checksum is only present if the object was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_crc32_c(self, input: impl Into<String>) -> Self

The Base64 encoded, 32-bit CRC32C checksum of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn set_checksum_crc32_c(self, input: Option<String>) -> Self

The Base64 encoded, 32-bit CRC32C checksum of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn get_checksum_crc32_c(&self) -> &Option<String>

The Base64 encoded, 32-bit CRC32C checksum of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_crc64_nvme(self, input: impl Into<String>) -> Self

The Base64 encoded, 64-bit CRC64NVME checksum of the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn set_checksum_crc64_nvme(self, input: Option<String>) -> Self

The Base64 encoded, 64-bit CRC64NVME checksum of the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn get_checksum_crc64_nvme(&self) -> &Option<String>

The Base64 encoded, 64-bit CRC64NVME checksum of the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_sha1(self, input: impl Into<String>) -> Self

The Base64 encoded, 160-bit SHA1 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn set_checksum_sha1(self, input: Option<String>) -> Self

The Base64 encoded, 160-bit SHA1 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn get_checksum_sha1(&self) -> &Option<String>

The Base64 encoded, 160-bit SHA1 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_sha256(self, input: impl Into<String>) -> Self

The Base64 encoded, 256-bit SHA256 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn set_checksum_sha256(self, input: Option<String>) -> Self

The Base64 encoded, 256-bit SHA256 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn get_checksum_sha256(&self) -> &Option<String>

The Base64 encoded, 256-bit SHA256 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_type(self, input: ChecksumType) -> Self

The checksum type, which determines how part-level checksums are combined to create an object-level checksum for multipart objects. You can use this header response to verify that the checksum type that is received is the same checksum type that was specified in the CreateMultipartUpload request. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn set_checksum_type(self, input: Option<ChecksumType>) -> Self

The checksum type, which determines how part-level checksums are combined to create an object-level checksum for multipart objects. You can use this header response to verify that the checksum type that is received is the same checksum type that was specified in the CreateMultipartUpload request. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn get_checksum_type(&self) -> &Option<ChecksumType>

The checksum type, which determines how part-level checksums are combined to create an object-level checksum for multipart objects. You can use this header response to verify that the checksum type that is received is the same checksum type that was specified in the CreateMultipartUpload request. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn missing_meta(self, input: i32) -> Self

This is set to the number of metadata entries not returned in the headers that are prefixed with x-amz-meta-. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

This functionality is not supported for directory buckets.

pub fn set_missing_meta(self, input: Option<i32>) -> Self

This is set to the number of metadata entries not returned in the headers that are prefixed with x-amz-meta-. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

This functionality is not supported for directory buckets.

pub fn get_missing_meta(&self) -> &Option<i32>

This is set to the number of metadata entries not returned in the headers that are prefixed with x-amz-meta-. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

This functionality is not supported for directory buckets.

pub fn version_id(self, input: impl Into<String>) -> Self

Version ID of the object.

This functionality is not supported for directory buckets.

pub fn set_version_id(self, input: Option<String>) -> Self

Version ID of the object.

This functionality is not supported for directory buckets.

pub fn get_version_id(&self) -> &Option<String>

Version ID of the object.

This functionality is not supported for directory buckets.

pub fn cache_control(self, input: impl Into<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn set_cache_control(self, input: Option<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn get_cache_control(&self) -> &Option<String>

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(self, input: impl Into<String>) -> Self

Specifies presentational information for the object.

pub fn set_content_disposition(self, input: Option<String>) -> Self

Specifies presentational information for the object.

pub fn get_content_disposition(&self) -> &Option<String>

Specifies presentational information for the object.

pub fn content_encoding(self, input: impl Into<String>) -> Self

Indicates what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn set_content_encoding(self, input: Option<String>) -> Self

Indicates what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn get_content_encoding(&self) -> &Option<String>

Indicates what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(self, input: impl Into<String>) -> Self

The language the content is in.

pub fn set_content_language(self, input: Option<String>) -> Self

The language the content is in.

pub fn get_content_language(&self) -> &Option<String>

The language the content is in.

pub fn content_range(self, input: impl Into<String>) -> Self

The portion of the object returned in the response.

pub fn set_content_range(self, input: Option<String>) -> Self

The portion of the object returned in the response.

pub fn get_content_range(&self) -> &Option<String>

The portion of the object returned in the response.

pub fn content_type(self, input: impl Into<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn set_content_type(self, input: Option<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn get_content_type(&self) -> &Option<String>

A standard MIME type describing the format of the object data.

pub fn website_redirect_location(self, input: impl Into<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

This functionality is not supported for directory buckets.

pub fn set_website_redirect_location(self, input: Option<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

This functionality is not supported for directory buckets.

pub fn get_website_redirect_location(&self) -> &Option<String>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

This functionality is not supported for directory buckets.

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

The server-side encryption algorithm used when you store this object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

pub fn set_server_side_encryption( self, input: Option<ServerSideEncryption>, ) -> Self

The server-side encryption algorithm used when you store this object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

pub fn get_server_side_encryption(&self) -> &Option<ServerSideEncryption>

The server-side encryption algorithm used when you store this object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

pub fn metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to metadata.

To override the contents of this collection use set_metadata.

A map of metadata to store with the object in S3.

pub fn set_metadata(self, input: Option<HashMap<String, String>>) -> Self

A map of metadata to store with the object in S3.

pub fn get_metadata(&self) -> &Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to confirm the encryption algorithm that's used.

This functionality is not supported for directory buckets.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to confirm the encryption algorithm that's used.

This functionality is not supported for directory buckets.

pub fn get_sse_customer_algorithm(&self) -> &Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to confirm the encryption algorithm that's used.

This functionality is not supported for directory buckets.

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide the round-trip message integrity verification of the customer-provided encryption key.

This functionality is not supported for directory buckets.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide the round-trip message integrity verification of the customer-provided encryption key.

This functionality is not supported for directory buckets.

pub fn get_sse_customer_key_md5(&self) -> &Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide the round-trip message integrity verification of the customer-provided encryption key.

This functionality is not supported for directory buckets.

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

If present, indicates the ID of the KMS key that was used for object encryption.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

If present, indicates the ID of the KMS key that was used for object encryption.

pub fn get_ssekms_key_id(&self) -> &Option<String>

If present, indicates the ID of the KMS key that was used for object encryption.

pub fn bucket_key_enabled(self, input: bool) -> Self

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Key Management Service (KMS) keys (SSE-KMS).

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Key Management Service (KMS) keys (SSE-KMS).

pub fn get_bucket_key_enabled(&self) -> &Option<bool>

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Key Management Service (KMS) keys (SSE-KMS).

pub fn storage_class(self, input: StorageClass) -> Self

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

Directory buckets - Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Self

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

Directory buckets - Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn get_storage_class(&self) -> &Option<StorageClass>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

Directory buckets - Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn get_request_charged(&self) -> &Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn replication_status(self, input: ReplicationStatus) -> Self

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

This functionality is not supported for directory buckets.

pub fn set_replication_status(self, input: Option<ReplicationStatus>) -> Self

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

This functionality is not supported for directory buckets.

pub fn get_replication_status(&self) -> &Option<ReplicationStatus>

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

This functionality is not supported for directory buckets.

pub fn parts_count(self, input: i32) -> Self

The count of parts this object has. This value is only returned if you specify partNumber in your request and the object was uploaded as a multipart upload.

pub fn set_parts_count(self, input: Option<i32>) -> Self

The count of parts this object has. This value is only returned if you specify partNumber in your request and the object was uploaded as a multipart upload.

pub fn get_parts_count(&self) -> &Option<i32>

The count of parts this object has. This value is only returned if you specify partNumber in your request and the object was uploaded as a multipart upload.

pub fn tag_count(self, input: i32) -> Self

The number of tags, if any, on the object, when you have the relevant permission to read object tags.

You can use GetObjectTagging to retrieve the tag set associated with an object.

This functionality is not supported for directory buckets.

pub fn set_tag_count(self, input: Option<i32>) -> Self

The number of tags, if any, on the object, when you have the relevant permission to read object tags.

You can use GetObjectTagging to retrieve the tag set associated with an object.

This functionality is not supported for directory buckets.

pub fn get_tag_count(&self) -> &Option<i32>

The number of tags, if any, on the object, when you have the relevant permission to read object tags.

You can use GetObjectTagging to retrieve the tag set associated with an object.

This functionality is not supported for directory buckets.

pub fn object_lock_mode(self, input: ObjectLockMode) -> Self

The Object Lock mode that's currently in place for this object.

This functionality is not supported for directory buckets.

pub fn set_object_lock_mode(self, input: Option<ObjectLockMode>) -> Self

The Object Lock mode that's currently in place for this object.

This functionality is not supported for directory buckets.

pub fn get_object_lock_mode(&self) -> &Option<ObjectLockMode>

The Object Lock mode that's currently in place for this object.

This functionality is not supported for directory buckets.

pub fn object_lock_retain_until_date(self, input: DateTime) -> Self

The date and time when this object's Object Lock will expire.

This functionality is not supported for directory buckets.

pub fn set_object_lock_retain_until_date(self, input: Option<DateTime>) -> Self

The date and time when this object's Object Lock will expire.

This functionality is not supported for directory buckets.

pub fn get_object_lock_retain_until_date(&self) -> &Option<DateTime>

The date and time when this object's Object Lock will expire.

This functionality is not supported for directory buckets.

pub fn object_lock_legal_hold_status( self, input: ObjectLockLegalHoldStatus, ) -> Self

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

This functionality is not supported for directory buckets.

pub fn set_object_lock_legal_hold_status( self, input: Option<ObjectLockLegalHoldStatus>, ) -> Self

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

This functionality is not supported for directory buckets.

pub fn get_object_lock_legal_hold_status( &self, ) -> &Option<ObjectLockLegalHoldStatus>

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

This functionality is not supported for directory buckets.

pub fn expires(self, input: DateTime) -> Self

๐Deprecated: Please use expires_string which contains the raw, unparsed value of this field.

The date and time at which the object is no longer cacheable.

pub fn set_expires(self, input: Option<DateTime>) -> Self

๐Deprecated: Please use expires_string which contains the raw, unparsed value of this field.

The date and time at which the object is no longer cacheable.

pub fn get_expires(&self) -> &Option<DateTime>

๐Deprecated: Please use expires_string which contains the raw, unparsed value of this field.

The date and time at which the object is no longer cacheable.

pub fn expires_string(self, input: impl Into<String>) -> Self

The date and time at which the object is no longer cacheable.

pub fn set_expires_string(self, input: Option<String>) -> Self

The date and time at which the object is no longer cacheable.

pub fn get_expires_string(&self) -> &Option<String>

The date and time at which the object is no longer cacheable.

pub fn build(self) -> GetObjectOutput

Consumes the builder and constructs a GetObjectOutput.