aws_sdk_s3::operation::get_object

Struct GetObjectOutput

#[non_exhaustive]
pub struct GetObjectOutput {

    pub body: ByteStream,
    pub delete_marker: Option<bool>,
    pub accept_ranges: Option<String>,
    pub expiration: Option<String>,
    pub restore: Option<String>,
    pub last_modified: Option<DateTime>,
    pub content_length: Option<i64>,
    pub e_tag: Option<String>,
    pub checksum_crc32: Option<String>,
    pub checksum_crc32_c: Option<String>,
    pub checksum_crc64_nvme: Option<String>,
    pub checksum_sha1: Option<String>,
    pub checksum_sha256: Option<String>,
    pub checksum_type: Option<ChecksumType>,
    pub missing_meta: Option<i32>,
    pub version_id: Option<String>,
    pub cache_control: Option<String>,
    pub content_disposition: Option<String>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_range: Option<String>,
    pub content_type: Option<String>,
    pub website_redirect_location: Option<String>,
    pub server_side_encryption: Option<ServerSideEncryption>,
    pub metadata: Option<HashMap<String, String>>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub ssekms_key_id: Option<String>,
    pub bucket_key_enabled: Option<bool>,
    pub storage_class: Option<StorageClass>,
    pub request_charged: Option<RequestCharged>,
    pub replication_status: Option<ReplicationStatus>,
    pub parts_count: Option<i32>,
    pub tag_count: Option<i32>,
    pub object_lock_mode: Option<ObjectLockMode>,
    pub object_lock_retain_until_date: Option<DateTime>,
    pub object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>,
    pub expires: Option<DateTime>,
    pub expires_string: Option<String>,
    /* private fields */

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

body: ByteStream

Object data.

delete_marker: Option<bool>

Indicates whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

If the current version of the object is a delete marker, Amazon S3 behaves as if the object was deleted and includes x-amz-delete-marker: true in the response.

If the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

accept_ranges: Option<String>

Indicates that a range of bytes was specified in the request.

expiration: Option<String>

If the object expiration is configured (see PutBucketLifecycleConfiguration ), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL-encoded.

Object expiration information is not returned in directory buckets and this header returns the value "NotImplemented" in all responses for directory buckets.

restore: Option<String>

Provides information about object restoration action and expiration time of the restored object copy.

This functionality is not supported for directory buckets. Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

last_modified: Option<DateTime>

Date and time when the object was last modified.

General purpose buckets - When you specify a versionId of the object in your request, if the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

content_length: Option<i64>

Size of the body in bytes.

e_tag: Option<String>

An entity tag (ETag) is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

checksum_crc32: Option<String>

The Base64 encoded, 32-bit CRC32 checksum of the object. This checksum is only present if the object was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

checksum_crc32_c: Option<String>

The Base64 encoded, 32-bit CRC32C checksum of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

checksum_crc64_nvme: Option<String>

The Base64 encoded, 64-bit CRC64NVME checksum of the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

checksum_sha1: Option<String>

The Base64 encoded, 160-bit SHA1 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

checksum_sha256: Option<String>

The Base64 encoded, 256-bit SHA256 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

checksum_type: Option<ChecksumType>

The checksum type, which determines how part-level checksums are combined to create an object-level checksum for multipart objects. You can use this header response to verify that the checksum type that is received is the same checksum type that was specified in the CreateMultipartUpload request. For more information, see Checking object integrity in the Amazon S3 User Guide.

missing_meta: Option<i32>

This is set to the number of metadata entries not returned in the headers that are prefixed with x-amz-meta-. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

This functionality is not supported for directory buckets.

version_id: Option<String>

Version ID of the object.

This functionality is not supported for directory buckets.

cache_control: Option<String>

Specifies caching behavior along the request/reply chain.

content_disposition: Option<String>

Specifies presentational information for the object.

content_encoding: Option<String>

Indicates what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

content_language: Option<String>

The language the content is in.

content_range: Option<String>

The portion of the object returned in the response.

content_type: Option<String>

A standard MIME type describing the format of the object data.

website_redirect_location: Option<String>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

This functionality is not supported for directory buckets.

server_side_encryption: Option<ServerSideEncryption>

The server-side encryption algorithm used when you store this object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

metadata: Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

sse_customer_algorithm: Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to confirm the encryption algorithm that's used.

This functionality is not supported for directory buckets.

sse_customer_key_md5: Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide the round-trip message integrity verification of the customer-provided encryption key.

This functionality is not supported for directory buckets.

ssekms_key_id: Option<String>

If present, indicates the ID of the KMS key that was used for object encryption.

bucket_key_enabled: Option<bool>

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Key Management Service (KMS) keys (SSE-KMS).

storage_class: Option<StorageClass>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

Directory buckets - Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

request_charged: Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

replication_status: Option<ReplicationStatus>

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

This functionality is not supported for directory buckets.

parts_count: Option<i32>

The count of parts this object has. This value is only returned if you specify partNumber in your request and the object was uploaded as a multipart upload.

tag_count: Option<i32>

The number of tags, if any, on the object, when you have the relevant permission to read object tags.

You can use GetObjectTagging to retrieve the tag set associated with an object.

This functionality is not supported for directory buckets.

object_lock_mode: Option<ObjectLockMode>

The Object Lock mode that's currently in place for this object.

This functionality is not supported for directory buckets.

object_lock_retain_until_date: Option<DateTime>

The date and time when this object's Object Lock will expire.

This functionality is not supported for directory buckets.

object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

This functionality is not supported for directory buckets.

expires: Option<DateTime>
👎Deprecated: Please use expires_string which contains the raw, unparsed value of this field.

The date and time at which the object is no longer cacheable.

expires_string: Option<String>

The date and time at which the object is no longer cacheable.

Implementations

impl GetObjectOutput

pub fn body(&self) -> &ByteStream

Object data.

pub fn delete_marker(&self) -> Option<bool>

Indicates whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

If the current version of the object is a delete marker, Amazon S3 behaves as if the object was deleted and includes x-amz-delete-marker: true in the response.

If the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

pub fn accept_ranges(&self) -> Option<&str>

Indicates that a range of bytes was specified in the request.

pub fn expiration(&self) -> Option<&str>

If the object expiration is configured (see PutBucketLifecycleConfiguration ), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL-encoded.

Object expiration information is not returned in directory buckets and this header returns the value "NotImplemented" in all responses for directory buckets.

pub fn restore(&self) -> Option<&str>

Provides information about object restoration action and expiration time of the restored object copy.

This functionality is not supported for directory buckets. Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn last_modified(&self) -> Option<&DateTime>

Date and time when the object was last modified.

General purpose buckets - When you specify a versionId of the object in your request, if the specified version in the request is a delete marker, the response returns a 405 Method Not Allowed error and the Last-Modified: timestamp response header.

pub fn content_length(&self) -> Option<i64>

Size of the body in bytes.

pub fn e_tag(&self) -> Option<&str>

An entity tag (ETag) is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn checksum_crc32(&self) -> Option<&str>

The Base64 encoded, 32-bit CRC32 checksum of the object. This checksum is only present if the object was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_crc32_c(&self) -> Option<&str>

The Base64 encoded, 32-bit CRC32C checksum of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_crc64_nvme(&self) -> Option<&str>

The Base64 encoded, 64-bit CRC64NVME checksum of the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_sha1(&self) -> Option<&str>

The Base64 encoded, 160-bit SHA1 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_sha256(&self) -> Option<&str>

The Base64 encoded, 256-bit SHA256 digest of the object. This checksum is only present if the checksum was uploaded with the object. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_type(&self) -> Option<&ChecksumType>

The checksum type, which determines how part-level checksums are combined to create an object-level checksum for multipart objects. You can use this header response to verify that the checksum type that is received is the same checksum type that was specified in the CreateMultipartUpload request. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn missing_meta(&self) -> Option<i32>

This is set to the number of metadata entries not returned in the headers that are prefixed with x-amz-meta-. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

This functionality is not supported for directory buckets.

pub fn version_id(&self) -> Option<&str>

Version ID of the object.

This functionality is not supported for directory buckets.

pub fn cache_control(&self) -> Option<&str>

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(&self) -> Option<&str>

Specifies presentational information for the object.

pub fn content_encoding(&self) -> Option<&str>

Indicates what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(&self) -> Option<&str>

The language the content is in.

pub fn content_range(&self) -> Option<&str>

The portion of the object returned in the response.

pub fn content_type(&self) -> Option<&str>

A standard MIME type describing the format of the object data.

pub fn website_redirect_location(&self) -> Option<&str>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

This functionality is not supported for directory buckets.

pub fn server_side_encryption(&self) -> Option<&ServerSideEncryption>

The server-side encryption algorithm used when you store this object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

pub fn metadata(&self) -> Option<&HashMap<String, String>>

A map of metadata to store with the object in S3.

pub fn sse_customer_algorithm(&self) -> Option<&str>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to confirm the encryption algorithm that's used.

This functionality is not supported for directory buckets.

pub fn sse_customer_key_md5(&self) -> Option<&str>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide the round-trip message integrity verification of the customer-provided encryption key.

This functionality is not supported for directory buckets.

pub fn ssekms_key_id(&self) -> Option<&str>

If present, indicates the ID of the KMS key that was used for object encryption.

pub fn bucket_key_enabled(&self) -> Option<bool>

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Key Management Service (KMS) keys (SSE-KMS).

pub fn storage_class(&self) -> Option<&StorageClass>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

Directory buckets - Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn request_charged(&self) -> Option<&RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn replication_status(&self) -> Option<&ReplicationStatus>

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

This functionality is not supported for directory buckets.

pub fn parts_count(&self) -> Option<i32>

The count of parts this object has. This value is only returned if you specify partNumber in your request and the object was uploaded as a multipart upload.

pub fn tag_count(&self) -> Option<i32>

The number of tags, if any, on the object, when you have the relevant permission to read object tags.

You can use GetObjectTagging to retrieve the tag set associated with an object.

This functionality is not supported for directory buckets.

pub fn object_lock_mode(&self) -> Option<&ObjectLockMode>

The Object Lock mode that's currently in place for this object.

This functionality is not supported for directory buckets.

pub fn object_lock_retain_until_date(&self) -> Option<&DateTime>

The date and time when this object's Object Lock will expire.

This functionality is not supported for directory buckets.

pub fn object_lock_legal_hold_status( &self, ) -> Option<&ObjectLockLegalHoldStatus>

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

This functionality is not supported for directory buckets.

pub fn expires(&self) -> Option<&DateTime>

👎Deprecated: Please use expires_string which contains the raw, unparsed value of this field.

The date and time at which the object is no longer cacheable.

pub fn expires_string(&self) -> Option<&str>

The date and time at which the object is no longer cacheable.

impl GetObjectOutput

pub fn builder() -> GetObjectOutputBuilder

Creates a new builder-style object to manufacture GetObjectOutput.