aws_sdk_s3::operation::list_multipart_uploads::builders

Struct ListMultipartUploadsOutputBuilder

#[non_exhaustive]
pub struct ListMultipartUploadsOutputBuilder { /* private fields */ }

A builder for ListMultipartUploadsOutput.

Implementations

impl ListMultipartUploadsOutputBuilder

pub fn bucket(self, input: impl Into<String>) -> Self

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

pub fn set_bucket(self, input: Option<String>) -> Self

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

pub fn get_bucket(&self) -> &Option<String>

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

pub fn key_marker(self, input: impl Into<String>) -> Self

The key at or after which the listing began.

pub fn set_key_marker(self, input: Option<String>) -> Self

The key at or after which the listing began.

pub fn get_key_marker(&self) -> &Option<String>

The key at or after which the listing began.

pub fn upload_id_marker(self, input: impl Into<String>) -> Self

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored. Otherwise, any multipart uploads for a key equal to the key-marker might be included in the list only if they have an upload ID lexicographically greater than the specified upload-id-marker.

This functionality is not supported for directory buckets.

pub fn set_upload_id_marker(self, input: Option<String>) -> Self

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored. Otherwise, any multipart uploads for a key equal to the key-marker might be included in the list only if they have an upload ID lexicographically greater than the specified upload-id-marker.

This functionality is not supported for directory buckets.

pub fn get_upload_id_marker(&self) -> &Option<String>

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored. Otherwise, any multipart uploads for a key equal to the key-marker might be included in the list only if they have an upload ID lexicographically greater than the specified upload-id-marker.

This functionality is not supported for directory buckets.

pub fn next_key_marker(self, input: impl Into<String>) -> Self

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

pub fn set_next_key_marker(self, input: Option<String>) -> Self

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

pub fn get_next_key_marker(&self) -> &Option<String>

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

pub fn prefix(self, input: impl Into<String>) -> Self

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

pub fn set_prefix(self, input: Option<String>) -> Self

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

pub fn get_prefix(&self) -> &Option<String>

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

pub fn delimiter(self, input: impl Into<String>) -> Self

Contains the delimiter you specified in the request. If you don't specify a delimiter in your request, this element is absent from the response.

Directory buckets - For directory buckets, / is the only supported delimiter.

pub fn set_delimiter(self, input: Option<String>) -> Self

Contains the delimiter you specified in the request. If you don't specify a delimiter in your request, this element is absent from the response.

Directory buckets - For directory buckets, / is the only supported delimiter.

pub fn get_delimiter(&self) -> &Option<String>

Contains the delimiter you specified in the request. If you don't specify a delimiter in your request, this element is absent from the response.

Directory buckets - For directory buckets, / is the only supported delimiter.

pub fn next_upload_id_marker(self, input: impl Into<String>) -> Self

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

This functionality is not supported for directory buckets.

pub fn set_next_upload_id_marker(self, input: Option<String>) -> Self

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

This functionality is not supported for directory buckets.

pub fn get_next_upload_id_marker(&self) -> &Option<String>

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

This functionality is not supported for directory buckets.

pub fn max_uploads(self, input: i32) -> Self

Maximum number of multipart uploads that could have been included in the response.

pub fn set_max_uploads(self, input: Option<i32>) -> Self

Maximum number of multipart uploads that could have been included in the response.

pub fn get_max_uploads(&self) -> &Option<i32>

Maximum number of multipart uploads that could have been included in the response.

pub fn is_truncated(self, input: bool) -> Self

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

pub fn set_is_truncated(self, input: Option<bool>) -> Self

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

pub fn get_is_truncated(&self) -> &Option<bool>

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

pub fn uploads(self, input: MultipartUpload) -> Self

Appends an item to uploads.

To override the contents of this collection use set_uploads.

Container for elements related to a particular multipart upload. A response can contain zero or more Upload elements.

pub fn set_uploads(self, input: Option<Vec<MultipartUpload>>) -> Self

Container for elements related to a particular multipart upload. A response can contain zero or more Upload elements.

pub fn get_uploads(&self) -> &Option<Vec<MultipartUpload>>

Container for elements related to a particular multipart upload. A response can contain zero or more Upload elements.

pub fn common_prefixes(self, input: CommonPrefix) -> Self

Appends an item to common_prefixes.

To override the contents of this collection use set_common_prefixes.

If you specify a delimiter in the request, then the result returns each distinct key prefix containing the delimiter in a CommonPrefixes element. The distinct key prefixes are returned in the Prefix child element.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

pub fn set_common_prefixes(self, input: Option<Vec<CommonPrefix>>) -> Self

If you specify a delimiter in the request, then the result returns each distinct key prefix containing the delimiter in a CommonPrefixes element. The distinct key prefixes are returned in the Prefix child element.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

pub fn get_common_prefixes(&self) -> &Option<Vec<CommonPrefix>>

If you specify a delimiter in the request, then the result returns each distinct key prefix containing the delimiter in a CommonPrefixes element. The distinct key prefixes are returned in the Prefix child element.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

pub fn encoding_type(self, input: EncodingType) -> Self

Encoding type used by Amazon S3 to encode object keys in the response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, KeyMarker, Prefix, NextKeyMarker, Key.

pub fn set_encoding_type(self, input: Option<EncodingType>) -> Self

Encoding type used by Amazon S3 to encode object keys in the response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, KeyMarker, Prefix, NextKeyMarker, Key.

pub fn get_encoding_type(&self) -> &Option<EncodingType>

Encoding type used by Amazon S3 to encode object keys in the response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, KeyMarker, Prefix, NextKeyMarker, Key.

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn get_request_charged(&self) -> &Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn build(self) -> ListMultipartUploadsOutput

Consumes the builder and constructs a ListMultipartUploadsOutput.