aws_sdk_s3::operation::list_object_versions::builders

Struct ListObjectVersionsOutputBuilder

#[non_exhaustive]
pub struct ListObjectVersionsOutputBuilder { /* private fields */ }

A builder for ListObjectVersionsOutput.

Implementations

impl ListObjectVersionsOutputBuilder

pub fn is_truncated(self, input: bool) -> Self

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request by using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

pub fn set_is_truncated(self, input: Option<bool>) -> Self

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request by using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

pub fn get_is_truncated(&self) -> &Option<bool>

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request by using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

pub fn key_marker(self, input: impl Into<String>) -> Self

Marks the last key returned in a truncated response.

pub fn set_key_marker(self, input: Option<String>) -> Self

Marks the last key returned in a truncated response.

pub fn get_key_marker(&self) -> &Option<String>

Marks the last key returned in a truncated response.

pub fn version_id_marker(self, input: impl Into<String>) -> Self

Marks the last version of the key returned in a truncated response.

pub fn set_version_id_marker(self, input: Option<String>) -> Self

Marks the last version of the key returned in a truncated response.

pub fn get_version_id_marker(&self) -> &Option<String>

Marks the last version of the key returned in a truncated response.

pub fn next_key_marker(self, input: impl Into<String>) -> Self

When the number of responses exceeds the value of MaxKeys, NextKeyMarker specifies the first key not returned that satisfies the search criteria. Use this value for the key-marker request parameter in a subsequent request.

pub fn set_next_key_marker(self, input: Option<String>) -> Self

When the number of responses exceeds the value of MaxKeys, NextKeyMarker specifies the first key not returned that satisfies the search criteria. Use this value for the key-marker request parameter in a subsequent request.

pub fn get_next_key_marker(&self) -> &Option<String>

When the number of responses exceeds the value of MaxKeys, NextKeyMarker specifies the first key not returned that satisfies the search criteria. Use this value for the key-marker request parameter in a subsequent request.

pub fn next_version_id_marker(self, input: impl Into<String>) -> Self

When the number of responses exceeds the value of MaxKeys, NextVersionIdMarker specifies the first object version not returned that satisfies the search criteria. Use this value for the version-id-marker request parameter in a subsequent request.

pub fn set_next_version_id_marker(self, input: Option<String>) -> Self

When the number of responses exceeds the value of MaxKeys, NextVersionIdMarker specifies the first object version not returned that satisfies the search criteria. Use this value for the version-id-marker request parameter in a subsequent request.

pub fn get_next_version_id_marker(&self) -> &Option<String>

When the number of responses exceeds the value of MaxKeys, NextVersionIdMarker specifies the first object version not returned that satisfies the search criteria. Use this value for the version-id-marker request parameter in a subsequent request.

pub fn versions(self, input: ObjectVersion) -> Self

Appends an item to versions.

To override the contents of this collection use set_versions.

Container for version information.

pub fn set_versions(self, input: Option<Vec<ObjectVersion>>) -> Self

Container for version information.

pub fn get_versions(&self) -> &Option<Vec<ObjectVersion>>

Container for version information.

pub fn delete_markers(self, input: DeleteMarkerEntry) -> Self

Appends an item to delete_markers.

To override the contents of this collection use set_delete_markers.

Container for an object that is a delete marker. To learn more about delete markers, see Working with delete markers.

pub fn set_delete_markers(self, input: Option<Vec<DeleteMarkerEntry>>) -> Self

Container for an object that is a delete marker. To learn more about delete markers, see Working with delete markers.

pub fn get_delete_markers(&self) -> &Option<Vec<DeleteMarkerEntry>>

Container for an object that is a delete marker. To learn more about delete markers, see Working with delete markers.

pub fn name(self, input: impl Into<String>) -> Self

The bucket name.

pub fn set_name(self, input: Option<String>) -> Self

The bucket name.

pub fn get_name(&self) -> &Option<String>

The bucket name.

pub fn prefix(self, input: impl Into<String>) -> Self

Selects objects that start with the value supplied by this parameter.

pub fn set_prefix(self, input: Option<String>) -> Self

Selects objects that start with the value supplied by this parameter.

pub fn get_prefix(&self) -> &Option<String>

Selects objects that start with the value supplied by this parameter.

pub fn delimiter(self, input: impl Into<String>) -> Self

The delimiter grouping the included keys. A delimiter is a character that you specify to group keys. All keys that contain the same string between the prefix and the first occurrence of the delimiter are grouped under a single result element in CommonPrefixes. These groups are counted as one result against the max-keys limitation. These keys are not returned elsewhere in the response.

pub fn set_delimiter(self, input: Option<String>) -> Self

The delimiter grouping the included keys. A delimiter is a character that you specify to group keys. All keys that contain the same string between the prefix and the first occurrence of the delimiter are grouped under a single result element in CommonPrefixes. These groups are counted as one result against the max-keys limitation. These keys are not returned elsewhere in the response.

pub fn get_delimiter(&self) -> &Option<String>

The delimiter grouping the included keys. A delimiter is a character that you specify to group keys. All keys that contain the same string between the prefix and the first occurrence of the delimiter are grouped under a single result element in CommonPrefixes. These groups are counted as one result against the max-keys limitation. These keys are not returned elsewhere in the response.

pub fn max_keys(self, input: i32) -> Self

Specifies the maximum number of objects to return.

pub fn set_max_keys(self, input: Option<i32>) -> Self

Specifies the maximum number of objects to return.

pub fn get_max_keys(&self) -> &Option<i32>

Specifies the maximum number of objects to return.

pub fn common_prefixes(self, input: CommonPrefix) -> Self

Appends an item to common_prefixes.

To override the contents of this collection use set_common_prefixes.

All of the keys rolled up into a common prefix count as a single return when calculating the number of returns.

pub fn set_common_prefixes(self, input: Option<Vec<CommonPrefix>>) -> Self

All of the keys rolled up into a common prefix count as a single return when calculating the number of returns.

pub fn get_common_prefixes(&self) -> &Option<Vec<CommonPrefix>>

All of the keys rolled up into a common prefix count as a single return when calculating the number of returns.

pub fn encoding_type(self, input: EncodingType) -> Self

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

KeyMarker, NextKeyMarker, Prefix, Key, and Delimiter.

pub fn set_encoding_type(self, input: Option<EncodingType>) -> Self

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

KeyMarker, NextKeyMarker, Prefix, Key, and Delimiter.

pub fn get_encoding_type(&self) -> &Option<EncodingType>

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

KeyMarker, NextKeyMarker, Prefix, Key, and Delimiter.

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn get_request_charged(&self) -> &Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn build(self) -> ListObjectVersionsOutput

Consumes the builder and constructs a ListObjectVersionsOutput.