aws_sdk_s3::operation::list_object_versions

Struct ListObjectVersionsOutput

#[non_exhaustive]
pub struct ListObjectVersionsOutput {

    pub is_truncated: Option<bool>,
    pub key_marker: Option<String>,
    pub version_id_marker: Option<String>,
    pub next_key_marker: Option<String>,
    pub next_version_id_marker: Option<String>,
    pub versions: Option<Vec<ObjectVersion>>,
    pub delete_markers: Option<Vec<DeleteMarkerEntry>>,
    pub name: Option<String>,
    pub prefix: Option<String>,
    pub delimiter: Option<String>,
    pub max_keys: Option<i32>,
    pub common_prefixes: Option<Vec<CommonPrefix>>,
    pub encoding_type: Option<EncodingType>,
    pub request_charged: Option<RequestCharged>,
    /* private fields */

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

is_truncated: Option<bool>

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request by using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

key_marker: Option<String>

Marks the last key returned in a truncated response.

version_id_marker: Option<String>

Marks the last version of the key returned in a truncated response.

next_key_marker: Option<String>

When the number of responses exceeds the value of MaxKeys, NextKeyMarker specifies the first key not returned that satisfies the search criteria. Use this value for the key-marker request parameter in a subsequent request.

next_version_id_marker: Option<String>

When the number of responses exceeds the value of MaxKeys, NextVersionIdMarker specifies the first object version not returned that satisfies the search criteria. Use this value for the version-id-marker request parameter in a subsequent request.

versions: Option<Vec<ObjectVersion>>

Container for version information.

delete_markers: Option<Vec<DeleteMarkerEntry>>

Container for an object that is a delete marker. To learn more about delete markers, see Working with delete markers.

name: Option<String>

The bucket name.

prefix: Option<String>

Selects objects that start with the value supplied by this parameter.

delimiter: Option<String>

The delimiter grouping the included keys. A delimiter is a character that you specify to group keys. All keys that contain the same string between the prefix and the first occurrence of the delimiter are grouped under a single result element in CommonPrefixes. These groups are counted as one result against the max-keys limitation. These keys are not returned elsewhere in the response.

max_keys: Option<i32>

Specifies the maximum number of objects to return.

common_prefixes: Option<Vec<CommonPrefix>>

All of the keys rolled up into a common prefix count as a single return when calculating the number of returns.

encoding_type: Option<EncodingType>

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

KeyMarker, NextKeyMarker, Prefix, Key, and Delimiter.

request_charged: Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

Implementations

impl ListObjectVersionsOutput

pub fn is_truncated(&self) -> Option<bool>

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request by using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

pub fn key_marker(&self) -> Option<&str>

Marks the last key returned in a truncated response.

pub fn version_id_marker(&self) -> Option<&str>

Marks the last version of the key returned in a truncated response.

pub fn next_key_marker(&self) -> Option<&str>

When the number of responses exceeds the value of MaxKeys, NextKeyMarker specifies the first key not returned that satisfies the search criteria. Use this value for the key-marker request parameter in a subsequent request.

pub fn next_version_id_marker(&self) -> Option<&str>

When the number of responses exceeds the value of MaxKeys, NextVersionIdMarker specifies the first object version not returned that satisfies the search criteria. Use this value for the version-id-marker request parameter in a subsequent request.

pub fn versions(&self) -> &[ObjectVersion]

Container for version information.

If no value was sent for this field, a default will be set. If you want to determine if no value was sent, use .versions.is_none().

pub fn delete_markers(&self) -> &[DeleteMarkerEntry]

Container for an object that is a delete marker. To learn more about delete markers, see Working with delete markers.

If no value was sent for this field, a default will be set. If you want to determine if no value was sent, use .delete_markers.is_none().

pub fn name(&self) -> Option<&str>

The bucket name.

pub fn prefix(&self) -> Option<&str>

Selects objects that start with the value supplied by this parameter.

pub fn delimiter(&self) -> Option<&str>

The delimiter grouping the included keys. A delimiter is a character that you specify to group keys. All keys that contain the same string between the prefix and the first occurrence of the delimiter are grouped under a single result element in CommonPrefixes. These groups are counted as one result against the max-keys limitation. These keys are not returned elsewhere in the response.

pub fn max_keys(&self) -> Option<i32>

Specifies the maximum number of objects to return.

pub fn common_prefixes(&self) -> &[CommonPrefix]

All of the keys rolled up into a common prefix count as a single return when calculating the number of returns.

If no value was sent for this field, a default will be set. If you want to determine if no value was sent, use .common_prefixes.is_none().

pub fn encoding_type(&self) -> Option<&EncodingType>

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

KeyMarker, NextKeyMarker, Prefix, Key, and Delimiter.

pub fn request_charged(&self) -> Option<&RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

impl ListObjectVersionsOutput

pub fn builder() -> ListObjectVersionsOutputBuilder

Creates a new builder-style object to manufacture ListObjectVersionsOutput.