aws_sdk_s3::operation::list_objects::builders

Struct ListObjectsOutputBuilder

#[non_exhaustive]
pub struct ListObjectsOutputBuilder { /* private fields */ }

A builder for ListObjectsOutput.

Implementations

impl ListObjectsOutputBuilder

pub fn is_truncated(self, input: bool) -> Self

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria.

pub fn set_is_truncated(self, input: Option<bool>) -> Self

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria.

pub fn get_is_truncated(&self) -> &Option<bool>

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria.

pub fn marker(self, input: impl Into<String>) -> Self

Indicates where in the bucket listing begins. Marker is included in the response if it was sent with the request.

pub fn set_marker(self, input: Option<String>) -> Self

Indicates where in the bucket listing begins. Marker is included in the response if it was sent with the request.

pub fn get_marker(&self) -> &Option<String>

Indicates where in the bucket listing begins. Marker is included in the response if it was sent with the request.

pub fn next_marker(self, input: impl Into<String>) -> Self

When the response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as the marker parameter in the subsequent request to get the next set of objects. Amazon S3 lists objects in alphabetical order.

This element is returned only if you have the delimiter request parameter specified. If the response does not include the NextMarker element and it is truncated, you can use the value of the last Key element in the response as the marker parameter in the subsequent request to get the next set of object keys.

pub fn set_next_marker(self, input: Option<String>) -> Self

When the response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as the marker parameter in the subsequent request to get the next set of objects. Amazon S3 lists objects in alphabetical order.

This element is returned only if you have the delimiter request parameter specified. If the response does not include the NextMarker element and it is truncated, you can use the value of the last Key element in the response as the marker parameter in the subsequent request to get the next set of object keys.

pub fn get_next_marker(&self) -> &Option<String>

When the response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as the marker parameter in the subsequent request to get the next set of objects. Amazon S3 lists objects in alphabetical order.

This element is returned only if you have the delimiter request parameter specified. If the response does not include the NextMarker element and it is truncated, you can use the value of the last Key element in the response as the marker parameter in the subsequent request to get the next set of object keys.

pub fn contents(self, input: Object) -> Self

Appends an item to contents.

To override the contents of this collection use set_contents.

Metadata about each object returned.

pub fn set_contents(self, input: Option<Vec<Object>>) -> Self

Metadata about each object returned.

pub fn get_contents(&self) -> &Option<Vec<Object>>

Metadata about each object returned.

pub fn name(self, input: impl Into<String>) -> Self

The bucket name.

pub fn set_name(self, input: Option<String>) -> Self

The bucket name.

pub fn get_name(&self) -> &Option<String>

The bucket name.

pub fn prefix(self, input: impl Into<String>) -> Self

Keys that begin with the indicated prefix.

pub fn set_prefix(self, input: Option<String>) -> Self

Keys that begin with the indicated prefix.

pub fn get_prefix(&self) -> &Option<String>

Keys that begin with the indicated prefix.

pub fn delimiter(self, input: impl Into<String>) -> Self

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

pub fn set_delimiter(self, input: Option<String>) -> Self

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

pub fn get_delimiter(&self) -> &Option<String>

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

pub fn max_keys(self, input: i32) -> Self

The maximum number of keys returned in the response body.

pub fn set_max_keys(self, input: Option<i32>) -> Self

The maximum number of keys returned in the response body.

pub fn get_max_keys(&self) -> &Option<i32>

The maximum number of keys returned in the response body.

pub fn common_prefixes(self, input: CommonPrefix) -> Self

Appends an item to common_prefixes.

To override the contents of this collection use set_common_prefixes.

All of the keys (up to 1,000) rolled up in a common prefix count as a single return when calculating the number of returns.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by the delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/), as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

pub fn set_common_prefixes(self, input: Option<Vec<CommonPrefix>>) -> Self

All of the keys (up to 1,000) rolled up in a common prefix count as a single return when calculating the number of returns.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by the delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/), as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

pub fn get_common_prefixes(&self) -> &Option<Vec<CommonPrefix>>

All of the keys (up to 1,000) rolled up in a common prefix count as a single return when calculating the number of returns.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by the delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/), as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

pub fn encoding_type(self, input: EncodingType) -> Self

Encoding type used by Amazon S3 to encode the object keys in the response. Responses are encoded only in UTF-8. An object key can contain any Unicode character. However, the XML 1.0 parser can't parse certain characters, such as characters with an ASCII value from 0 to 10. For characters that aren't supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response. For more information about characters to avoid in object key names, see Object key naming guidelines.

When using the URL encoding type, non-ASCII characters that are used in an object's key name will be percent-encoded according to UTF-8 code values. For example, the object test_file(3).png will appear as test_file%283%29.png.

pub fn set_encoding_type(self, input: Option<EncodingType>) -> Self

Encoding type used by Amazon S3 to encode the object keys in the response. Responses are encoded only in UTF-8. An object key can contain any Unicode character. However, the XML 1.0 parser can't parse certain characters, such as characters with an ASCII value from 0 to 10. For characters that aren't supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response. For more information about characters to avoid in object key names, see Object key naming guidelines.

When using the URL encoding type, non-ASCII characters that are used in an object's key name will be percent-encoded according to UTF-8 code values. For example, the object test_file(3).png will appear as test_file%283%29.png.

pub fn get_encoding_type(&self) -> &Option<EncodingType>

Encoding type used by Amazon S3 to encode the object keys in the response. Responses are encoded only in UTF-8. An object key can contain any Unicode character. However, the XML 1.0 parser can't parse certain characters, such as characters with an ASCII value from 0 to 10. For characters that aren't supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response. For more information about characters to avoid in object key names, see Object key naming guidelines.

When using the URL encoding type, non-ASCII characters that are used in an object's key name will be percent-encoded according to UTF-8 code values. For example, the object test_file(3).png will appear as test_file%283%29.png.

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn get_request_charged(&self) -> &Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn build(self) -> ListObjectsOutput

Consumes the builder and constructs a ListObjectsOutput.