aws_sdk_s3::operation::list_objects_v2

Struct ListObjectsV2Output

#[non_exhaustive]
pub struct ListObjectsV2Output {

    pub is_truncated: Option<bool>,
    pub contents: Option<Vec<Object>>,
    pub name: Option<String>,
    pub prefix: Option<String>,
    pub delimiter: Option<String>,
    pub max_keys: Option<i32>,
    pub common_prefixes: Option<Vec<CommonPrefix>>,
    pub encoding_type: Option<EncodingType>,
    pub key_count: Option<i32>,
    pub continuation_token: Option<String>,
    pub next_continuation_token: Option<String>,
    pub start_after: Option<String>,
    pub request_charged: Option<RequestCharged>,
    /* private fields */

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

is_truncated: Option<bool>

Set to false if all of the results were returned. Set to true if more keys are available to return. If the number of results exceeds that specified by MaxKeys, all of the results might not be returned.

contents: Option<Vec<Object>>

Metadata about each object returned.

name: Option<String>

The bucket name.

prefix: Option<String>

Keys that begin with the indicated prefix.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

delimiter: Option<String>

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

Directory buckets - For directory buckets, / is the only supported delimiter.

max_keys: Option<i32>

Sets the maximum number of keys returned in the response. By default, the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

common_prefixes: Option<Vec<CommonPrefix>>

All of the keys (up to 1,000) that share the same prefix are grouped together. When counting the total numbers of returns by this API operation, this group of keys is considered as one item.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by a delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/) as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

Directory buckets - When you query ListObjectsV2 with a delimiter during in-progress multipart uploads, the CommonPrefixes response parameter contains the prefixes that are associated with the in-progress multipart uploads. For more information about multipart uploads, see Multipart Upload Overview in the Amazon S3 User Guide.

encoding_type: Option<EncodingType>

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, Prefix, Key, and StartAfter.

key_count: Option<i32>

KeyCount is the number of keys returned with this request. KeyCount will always be less than or equal to the MaxKeys field. For example, if you ask for 50 keys, your result will include 50 keys or fewer.

continuation_token: Option<String>

If ContinuationToken was sent with the request, it is included in the response. You can use the returned ContinuationToken for pagination of the list response.

next_continuation_token: Option<String>

NextContinuationToken is sent when isTruncated is true, which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

start_after: Option<String>

If StartAfter was sent with the request, it is included in the response.

This functionality is not supported for directory buckets.

request_charged: Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

Implementations

impl ListObjectsV2Output

pub fn is_truncated(&self) -> Option<bool>

Set to false if all of the results were returned. Set to true if more keys are available to return. If the number of results exceeds that specified by MaxKeys, all of the results might not be returned.

pub fn contents(&self) -> &[Object]

Metadata about each object returned.

If no value was sent for this field, a default will be set. If you want to determine if no value was sent, use .contents.is_none().

pub fn name(&self) -> Option<&str>

The bucket name.

pub fn prefix(&self) -> Option<&str>

Keys that begin with the indicated prefix.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

pub fn delimiter(&self) -> Option<&str>

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

Directory buckets - For directory buckets, / is the only supported delimiter.

pub fn max_keys(&self) -> Option<i32>

Sets the maximum number of keys returned in the response. By default, the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

pub fn common_prefixes(&self) -> &[CommonPrefix]

All of the keys (up to 1,000) that share the same prefix are grouped together. When counting the total numbers of returns by this API operation, this group of keys is considered as one item.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by a delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/) as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

Directory buckets - For directory buckets, only prefixes that end in a delimiter (/) are supported.

Directory buckets - When you query ListObjectsV2 with a delimiter during in-progress multipart uploads, the CommonPrefixes response parameter contains the prefixes that are associated with the in-progress multipart uploads. For more information about multipart uploads, see Multipart Upload Overview in the Amazon S3 User Guide.

If no value was sent for this field, a default will be set. If you want to determine if no value was sent, use .common_prefixes.is_none().

pub fn encoding_type(&self) -> Option<&EncodingType>

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, Prefix, Key, and StartAfter.

pub fn key_count(&self) -> Option<i32>

KeyCount is the number of keys returned with this request. KeyCount will always be less than or equal to the MaxKeys field. For example, if you ask for 50 keys, your result will include 50 keys or fewer.

pub fn continuation_token(&self) -> Option<&str>

If ContinuationToken was sent with the request, it is included in the response. You can use the returned ContinuationToken for pagination of the list response.

pub fn next_continuation_token(&self) -> Option<&str>

NextContinuationToken is sent when isTruncated is true, which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

pub fn start_after(&self) -> Option<&str>

If StartAfter was sent with the request, it is included in the response.

This functionality is not supported for directory buckets.

pub fn request_charged(&self) -> Option<&RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

impl ListObjectsV2Output

pub fn builder() -> ListObjectsV2OutputBuilder

Creates a new builder-style object to manufacture ListObjectsV2Output.