aws_sdk_s3::operation::list_parts::builders

Struct ListPartsOutputBuilder

#[non_exhaustive]
pub struct ListPartsOutputBuilder { /* private fields */ }

A builder for ListPartsOutput.

Implementations

impl ListPartsOutputBuilder

pub fn abort_date(self, input: DateTime) -> Self

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, then the response includes this header indicating when the initiated multipart upload will become eligible for abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Configuration.

The response will also include the x-amz-abort-rule-id header that will provide the ID of the lifecycle configuration rule that defines this action.

This functionality is not supported for directory buckets.

pub fn set_abort_date(self, input: Option<DateTime>) -> Self

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, then the response includes this header indicating when the initiated multipart upload will become eligible for abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Configuration.

The response will also include the x-amz-abort-rule-id header that will provide the ID of the lifecycle configuration rule that defines this action.

This functionality is not supported for directory buckets.

pub fn get_abort_date(&self) -> &Option<DateTime>

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, then the response includes this header indicating when the initiated multipart upload will become eligible for abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Configuration.

The response will also include the x-amz-abort-rule-id header that will provide the ID of the lifecycle configuration rule that defines this action.

This functionality is not supported for directory buckets.

pub fn abort_rule_id(self, input: impl Into<String>) -> Self

This header is returned along with the x-amz-abort-date header. It identifies applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

This functionality is not supported for directory buckets.

pub fn set_abort_rule_id(self, input: Option<String>) -> Self

This header is returned along with the x-amz-abort-date header. It identifies applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

This functionality is not supported for directory buckets.

pub fn get_abort_rule_id(&self) -> &Option<String>

This header is returned along with the x-amz-abort-date header. It identifies applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

This functionality is not supported for directory buckets.

pub fn bucket(self, input: impl Into<String>) -> Self

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

pub fn set_bucket(self, input: Option<String>) -> Self

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

pub fn get_bucket(&self) -> &Option<String>

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

pub fn key(self, input: impl Into<String>) -> Self

Object key for which the multipart upload was initiated.

pub fn set_key(self, input: Option<String>) -> Self

Object key for which the multipart upload was initiated.

pub fn get_key(&self) -> &Option<String>

Object key for which the multipart upload was initiated.

pub fn upload_id(self, input: impl Into<String>) -> Self

Upload ID identifying the multipart upload whose parts are being listed.

pub fn set_upload_id(self, input: Option<String>) -> Self

Upload ID identifying the multipart upload whose parts are being listed.

pub fn get_upload_id(&self) -> &Option<String>

Upload ID identifying the multipart upload whose parts are being listed.

pub fn part_number_marker(self, input: impl Into<String>) -> Self

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

pub fn set_part_number_marker(self, input: Option<String>) -> Self

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

pub fn get_part_number_marker(&self) -> &Option<String>

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

pub fn next_part_number_marker(self, input: impl Into<String>) -> Self

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

pub fn set_next_part_number_marker(self, input: Option<String>) -> Self

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

pub fn get_next_part_number_marker(&self) -> &Option<String>

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

pub fn max_parts(self, input: i32) -> Self

Maximum number of parts that were allowed in the response.

pub fn set_max_parts(self, input: Option<i32>) -> Self

Maximum number of parts that were allowed in the response.

pub fn get_max_parts(&self) -> &Option<i32>

Maximum number of parts that were allowed in the response.

pub fn is_truncated(self, input: bool) -> Self

Indicates whether the returned list of parts is truncated. A true value indicates that the list was truncated. A list can be truncated if the number of parts exceeds the limit returned in the MaxParts element.

pub fn set_is_truncated(self, input: Option<bool>) -> Self

Indicates whether the returned list of parts is truncated. A true value indicates that the list was truncated. A list can be truncated if the number of parts exceeds the limit returned in the MaxParts element.

pub fn get_is_truncated(&self) -> &Option<bool>

Indicates whether the returned list of parts is truncated. A true value indicates that the list was truncated. A list can be truncated if the number of parts exceeds the limit returned in the MaxParts element.

pub fn parts(self, input: Part) -> Self

Appends an item to parts.

To override the contents of this collection use set_parts.

Container for elements related to a particular part. A response can contain zero or more Part elements.

pub fn set_parts(self, input: Option<Vec<Part>>) -> Self

Container for elements related to a particular part. A response can contain zero or more Part elements.

pub fn get_parts(&self) -> &Option<Vec<Part>>

Container for elements related to a particular part. A response can contain zero or more Part elements.

pub fn initiator(self, input: Initiator) -> Self

Container element that identifies who initiated the multipart upload. If the initiator is an Amazon Web Services account, this element provides the same information as the Owner element. If the initiator is an IAM User, this element provides the user ARN.

pub fn set_initiator(self, input: Option<Initiator>) -> Self

Container element that identifies who initiated the multipart upload. If the initiator is an Amazon Web Services account, this element provides the same information as the Owner element. If the initiator is an IAM User, this element provides the user ARN.

pub fn get_initiator(&self) -> &Option<Initiator>

Container element that identifies who initiated the multipart upload. If the initiator is an Amazon Web Services account, this element provides the same information as the Owner element. If the initiator is an IAM User, this element provides the user ARN.

pub fn owner(self, input: Owner) -> Self

Container element that identifies the object owner, after the object is created. If multipart upload is initiated by an IAM user, this element provides the parent account ID.

Directory buckets - The bucket owner is returned as the object owner for all the parts.

pub fn set_owner(self, input: Option<Owner>) -> Self

Container element that identifies the object owner, after the object is created. If multipart upload is initiated by an IAM user, this element provides the parent account ID.

Directory buckets - The bucket owner is returned as the object owner for all the parts.

pub fn get_owner(&self) -> &Option<Owner>

Container element that identifies the object owner, after the object is created. If multipart upload is initiated by an IAM user, this element provides the parent account ID.

Directory buckets - The bucket owner is returned as the object owner for all the parts.

pub fn storage_class(self, input: StorageClass) -> Self

The class of storage used to store the uploaded object.

Directory buckets - Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Self

The class of storage used to store the uploaded object.

Directory buckets - Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn get_storage_class(&self) -> &Option<StorageClass>

The class of storage used to store the uploaded object.

Directory buckets - Directory buckets only support EXPRESS_ONEZONE (the S3 Express One Zone storage class) in Availability Zones and ONEZONE_IA (the S3 One Zone-Infrequent Access storage class) in Dedicated Local Zones.

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn get_request_charged(&self) -> &Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn checksum_algorithm(self, input: ChecksumAlgorithm) -> Self

The algorithm that was used to create a checksum of the object.

pub fn set_checksum_algorithm(self, input: Option<ChecksumAlgorithm>) -> Self

The algorithm that was used to create a checksum of the object.

pub fn get_checksum_algorithm(&self) -> &Option<ChecksumAlgorithm>

The algorithm that was used to create a checksum of the object.

pub fn checksum_type(self, input: ChecksumType) -> Self

The checksum type, which determines how part-level checksums are combined to create an object-level checksum for multipart objects. You can use this header response to verify that the checksum type that is received is the same checksum type that was specified in CreateMultipartUpload request. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn set_checksum_type(self, input: Option<ChecksumType>) -> Self

The checksum type, which determines how part-level checksums are combined to create an object-level checksum for multipart objects. You can use this header response to verify that the checksum type that is received is the same checksum type that was specified in CreateMultipartUpload request. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn get_checksum_type(&self) -> &Option<ChecksumType>

The checksum type, which determines how part-level checksums are combined to create an object-level checksum for multipart objects. You can use this header response to verify that the checksum type that is received is the same checksum type that was specified in CreateMultipartUpload request. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn build(self) -> ListPartsOutput

Consumes the builder and constructs a ListPartsOutput.