aws_sdk_s3::operation::select_object_content

Struct SelectObjectContentInput

#[non_exhaustive]
pub struct SelectObjectContentInput {
    pub bucket: Option<String>,
    pub key: Option<String>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub expression: Option<String>,
    pub expression_type: Option<ExpressionType>,
    pub request_progress: Option<RequestProgress>,
    pub input_serialization: Option<InputSerialization>,
    pub output_serialization: Option<OutputSerialization>,
    pub scan_range: Option<ScanRange>,
    pub expected_bucket_owner: Option<String>,
}

Learn Amazon S3 Select is no longer available to new customers. Existing customers of Amazon S3 Select can continue to use the feature as usual. Learn more

Request to filter the contents of an Amazon S3 object based on a simple Structured Query Language (SQL) statement. In the request, along with the SQL expression, you must specify a data serialization format (JSON or CSV) of the object. Amazon S3 uses this to parse object data into records. It returns only records that match the specified SQL expression. You must also specify the data serialization format for the response. For more information, see S3Select API Documentation.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

bucket: Option<String>

The S3 bucket.

key: Option<String>

The object key.

sse_customer_algorithm: Option<String>

The server-side encryption (SSE) algorithm used to encrypt the object. This parameter is needed only when the object was created using a checksum algorithm. For more information, see Protecting data using SSE-C keys in the Amazon S3 User Guide.

sse_customer_key: Option<String>

The server-side encryption (SSE) customer managed key. This parameter is needed only when the object was created using a checksum algorithm. For more information, see Protecting data using SSE-C keys in the Amazon S3 User Guide.

sse_customer_key_md5: Option<String>

The MD5 server-side encryption (SSE) customer managed key. This parameter is needed only when the object was created using a checksum algorithm. For more information, see Protecting data using SSE-C keys in the Amazon S3 User Guide.

expression: Option<String>

The expression that is used to query the object.

expression_type: Option<ExpressionType>

The type of the provided expression (for example, SQL).

request_progress: Option<RequestProgress>

Specifies if periodic request progress information should be enabled.

input_serialization: Option<InputSerialization>

Describes the format of the data in the object that is being queried.

output_serialization: Option<OutputSerialization>

Describes the format of the data that you want Amazon S3 to return in response.

scan_range: Option<ScanRange>

Specifies the byte range of the object to get the records from. A record is processed when its first byte is contained by the range. This parameter is optional, but when specified, it must not be empty. See RFC 2616, Section 14.35.1 about how to specify the start and end of the range.

ScanRangemay be used in the following ways:

50 100 - process only the records starting between the bytes 50 and 100 (inclusive, counting from zero)

50 - process only the records starting after the byte 50

50 - process only the records within the last 50 bytes of the file.

expected_bucket_owner: Option<String>

The account ID of the expected bucket owner. If the account ID that you provide does not match the actual owner of the bucket, the request fails with the HTTP status code 403 Forbidden (access denied).

Implementations

impl SelectObjectContentInput

pub fn bucket(&self) -> Option<&str>

The S3 bucket.

pub fn key(&self) -> Option<&str>

The object key.

pub fn sse_customer_algorithm(&self) -> Option<&str>

The server-side encryption (SSE) algorithm used to encrypt the object. This parameter is needed only when the object was created using a checksum algorithm. For more information, see Protecting data using SSE-C keys in the Amazon S3 User Guide.

pub fn sse_customer_key(&self) -> Option<&str>

The server-side encryption (SSE) customer managed key. This parameter is needed only when the object was created using a checksum algorithm. For more information, see Protecting data using SSE-C keys in the Amazon S3 User Guide.

pub fn sse_customer_key_md5(&self) -> Option<&str>

The MD5 server-side encryption (SSE) customer managed key. This parameter is needed only when the object was created using a checksum algorithm. For more information, see Protecting data using SSE-C keys in the Amazon S3 User Guide.

pub fn expression(&self) -> Option<&str>

The expression that is used to query the object.

pub fn expression_type(&self) -> Option<&ExpressionType>

The type of the provided expression (for example, SQL).

pub fn request_progress(&self) -> Option<&RequestProgress>

Specifies if periodic request progress information should be enabled.

pub fn input_serialization(&self) -> Option<&InputSerialization>

Describes the format of the data in the object that is being queried.

pub fn output_serialization(&self) -> Option<&OutputSerialization>

Describes the format of the data that you want Amazon S3 to return in response.

pub fn scan_range(&self) -> Option<&ScanRange>

Specifies the byte range of the object to get the records from. A record is processed when its first byte is contained by the range. This parameter is optional, but when specified, it must not be empty. See RFC 2616, Section 14.35.1 about how to specify the start and end of the range.

ScanRangemay be used in the following ways:

50 100 - process only the records starting between the bytes 50 and 100 (inclusive, counting from zero)

50 - process only the records starting after the byte 50

50 - process only the records within the last 50 bytes of the file.

pub fn expected_bucket_owner(&self) -> Option<&str>

The account ID of the expected bucket owner. If the account ID that you provide does not match the actual owner of the bucket, the request fails with the HTTP status code 403 Forbidden (access denied).

impl SelectObjectContentInput

pub fn builder() -> SelectObjectContentInputBuilder

Creates a new builder-style object to manufacture SelectObjectContentInput.