aws_sdk_s3::operation::write_get_object_response::builders

Struct WriteGetObjectResponseInputBuilder

#[non_exhaustive]
pub struct WriteGetObjectResponseInputBuilder { /* private fields */ }

A builder for WriteGetObjectResponseInput.

Implementations

impl WriteGetObjectResponseInputBuilder

pub fn request_route(self, input: impl Into<String>) -> Self

Route prefix to the HTTP URL generated.
This field is required.

pub fn set_request_route(self, input: Option<String>) -> Self

Route prefix to the HTTP URL generated.

pub fn get_request_route(&self) -> &Option<String>

Route prefix to the HTTP URL generated.

pub fn request_token(self, input: impl Into<String>) -> Self

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.
This field is required.

pub fn set_request_token(self, input: Option<String>) -> Self

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

pub fn get_request_token(&self) -> &Option<String>

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

pub fn body(self, input: ByteStream) -> Self

The object data.

pub fn set_body(self, input: Option<ByteStream>) -> Self

The object data.

pub fn get_body(&self) -> &Option<ByteStream>

The object data.

pub fn status_code(self, input: i32) -> Self

The integer status code for an HTTP response of a corresponding GetObject request. The following is a list of status codes.

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

pub fn set_status_code(self, input: Option<i32>) -> Self

The integer status code for an HTTP response of a corresponding GetObject request. The following is a list of status codes.

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

pub fn get_status_code(&self) -> &Option<i32>

The integer status code for an HTTP response of a corresponding GetObject request. The following is a list of status codes.

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

pub fn error_code(self, input: impl Into<String>) -> Self

A string that uniquely identifies an error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. The regular expression (regex) value is "^\[A-Z\]\[a-zA-Z\]+$".

pub fn set_error_code(self, input: Option<String>) -> Self

A string that uniquely identifies an error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. The regular expression (regex) value is "^\[A-Z\]\[a-zA-Z\]+$".

pub fn get_error_code(&self) -> &Option<String>

A string that uniquely identifies an error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. The regular expression (regex) value is "^\[A-Z\]\[a-zA-Z\]+$".

pub fn error_message(self, input: impl Into<String>) -> Self

Contains a generic description of the error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

pub fn set_error_message(self, input: Option<String>) -> Self

Contains a generic description of the error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

pub fn get_error_message(&self) -> &Option<String>

Contains a generic description of the error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

pub fn accept_ranges(self, input: impl Into<String>) -> Self

Indicates that a range of bytes was specified.

pub fn set_accept_ranges(self, input: Option<String>) -> Self

Indicates that a range of bytes was specified.

pub fn get_accept_ranges(&self) -> &Option<String>

Indicates that a range of bytes was specified.

pub fn cache_control(self, input: impl Into<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn set_cache_control(self, input: Option<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn get_cache_control(&self) -> &Option<String>

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(self, input: impl Into<String>) -> Self

Specifies presentational information for the object.

pub fn set_content_disposition(self, input: Option<String>) -> Self

Specifies presentational information for the object.

pub fn get_content_disposition(&self) -> &Option<String>

Specifies presentational information for the object.

pub fn content_encoding(self, input: impl Into<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn set_content_encoding(self, input: Option<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn get_content_encoding(&self) -> &Option<String>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(self, input: impl Into<String>) -> Self

The language the content is in.

pub fn set_content_language(self, input: Option<String>) -> Self

The language the content is in.

pub fn get_content_language(&self) -> &Option<String>

The language the content is in.

pub fn content_length(self, input: i64) -> Self

The size of the content body in bytes.

pub fn set_content_length(self, input: Option<i64>) -> Self

The size of the content body in bytes.

pub fn get_content_length(&self) -> &Option<i64>

The size of the content body in bytes.

pub fn content_range(self, input: impl Into<String>) -> Self

The portion of the object returned in the response.

pub fn set_content_range(self, input: Option<String>) -> Self

The portion of the object returned in the response.

pub fn get_content_range(&self) -> &Option<String>

The portion of the object returned in the response.

pub fn content_type(self, input: impl Into<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn set_content_type(self, input: Option<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn get_content_type(&self) -> &Option<String>

A standard MIME type describing the format of the object data.

pub fn checksum_crc32(self, input: impl Into<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32 checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn set_checksum_crc32(self, input: Option<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32 checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn get_checksum_crc32(&self) -> &Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32 checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn checksum_crc32_c(self, input: impl Into<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32C checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn set_checksum_crc32_c(self, input: Option<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32C checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn get_checksum_crc32_c(&self) -> &Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32C checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn checksum_crc64_nvme(self, input: impl Into<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This header specifies the Base64 encoded, 64-bit CRC64NVME checksum of the part. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn set_checksum_crc64_nvme(self, input: Option<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This header specifies the Base64 encoded, 64-bit CRC64NVME checksum of the part. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn get_checksum_crc64_nvme(&self) -> &Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This header specifies the Base64 encoded, 64-bit CRC64NVME checksum of the part. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_sha1(self, input: impl Into<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 160-bit SHA1 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn set_checksum_sha1(self, input: Option<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 160-bit SHA1 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn get_checksum_sha1(&self) -> &Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 160-bit SHA1 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn checksum_sha256(self, input: impl Into<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 256-bit SHA256 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn set_checksum_sha256(self, input: Option<String>) -> Self

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 256-bit SHA256 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn get_checksum_sha256(&self) -> &Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 256-bit SHA256 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn delete_marker(self, input: bool) -> Self

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker. To learn more about delete markers, see Working with delete markers.

pub fn set_delete_marker(self, input: Option<bool>) -> Self

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker. To learn more about delete markers, see Working with delete markers.

pub fn get_delete_marker(&self) -> &Option<bool>

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker. To learn more about delete markers, see Working with delete markers.

pub fn e_tag(self, input: impl Into<String>) -> Self

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn set_e_tag(self, input: Option<String>) -> Self

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn get_e_tag(&self) -> &Option<String>

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn expires(self, input: DateTime) -> Self

The date and time at which the object is no longer cacheable.

pub fn set_expires(self, input: Option<DateTime>) -> Self

The date and time at which the object is no longer cacheable.

pub fn get_expires(&self) -> &Option<DateTime>

The date and time at which the object is no longer cacheable.

pub fn expiration(self, input: impl Into<String>) -> Self

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide the object expiration information. The value of the rule-id is URL-encoded.

pub fn set_expiration(self, input: Option<String>) -> Self

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide the object expiration information. The value of the rule-id is URL-encoded.

pub fn get_expiration(&self) -> &Option<String>

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide the object expiration information. The value of the rule-id is URL-encoded.

pub fn last_modified(self, input: DateTime) -> Self

The date and time that the object was last modified.

pub fn set_last_modified(self, input: Option<DateTime>) -> Self

The date and time that the object was last modified.

pub fn get_last_modified(&self) -> &Option<DateTime>

The date and time that the object was last modified.

pub fn missing_meta(self, input: i32) -> Self

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn set_missing_meta(self, input: Option<i32>) -> Self

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn get_missing_meta(&self) -> &Option<i32>

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to metadata.

To override the contents of this collection use set_metadata.

A map of metadata to store with the object in S3.

pub fn set_metadata(self, input: Option<HashMap<String, String>>) -> Self

A map of metadata to store with the object in S3.

pub fn get_metadata(&self) -> &Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

pub fn object_lock_mode(self, input: ObjectLockMode) -> Self

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

pub fn set_object_lock_mode(self, input: Option<ObjectLockMode>) -> Self

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

pub fn get_object_lock_mode(&self) -> &Option<ObjectLockMode>

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

pub fn object_lock_legal_hold_status( self, input: ObjectLockLegalHoldStatus, ) -> Self

Indicates whether an object stored in Amazon S3 has an active legal hold.

pub fn set_object_lock_legal_hold_status( self, input: Option<ObjectLockLegalHoldStatus>, ) -> Self

Indicates whether an object stored in Amazon S3 has an active legal hold.

pub fn get_object_lock_legal_hold_status( &self, ) -> &Option<ObjectLockLegalHoldStatus>

Indicates whether an object stored in Amazon S3 has an active legal hold.

pub fn object_lock_retain_until_date(self, input: DateTime) -> Self

The date and time when Object Lock is configured to expire.

pub fn set_object_lock_retain_until_date(self, input: Option<DateTime>) -> Self

The date and time when Object Lock is configured to expire.

pub fn get_object_lock_retain_until_date(&self) -> &Option<DateTime>

The date and time when Object Lock is configured to expire.

pub fn parts_count(self, input: i32) -> Self

The count of parts this object has.

pub fn set_parts_count(self, input: Option<i32>) -> Self

The count of parts this object has.

pub fn get_parts_count(&self) -> &Option<i32>

The count of parts this object has.

pub fn replication_status(self, input: ReplicationStatus) -> Self

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

pub fn set_replication_status(self, input: Option<ReplicationStatus>) -> Self

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

pub fn get_replication_status(&self) -> &Option<ReplicationStatus>

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn get_request_charged(&self) -> &Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn restore(self, input: impl Into<String>) -> Self

Provides information about object restoration operation and expiration time of the restored object copy.

pub fn set_restore(self, input: Option<String>) -> Self

Provides information about object restoration operation and expiration time of the restored object copy.

pub fn get_restore(&self) -> &Option<String>

Provides information about object restoration operation and expiration time of the restored object copy.

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

The server-side encryption algorithm used when storing requested object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

pub fn set_server_side_encryption( self, input: Option<ServerSideEncryption>, ) -> Self

The server-side encryption algorithm used when storing requested object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

pub fn get_server_side_encryption(&self) -> &Option<ServerSideEncryption>

The server-side encryption algorithm used when storing requested object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

pub fn get_sse_customer_algorithm(&self) -> &Option<String>

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

If present, specifies the ID (Key ID, Key ARN, or Key Alias) of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric encryption customer managed key that was used for stored in Amazon S3 object.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

If present, specifies the ID (Key ID, Key ARN, or Key Alias) of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric encryption customer managed key that was used for stored in Amazon S3 object.

pub fn get_ssekms_key_id(&self) -> &Option<String>

If present, specifies the ID (Key ID, Key ARN, or Key Alias) of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric encryption customer managed key that was used for stored in Amazon S3 object.

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

pub fn get_sse_customer_key_md5(&self) -> &Option<String>

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

pub fn storage_class(self, input: StorageClass) -> Self

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Self

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

pub fn get_storage_class(&self) -> &Option<StorageClass>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

pub fn tag_count(self, input: i32) -> Self

The number of tags, if any, on the object.

pub fn set_tag_count(self, input: Option<i32>) -> Self

The number of tags, if any, on the object.

pub fn get_tag_count(&self) -> &Option<i32>

The number of tags, if any, on the object.

pub fn version_id(self, input: impl Into<String>) -> Self

An ID used to reference a specific version of the object.

pub fn set_version_id(self, input: Option<String>) -> Self

An ID used to reference a specific version of the object.

pub fn get_version_id(&self) -> &Option<String>

An ID used to reference a specific version of the object.

pub fn bucket_key_enabled(self, input: bool) -> Self

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn get_bucket_key_enabled(&self) -> &Option<bool>

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn build(self) -> Result<WriteGetObjectResponseInput, BuildError>

Consumes the builder and constructs a WriteGetObjectResponseInput.

impl WriteGetObjectResponseInputBuilder

pub async fn send_with( self, client: &Client, ) -> Result<WriteGetObjectResponseOutput, SdkError<WriteGetObjectResponseError, HttpResponse>>

Sends a request with this input using the given client.