aws_sdk_s3::operation::write_get_object_response

Struct WriteGetObjectResponseInput

#[non_exhaustive]
pub struct WriteGetObjectResponseInput {

    pub request_route: Option<String>,
    pub request_token: Option<String>,
    pub body: ByteStream,
    pub status_code: Option<i32>,
    pub error_code: Option<String>,
    pub error_message: Option<String>,
    pub accept_ranges: Option<String>,
    pub cache_control: Option<String>,
    pub content_disposition: Option<String>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_length: Option<i64>,
    pub content_range: Option<String>,
    pub content_type: Option<String>,
    pub checksum_crc32: Option<String>,
    pub checksum_crc32_c: Option<String>,
    pub checksum_crc64_nvme: Option<String>,
    pub checksum_sha1: Option<String>,
    pub checksum_sha256: Option<String>,
    pub delete_marker: Option<bool>,
    pub e_tag: Option<String>,
    pub expires: Option<DateTime>,
    pub expiration: Option<String>,
    pub last_modified: Option<DateTime>,
    pub missing_meta: Option<i32>,
    pub metadata: Option<HashMap<String, String>>,
    pub object_lock_mode: Option<ObjectLockMode>,
    pub object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>,
    pub object_lock_retain_until_date: Option<DateTime>,
    pub parts_count: Option<i32>,
    pub replication_status: Option<ReplicationStatus>,
    pub request_charged: Option<RequestCharged>,
    pub restore: Option<String>,
    pub server_side_encryption: Option<ServerSideEncryption>,
    pub sse_customer_algorithm: Option<String>,
    pub ssekms_key_id: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub storage_class: Option<StorageClass>,
    pub tag_count: Option<i32>,
    pub version_id: Option<String>,
    pub bucket_key_enabled: Option<bool>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

request_route: Option<String>

Route prefix to the HTTP URL generated.

request_token: Option<String>

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

body: ByteStream

The object data.

status_code: Option<i32>

The integer status code for an HTTP response of a corresponding GetObject request. The following is a list of status codes.

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

error_code: Option<String>

A string that uniquely identifies an error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. The regular expression (regex) value is "^\[A-Z\]\[a-zA-Z\]+$".

error_message: Option<String>

Contains a generic description of the error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

accept_ranges: Option<String>

Indicates that a range of bytes was specified.

cache_control: Option<String>

Specifies caching behavior along the request/reply chain.

content_disposition: Option<String>

Specifies presentational information for the object.

content_encoding: Option<String>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

content_language: Option<String>

The language the content is in.

content_length: Option<i64>

The size of the content body in bytes.

content_range: Option<String>

The portion of the object returned in the response.

content_type: Option<String>

A standard MIME type describing the format of the object data.

checksum_crc32: Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32 checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

checksum_crc32_c: Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32C checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

checksum_crc64_nvme: Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This header specifies the Base64 encoded, 64-bit CRC64NVME checksum of the part. For more information, see Checking object integrity in the Amazon S3 User Guide.

checksum_sha1: Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 160-bit SHA1 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

checksum_sha256: Option<String>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 256-bit SHA256 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

delete_marker: Option<bool>

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker. To learn more about delete markers, see Working with delete markers.

e_tag: Option<String>

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

expires: Option<DateTime>

The date and time at which the object is no longer cacheable.

expiration: Option<String>

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide the object expiration information. The value of the rule-id is URL-encoded.

last_modified: Option<DateTime>

The date and time that the object was last modified.

missing_meta: Option<i32>

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

metadata: Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

object_lock_mode: Option<ObjectLockMode>

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>

Indicates whether an object stored in Amazon S3 has an active legal hold.

object_lock_retain_until_date: Option<DateTime>

The date and time when Object Lock is configured to expire.

parts_count: Option<i32>

The count of parts this object has.

replication_status: Option<ReplicationStatus>

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

request_charged: Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

restore: Option<String>

Provides information about object restoration operation and expiration time of the restored object copy.

server_side_encryption: Option<ServerSideEncryption>

The server-side encryption algorithm used when storing requested object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

sse_customer_algorithm: Option<String>

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

ssekms_key_id: Option<String>

If present, specifies the ID (Key ID, Key ARN, or Key Alias) of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric encryption customer managed key that was used for stored in Amazon S3 object.

sse_customer_key_md5: Option<String>

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

storage_class: Option<StorageClass>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

tag_count: Option<i32>

The number of tags, if any, on the object.

version_id: Option<String>

An ID used to reference a specific version of the object.

bucket_key_enabled: Option<bool>

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

Implementations

impl WriteGetObjectResponseInput

pub fn request_route(&self) -> Option<&str>

Route prefix to the HTTP URL generated.

pub fn request_token(&self) -> Option<&str>

A single use encrypted token that maps WriteGetObjectResponse to the end user GetObject request.

pub fn body(&self) -> &ByteStream

The object data.

pub fn status_code(&self) -> Option<i32>

The integer status code for an HTTP response of a corresponding GetObject request. The following is a list of status codes.

200 - OK

206 - Partial Content

304 - Not Modified

400 - Bad Request

401 - Unauthorized

403 - Forbidden

404 - Not Found

405 - Method Not Allowed

409 - Conflict

411 - Length Required

412 - Precondition Failed

416 - Range Not Satisfiable

500 - Internal Server Error

503 - Service Unavailable

pub fn error_code(&self) -> Option<&str>

A string that uniquely identifies an error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in the body. All error codes from S3 are sentence-cased. The regular expression (regex) value is "^\[A-Z\]\[a-zA-Z\]+$".

pub fn error_message(&self) -> Option<&str>

Contains a generic description of the error condition. Returned in the tag of the error XML response for a corresponding GetObject call. Cannot be used with a successful StatusCode header or when the transformed object is provided in body.

pub fn accept_ranges(&self) -> Option<&str>

Indicates that a range of bytes was specified.

pub fn cache_control(&self) -> Option<&str>

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(&self) -> Option<&str>

Specifies presentational information for the object.

pub fn content_encoding(&self) -> Option<&str>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(&self) -> Option<&str>

The language the content is in.

pub fn content_length(&self) -> Option<i64>

The size of the content body in bytes.

pub fn content_range(&self) -> Option<&str>

The portion of the object returned in the response.

pub fn content_type(&self) -> Option<&str>

A standard MIME type describing the format of the object data.

pub fn checksum_crc32(&self) -> Option<&str>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32 checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn checksum_crc32_c(&self) -> Option<&str>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 32-bit CRC32C checksum of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn checksum_crc64_nvme(&self) -> Option<&str>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This header specifies the Base64 encoded, 64-bit CRC64NVME checksum of the part. For more information, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_sha1(&self) -> Option<&str>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 160-bit SHA1 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn checksum_sha256(&self) -> Option<&str>

This header can be used as a data integrity check to verify that the data received is the same data that was originally sent. This specifies the Base64 encoded, 256-bit SHA256 digest of the object returned by the Object Lambda function. This may not match the checksum for the object stored in Amazon S3. Amazon S3 will perform validation of the checksum values only when the original GetObject request required checksum validation. For more information about checksums, see Checking object integrity in the Amazon S3 User Guide.

Only one checksum header can be specified at a time. If you supply multiple checksum headers, this request will fail.

pub fn delete_marker(&self) -> Option<bool>

Specifies whether an object stored in Amazon S3 is (true) or is not (false) a delete marker. To learn more about delete markers, see Working with delete markers.

pub fn e_tag(&self) -> Option<&str>

An opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn expires(&self) -> Option<&DateTime>

The date and time at which the object is no longer cacheable.

pub fn expiration(&self) -> Option<&str>

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide the object expiration information. The value of the rule-id is URL-encoded.

pub fn last_modified(&self) -> Option<&DateTime>

The date and time that the object was last modified.

pub fn missing_meta(&self) -> Option<i32>

Set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn metadata(&self) -> Option<&HashMap<String, String>>

A map of metadata to store with the object in S3.

pub fn object_lock_mode(&self) -> Option<&ObjectLockMode>

Indicates whether an object stored in Amazon S3 has Object Lock enabled. For more information about S3 Object Lock, see Object Lock.

pub fn object_lock_legal_hold_status( &self, ) -> Option<&ObjectLockLegalHoldStatus>

Indicates whether an object stored in Amazon S3 has an active legal hold.

pub fn object_lock_retain_until_date(&self) -> Option<&DateTime>

The date and time when Object Lock is configured to expire.

pub fn parts_count(&self) -> Option<i32>

The count of parts this object has.

pub fn replication_status(&self) -> Option<&ReplicationStatus>

Indicates if request involves bucket that is either a source or destination in a Replication rule. For more information about S3 Replication, see Replication.

pub fn request_charged(&self) -> Option<&RequestCharged>

If present, indicates that the requester was successfully charged for the request. For more information, see Using Requester Pays buckets for storage transfers and usage in the Amazon Simple Storage Service user guide.

This functionality is not supported for directory buckets.

pub fn restore(&self) -> Option<&str>

Provides information about object restoration operation and expiration time of the restored object copy.

pub fn server_side_encryption(&self) -> Option<&ServerSideEncryption>

The server-side encryption algorithm used when storing requested object in Amazon S3 or Amazon FSx.

When accessing data stored in Amazon FSx file systems using S3 access points, the only valid server side encryption option is aws:fsx.

pub fn sse_customer_algorithm(&self) -> Option<&str>

Encryption algorithm used if server-side encryption with a customer-provided encryption key was specified for object stored in Amazon S3.

pub fn ssekms_key_id(&self) -> Option<&str>

If present, specifies the ID (Key ID, Key ARN, or Key Alias) of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric encryption customer managed key that was used for stored in Amazon S3 object.

pub fn sse_customer_key_md5(&self) -> Option<&str>

128-bit MD5 digest of customer-provided encryption key used in Amazon S3 to encrypt data stored in S3. For more information, see Protecting data using server-side encryption with customer-provided encryption keys (SSE-C).

pub fn storage_class(&self) -> Option<&StorageClass>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

pub fn tag_count(&self) -> Option<i32>

The number of tags, if any, on the object.

pub fn version_id(&self) -> Option<&str>

An ID used to reference a specific version of the object.

pub fn bucket_key_enabled(&self) -> Option<bool>

Indicates whether the object stored in Amazon S3 uses an S3 bucket key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

impl WriteGetObjectResponseInput

pub fn builder() -> WriteGetObjectResponseInputBuilder

Creates a new builder-style object to manufacture WriteGetObjectResponseInput.