Module aws_sdk_s3::output

Output structures for operations.

Modules

abort_multipart_upload_output

See AbortMultipartUploadOutput

complete_multipart_upload_output

See CompleteMultipartUploadOutput

copy_object_output

See CopyObjectOutput

create_bucket_output

See CreateBucketOutput

create_multipart_upload_output

See CreateMultipartUploadOutput

delete_bucket_analytics_configuration_output

See DeleteBucketAnalyticsConfigurationOutput

delete_bucket_cors_output

See DeleteBucketCorsOutput

delete_bucket_encryption_output

See DeleteBucketEncryptionOutput

delete_bucket_intelligent_tiering_configuration_output

See DeleteBucketIntelligentTieringConfigurationOutput

delete_bucket_inventory_configuration_output

See DeleteBucketInventoryConfigurationOutput

delete_bucket_lifecycle_output

See DeleteBucketLifecycleOutput

delete_bucket_metrics_configuration_output

See DeleteBucketMetricsConfigurationOutput

delete_bucket_output

See DeleteBucketOutput

delete_bucket_ownership_controls_output

See DeleteBucketOwnershipControlsOutput

delete_bucket_policy_output

See DeleteBucketPolicyOutput

delete_bucket_replication_output

See DeleteBucketReplicationOutput

delete_bucket_tagging_output

See DeleteBucketTaggingOutput

delete_bucket_website_output

See DeleteBucketWebsiteOutput

delete_object_output

See DeleteObjectOutput

delete_object_tagging_output

See DeleteObjectTaggingOutput

delete_objects_output

See DeleteObjectsOutput

delete_public_access_block_output

See DeletePublicAccessBlockOutput

get_bucket_accelerate_configuration_output

See GetBucketAccelerateConfigurationOutput

get_bucket_acl_output

See GetBucketAclOutput

get_bucket_analytics_configuration_output

See GetBucketAnalyticsConfigurationOutput

get_bucket_cors_output

See GetBucketCorsOutput

get_bucket_encryption_output

See GetBucketEncryptionOutput

get_bucket_intelligent_tiering_configuration_output

See GetBucketIntelligentTieringConfigurationOutput

get_bucket_inventory_configuration_output

See GetBucketInventoryConfigurationOutput

get_bucket_lifecycle_configuration_output

See GetBucketLifecycleConfigurationOutput

get_bucket_location_output

See GetBucketLocationOutput

get_bucket_logging_output

See GetBucketLoggingOutput

get_bucket_metrics_configuration_output

See GetBucketMetricsConfigurationOutput

get_bucket_notification_configuration_output

See GetBucketNotificationConfigurationOutput

get_bucket_ownership_controls_output

See GetBucketOwnershipControlsOutput

get_bucket_policy_output

See GetBucketPolicyOutput

get_bucket_policy_status_output

See GetBucketPolicyStatusOutput

get_bucket_replication_output

See GetBucketReplicationOutput

get_bucket_request_payment_output

See GetBucketRequestPaymentOutput

get_bucket_tagging_output

See GetBucketTaggingOutput

get_bucket_versioning_output

See GetBucketVersioningOutput

get_bucket_website_output

See GetBucketWebsiteOutput

get_object_acl_output

See GetObjectAclOutput

get_object_legal_hold_output

See GetObjectLegalHoldOutput

get_object_lock_configuration_output

See GetObjectLockConfigurationOutput

get_object_output

See GetObjectOutput

get_object_retention_output

See GetObjectRetentionOutput

get_object_tagging_output

See GetObjectTaggingOutput

get_object_torrent_output

See GetObjectTorrentOutput

get_public_access_block_output

See GetPublicAccessBlockOutput

head_bucket_output

See HeadBucketOutput

head_object_output

See HeadObjectOutput

list_bucket_analytics_configurations_output

See ListBucketAnalyticsConfigurationsOutput

list_bucket_intelligent_tiering_configurations_output

See ListBucketIntelligentTieringConfigurationsOutput

list_bucket_inventory_configurations_output

See ListBucketInventoryConfigurationsOutput

list_bucket_metrics_configurations_output

See ListBucketMetricsConfigurationsOutput

list_buckets_output

See ListBucketsOutput

list_multipart_uploads_output

See ListMultipartUploadsOutput

list_object_versions_output

See ListObjectVersionsOutput

list_objects_output

See ListObjectsOutput

list_objects_v2_output

See ListObjectsV2Output

list_parts_output

See ListPartsOutput

put_bucket_accelerate_configuration_output

See PutBucketAccelerateConfigurationOutput

put_bucket_acl_output

See PutBucketAclOutput

put_bucket_analytics_configuration_output

See PutBucketAnalyticsConfigurationOutput

put_bucket_cors_output

See PutBucketCorsOutput

put_bucket_encryption_output

See PutBucketEncryptionOutput

put_bucket_intelligent_tiering_configuration_output

See PutBucketIntelligentTieringConfigurationOutput

put_bucket_inventory_configuration_output

See PutBucketInventoryConfigurationOutput

put_bucket_lifecycle_configuration_output

See PutBucketLifecycleConfigurationOutput

put_bucket_logging_output

See PutBucketLoggingOutput

put_bucket_metrics_configuration_output

See PutBucketMetricsConfigurationOutput

put_bucket_notification_configuration_output

See PutBucketNotificationConfigurationOutput

put_bucket_ownership_controls_output

See PutBucketOwnershipControlsOutput

put_bucket_policy_output

See PutBucketPolicyOutput

put_bucket_replication_output

See PutBucketReplicationOutput

put_bucket_request_payment_output

See PutBucketRequestPaymentOutput

put_bucket_tagging_output

See PutBucketTaggingOutput

put_bucket_versioning_output

See PutBucketVersioningOutput

put_bucket_website_output

See PutBucketWebsiteOutput

put_object_acl_output

See PutObjectAclOutput

put_object_legal_hold_output

See PutObjectLegalHoldOutput

put_object_lock_configuration_output

See PutObjectLockConfigurationOutput

put_object_output

See PutObjectOutput

put_object_retention_output

See PutObjectRetentionOutput

put_object_tagging_output

See PutObjectTaggingOutput

put_public_access_block_output

See PutPublicAccessBlockOutput

restore_object_output

See RestoreObjectOutput

select_object_content_output

See SelectObjectContentOutput

upload_part_copy_output

See UploadPartCopyOutput

upload_part_output

See UploadPartOutput

write_get_object_response_output

See WriteGetObjectResponseOutput

Structs

AbortMultipartUploadOutput

CompleteMultipartUploadOutput

CopyObjectOutput

CreateBucketOutput

CreateMultipartUploadOutput

DeleteBucketAnalyticsConfigurationOutput

DeleteBucketCorsOutput

DeleteBucketEncryptionOutput

DeleteBucketIntelligentTieringConfigurationOutput

DeleteBucketInventoryConfigurationOutput

DeleteBucketLifecycleOutput

DeleteBucketMetricsConfigurationOutput

DeleteBucketOutput

DeleteBucketOwnershipControlsOutput

DeleteBucketPolicyOutput

DeleteBucketReplicationOutput

DeleteBucketTaggingOutput

DeleteBucketWebsiteOutput

DeleteObjectOutput

DeleteObjectTaggingOutput

DeleteObjectsOutput

DeletePublicAccessBlockOutput

GetBucketAccelerateConfigurationOutput

GetBucketAclOutput

GetBucketAnalyticsConfigurationOutput

GetBucketCorsOutput

GetBucketEncryptionOutput

GetBucketIntelligentTieringConfigurationOutput

GetBucketInventoryConfigurationOutput

GetBucketLifecycleConfigurationOutput

GetBucketLocationOutput

GetBucketLoggingOutput

GetBucketMetricsConfigurationOutput

GetBucketNotificationConfigurationOutput

A container for specifying the notification configuration of the bucket. If this element is empty, notifications are turned off for the bucket.

GetBucketOwnershipControlsOutput

GetBucketPolicyOutput

GetBucketPolicyStatusOutput

GetBucketReplicationOutput

GetBucketRequestPaymentOutput

GetBucketTaggingOutput

GetBucketVersioningOutput

GetBucketWebsiteOutput

GetObjectAclOutput

GetObjectLegalHoldOutput

GetObjectLockConfigurationOutput

GetObjectOutput

GetObjectRetentionOutput

GetObjectTaggingOutput

GetObjectTorrentOutput

GetPublicAccessBlockOutput

HeadBucketOutput

HeadObjectOutput

ListBucketAnalyticsConfigurationsOutput

ListBucketIntelligentTieringConfigurationsOutput

ListBucketInventoryConfigurationsOutput

ListBucketMetricsConfigurationsOutput

ListBucketsOutput

ListMultipartUploadsOutput

ListObjectVersionsOutput

ListObjectsOutput

ListObjectsV2Output

ListPartsOutput

PutBucketAccelerateConfigurationOutput

PutBucketAclOutput

PutBucketAnalyticsConfigurationOutput

PutBucketCorsOutput

PutBucketEncryptionOutput

PutBucketIntelligentTieringConfigurationOutput

PutBucketInventoryConfigurationOutput

PutBucketLifecycleConfigurationOutput

PutBucketLoggingOutput

PutBucketMetricsConfigurationOutput

PutBucketNotificationConfigurationOutput

PutBucketOwnershipControlsOutput

PutBucketPolicyOutput

PutBucketReplicationOutput

PutBucketRequestPaymentOutput

PutBucketTaggingOutput

PutBucketVersioningOutput

PutBucketWebsiteOutput

PutObjectAclOutput

PutObjectLegalHoldOutput

PutObjectLockConfigurationOutput

PutObjectOutput

PutObjectRetentionOutput

PutObjectTaggingOutput

PutPublicAccessBlockOutput

RestoreObjectOutput

SelectObjectContentOutput

UploadPartCopyOutput

UploadPartOutput

WriteGetObjectResponseOutput