Struct aws_sdk_s3::output::create_multipart_upload_output::Builder

#[non_exhaustive]
pub struct Builder { /* private fields */ }

A builder for CreateMultipartUploadOutput

Implementations

impl Builder

pub fn abort_date(self, input: DateTime) -> Self

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, the response includes this header. The header indicates when the initiated multipart upload becomes eligible for an abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response also includes the x-amz-abort-rule-id header that provides the ID of the lifecycle configuration rule that defines this action.

pub fn set_abort_date(self, input: Option<DateTime>) -> Self

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, the response includes this header. The header indicates when the initiated multipart upload becomes eligible for an abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response also includes the x-amz-abort-rule-id header that provides the ID of the lifecycle configuration rule that defines this action.

pub fn abort_rule_id(self, input: impl Into<String>) -> Self

This header is returned along with the x-amz-abort-date header. It identifies the applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

pub fn set_abort_rule_id(self, input: Option<String>) -> Self

This header is returned along with the x-amz-abort-date header. It identifies the applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

pub fn bucket(self, input: impl Into<String>) -> Self

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_bucket(self, input: Option<String>) -> Self

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn key(self, input: impl Into<String>) -> Self

Object key for which the multipart upload was initiated.

pub fn set_key(self, input: Option<String>) -> Self

Object key for which the multipart upload was initiated.

pub fn upload_id(self, input: impl Into<String>) -> Self

ID for the initiated multipart upload.

pub fn set_upload_id(self, input: Option<String>) -> Self

ID for the initiated multipart upload.

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn set_server_side_encryption( self, input: Option<ServerSideEncryption> ) -> Self

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

pub fn ssekms_encryption_context(self, input: impl Into<String>) -> Self

If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

pub fn set_ssekms_encryption_context(self, input: Option<String>) -> Self

If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

pub fn bucket_key_enabled(self, input: bool) -> Self

Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

If present, indicates that the requester was successfully charged for the request.

pub fn build(self) -> CreateMultipartUploadOutput

Consumes the builder and constructs a CreateMultipartUploadOutput