Struct aws_sdk_s3::output::get_object_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for GetObjectOutput

Implementations

impl Builder

pub fn body(self, input: ByteStream) -> Self

Object data.

pub fn set_body(self, input: Option<ByteStream>) -> Self

pub fn delete_marker(self, input: bool) -> Self

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

pub fn set_delete_marker(self, input: Option<bool>) -> Self

pub fn accept_ranges(self, input: impl Into<String>) -> Self

Indicates that a range of bytes was specified.

pub fn set_accept_ranges(self, input: Option<String>) -> Self

pub fn expiration(self, input: impl Into<String>) -> Self

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

pub fn set_expiration(self, input: Option<String>) -> Self

pub fn restore(self, input: impl Into<String>) -> Self

Provides information about object restoration action and expiration time of the restored object copy.

pub fn set_restore(self, input: Option<String>) -> Self

pub fn last_modified(self, input: Instant) -> Self

Creation date of the object.

pub fn set_last_modified(self, input: Option<Instant>) -> Self

pub fn content_length(self, input: i64) -> Self

Size of the body in bytes.

pub fn set_content_length(self, input: Option<i64>) -> Self

pub fn e_tag(self, input: impl Into<String>) -> Self

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn set_e_tag(self, input: Option<String>) -> Self

pub fn missing_meta(self, input: i32) -> Self

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn set_missing_meta(self, input: Option<i32>) -> Self

pub fn version_id(self, input: impl Into<String>) -> Self

Version of the object.

pub fn set_version_id(self, input: Option<String>) -> Self

pub fn cache_control(self, input: impl Into<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn set_cache_control(self, input: Option<String>) -> Self

pub fn content_disposition(self, input: impl Into<String>) -> Self

Specifies presentational information for the object.

pub fn set_content_disposition(self, input: Option<String>) -> Self

pub fn content_encoding(self, input: impl Into<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn set_content_encoding(self, input: Option<String>) -> Self

pub fn content_language(self, input: impl Into<String>) -> Self

The language the content is in.

pub fn set_content_language(self, input: Option<String>) -> Self

pub fn content_range(self, input: impl Into<String>) -> Self

The portion of the object returned in the response.

pub fn set_content_range(self, input: Option<String>) -> Self

pub fn content_type(self, input: impl Into<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn set_content_type(self, input: Option<String>) -> Self

pub fn expires(self, input: Instant) -> Self

The date and time at which the object is no longer cacheable.

pub fn set_expires(self, input: Option<Instant>) -> Self

pub fn website_redirect_location(self, input: impl Into<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

pub fn set_website_redirect_location(self, input: Option<String>) -> Self

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn set_server_side_encryption(
self,
input: Option<ServerSideEncryption>
) -> Self

pub fn metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_metadata(self, input: Option<HashMap<String, String>>) -> Self

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

If present, specifies the ID of the AWS Key Management Service (AWS KMS) symmetric customer managed customer master key (CMK) that was used for the object.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

pub fn bucket_key_enabled(self, input: bool) -> Self

Indicates whether the object uses an S3 Bucket Key for server-side encryption with AWS KMS (SSE-KMS).

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

pub fn storage_class(self, input: StorageClass) -> Self

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Self

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

pub fn replication_status(self, input: ReplicationStatus) -> Self

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

pub fn set_replication_status(self, input: Option<ReplicationStatus>) -> Self

pub fn parts_count(self, input: i32) -> Self

The count of parts this object has.

pub fn set_parts_count(self, input: Option<i32>) -> Self

pub fn tag_count(self, input: i32) -> Self

The number of tags, if any, on the object.

pub fn set_tag_count(self, input: Option<i32>) -> Self

pub fn object_lock_mode(self, input: ObjectLockMode) -> Self

The Object Lock mode currently in place for this object.

pub fn set_object_lock_mode(self, input: Option<ObjectLockMode>) -> Self

pub fn object_lock_retain_until_date(self, input: Instant) -> Self

The date and time when this object's Object Lock will expire.

pub fn set_object_lock_retain_until_date(self, input: Option<Instant>) -> Self

pub fn object_lock_legal_hold_status(
self,
input: ObjectLockLegalHoldStatus
) -> Self

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

pub fn set_object_lock_legal_hold_status(
self,
input: Option<ObjectLockLegalHoldStatus>
) -> Self

pub fn build(self) -> GetObjectOutput

Consumes the builder and constructs a GetObjectOutput