Struct aws_sdk_s3::output::list_objects_v2_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ListObjectsV2Output

Implementations

impl Builder

pub fn is_truncated(self, input: bool) -> Self

Set to false if all of the results were returned. Set to true if more keys are available to return. If the number of results exceeds that specified by MaxKeys, all of the results might not be returned.

pub fn set_is_truncated(self, input: Option<bool>) -> Self

pub fn contents(self, input: impl Into<Object>) -> Self

pub fn set_contents(self, input: Option<Vec<Object>>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The bucket name.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the AWS SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the AWS SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn set_name(self, input: Option<String>) -> Self

pub fn prefix(self, input: impl Into<String>) -> Self

Keys that begin with the indicated prefix.

pub fn set_prefix(self, input: Option<String>) -> Self

pub fn delimiter(self, input: impl Into<String>) -> Self

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

pub fn set_delimiter(self, input: Option<String>) -> Self

pub fn max_keys(self, input: i32) -> Self

Sets the maximum number of keys returned in the response. By default the action returns up to 1,000 key names. The response might contain fewer keys but will never contain more.

pub fn set_max_keys(self, input: Option<i32>) -> Self

pub fn common_prefixes(self, input: impl Into<CommonPrefix>) -> Self

pub fn set_common_prefixes(self, input: Option<Vec<CommonPrefix>>) -> Self

pub fn encoding_type(self, input: EncodingType) -> Self

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify the encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

Delimiter, Prefix, Key, and StartAfter.

pub fn set_encoding_type(self, input: Option<EncodingType>) -> Self

pub fn key_count(self, input: i32) -> Self

KeyCount is the number of keys returned with this request. KeyCount will always be less than or equals to MaxKeys field. Say you ask for 50 keys, your result will include less than equals 50 keys

pub fn set_key_count(self, input: Option<i32>) -> Self

pub fn continuation_token(self, input: impl Into<String>) -> Self

If ContinuationToken was sent with the request, it is included in the response.

pub fn set_continuation_token(self, input: Option<String>) -> Self

pub fn next_continuation_token(self, input: impl Into<String>) -> Self

NextContinuationToken is sent when isTruncated is true, which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

pub fn set_next_continuation_token(self, input: Option<String>) -> Self

pub fn start_after(self, input: impl Into<String>) -> Self

If StartAfter was sent with the request, it is included in the response.

pub fn set_start_after(self, input: Option<String>) -> Self

pub fn build(self) -> ListObjectsV2Output

Consumes the builder and constructs a ListObjectsV2Output