Struct aws_sdk_s3::output::put_object_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for PutObjectOutput

Implementations

impl Builder

pub fn expiration(self, input: impl Into<String>) -> Self

If the expiration is configured for the object (see PutBucketLifecycleConfiguration), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide information about object expiration. The value of the rule-id is URL encoded.

pub fn set_expiration(self, input: Option<String>) -> Self

pub fn e_tag(self, input: impl Into<String>) -> Self

Entity tag for the uploaded object.

pub fn set_e_tag(self, input: Option<String>) -> Self

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

If you specified server-side encryption either with an AWS KMS customer master key (CMK) or Amazon S3-managed encryption key in your PUT request, the response includes this header. It confirms the encryption algorithm that Amazon S3 used to encrypt the object.

pub fn set_server_side_encryption(
self,
input: Option<ServerSideEncryption>
) -> Self

pub fn version_id(self, input: impl Into<String>) -> Self

Version of the object.

pub fn set_version_id(self, input: Option<String>) -> Self

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the AWS Key Management Service (AWS KMS) symmetric customer managed customer master key (CMK) that was used for the object.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

pub fn ssekms_encryption_context(self, input: impl Into<String>) -> Self

If present, specifies the AWS KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

pub fn set_ssekms_encryption_context(self, input: Option<String>) -> Self

pub fn bucket_key_enabled(self, input: bool) -> Self

Indicates whether the uploaded object uses an S3 Bucket Key for server-side encryption with AWS KMS (SSE-KMS).

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

pub fn build(self) -> PutObjectOutput

Consumes the builder and constructs a PutObjectOutput