Struct aws_sdk_s3::output::HeadObjectOutput

#[non_exhaustive]
pub struct HeadObjectOutput {

    pub delete_marker: bool,
    pub accept_ranges: Option<String>,
    pub expiration: Option<String>,
    pub restore: Option<String>,
    pub archive_status: Option<ArchiveStatus>,
    pub last_modified: Option<Instant>,
    pub content_length: i64,
    pub e_tag: Option<String>,
    pub missing_meta: i32,
    pub version_id: Option<String>,
    pub cache_control: Option<String>,
    pub content_disposition: Option<String>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_type: Option<String>,
    pub expires: Option<Instant>,
    pub website_redirect_location: Option<String>,
    pub server_side_encryption: Option<ServerSideEncryption>,
    pub metadata: Option<HashMap<String, String>>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub ssekms_key_id: Option<String>,
    pub bucket_key_enabled: bool,
    pub storage_class: Option<StorageClass>,
    pub request_charged: Option<RequestCharged>,
    pub replication_status: Option<ReplicationStatus>,
    pub parts_count: i32,
    pub object_lock_mode: Option<ObjectLockMode>,
    pub object_lock_retain_until_date: Option<Instant>,
    pub object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

delete_marker: bool

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

accept_ranges: Option<String>

Indicates that a range of bytes was specified.

expiration: Option<String>

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

restore: Option<String>

If the object is an archived object (an object whose storage class is GLACIER), the response includes this header if either the archive restoration is in progress (see RestoreObject or an archive copy is already restored.

If an archive copy is already restored, the header value indicates when Amazon S3 is scheduled to delete the object copy. For example:

x-amz-restore: ongoing-request="false", expiry-date="Fri, 21 Dec 2012 00:00:00 GMT"

If the object restoration is in progress, the header returns the value ongoing-request="true".

For more information about archiving objects, see Transitioning Objects: General Considerations.

archive_status: Option<ArchiveStatus>

The archive state of the head object.

last_modified: Option<Instant>

Creation date of the object.

content_length: i64

Size of the body in bytes.

e_tag: Option<String>

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

missing_meta: i32

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

version_id: Option<String>

Version of the object.

cache_control: Option<String>

Specifies caching behavior along the request/reply chain.

content_disposition: Option<String>

Specifies presentational information for the object.

content_encoding: Option<String>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

content_language: Option<String>

The language the content is in.

content_type: Option<String>

A standard MIME type describing the format of the object data.

expires: Option<Instant>

The date and time at which the object is no longer cacheable.

website_redirect_location: Option<String>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

server_side_encryption: Option<ServerSideEncryption>

If the object is stored using server-side encryption either with an AWS KMS customer master key (CMK) or an Amazon S3-managed encryption key, the response includes this header with the value of the server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

metadata: Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

sse_customer_algorithm: Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

sse_customer_key_md5: Option<String>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

ssekms_key_id: Option<String>

If present, specifies the ID of the AWS Key Management Service (AWS KMS) symmetric customer managed customer master key (CMK) that was used for the object.

bucket_key_enabled: bool

Indicates whether the object uses an S3 Bucket Key for server-side encryption with AWS KMS (SSE-KMS).

storage_class: Option<StorageClass>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

request_charged: Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request.

replication_status: Option<ReplicationStatus>

Amazon S3 can return this header if your request involves a bucket that is either a source or a destination in a replication rule.

In replication, you have a source bucket on which you configure replication and destination bucket or buckets where Amazon S3 stores object replicas. When you request an object (GetObject) or object metadata (HeadObject) from these buckets, Amazon S3 will return the x-amz-replication-status header in the response as follows:

If requesting an object from the source bucket — Amazon S3 will return the x-amz-replication-status header if the object in your request is eligible for replication.

For example, suppose that in your replication configuration, you specify object prefix TaxDocs requesting Amazon S3 to replicate objects with key prefix TaxDocs. Any objects you upload with this key name prefix, for example TaxDocs/document1.pdf, are eligible for replication. For any object request with this key name prefix, Amazon S3 will return the x-amz-replication-status header with value PENDING, COMPLETED or FAILED indicating object replication status.

If requesting an object from a destination bucket — Amazon S3 will return the x-amz-replication-status header with value REPLICA if the object in your request is a replica that Amazon S3 created and there is no replica modification replication in progress.

When replicating objects to multiple destination buckets the x-amz-replication-status header acts differently. The header of the source object will only return a value of COMPLETED when replication is successful to all destinations. The header will remain at value PENDING until replication has completed for all destinations. If one or more destinations fails replication the header will return FAILED.

For more information, see Replication.

parts_count: i32

The count of parts this object has.

object_lock_mode: Option<ObjectLockMode>

The Object Lock mode, if any, that's in effect for this object. This header is only returned if the requester has the s3:GetObjectRetention permission. For more information about S3 Object Lock, see Object Lock.

object_lock_retain_until_date: Option<Instant>

The date and time when the Object Lock retention period expires. This header is only returned if the requester has the s3:GetObjectRetention permission.

object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>

Specifies whether a legal hold is in effect for this object. This header is only returned if the requester has the s3:GetObjectLegalHold permission. This header is not returned if the specified version of this object has never had a legal hold applied. For more information about S3 Object Lock, see Object Lock.

Implementations

impl HeadObjectOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture HeadObjectOutput