Struct aws_sdk_s3::output::ListObjectVersionsOutput

#[non_exhaustive]
pub struct ListObjectVersionsOutput {

    pub is_truncated: bool,
    pub key_marker: Option<String>,
    pub version_id_marker: Option<String>,
    pub next_key_marker: Option<String>,
    pub next_version_id_marker: Option<String>,
    pub versions: Option<Vec<ObjectVersion>>,
    pub delete_markers: Option<Vec<DeleteMarkerEntry>>,
    pub name: Option<String>,
    pub prefix: Option<String>,
    pub delimiter: Option<String>,
    pub max_keys: i32,
    pub common_prefixes: Option<Vec<CommonPrefix>>,
    pub encoding_type: Option<EncodingType>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

is_truncated: bool

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

key_marker: Option<String>

Marks the last key returned in a truncated response.

version_id_marker: Option<String>

Marks the last version of the key returned in a truncated response.

next_key_marker: Option<String>

When the number of responses exceeds the value of MaxKeys, NextKeyMarker specifies the first key not returned that satisfies the search criteria. Use this value for the key-marker request parameter in a subsequent request.

next_version_id_marker: Option<String>

When the number of responses exceeds the value of MaxKeys, NextVersionIdMarker specifies the first object version not returned that satisfies the search criteria. Use this value for the version-id-marker request parameter in a subsequent request.

versions: Option<Vec<ObjectVersion>>

Container for version information.

delete_markers: Option<Vec<DeleteMarkerEntry>>

Container for an object that is a delete marker.

name: Option<String>

The bucket name.

prefix: Option<String>

Selects objects that start with the value supplied by this parameter.

delimiter: Option<String>

The delimiter grouping the included keys. A delimiter is a character that you specify to group keys. All keys that contain the same string between the prefix and the first occurrence of the delimiter are grouped under a single result element in CommonPrefixes. These groups are counted as one result against the max-keys limitation. These keys are not returned elsewhere in the response.

max_keys: i32

Specifies the maximum number of objects to return.

common_prefixes: Option<Vec<CommonPrefix>>

All of the keys rolled up into a common prefix count as a single return when calculating the number of returns.

encoding_type: Option<EncodingType>

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

KeyMarker, NextKeyMarker, Prefix, Key, and Delimiter.

Implementations

impl ListObjectVersionsOutput

pub fn is_truncated(&self) -> bool

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

pub fn key_marker(&self) -> Option<&str>

Marks the last key returned in a truncated response.

pub fn version_id_marker(&self) -> Option<&str>

Marks the last version of the key returned in a truncated response.

pub fn next_key_marker(&self) -> Option<&str>

When the number of responses exceeds the value of MaxKeys, NextKeyMarker specifies the first key not returned that satisfies the search criteria. Use this value for the key-marker request parameter in a subsequent request.

pub fn next_version_id_marker(&self) -> Option<&str>

When the number of responses exceeds the value of MaxKeys, NextVersionIdMarker specifies the first object version not returned that satisfies the search criteria. Use this value for the version-id-marker request parameter in a subsequent request.

pub fn versions(&self) -> Option<&[ObjectVersion]>

Container for version information.

pub fn delete_markers(&self) -> Option<&[DeleteMarkerEntry]>

Container for an object that is a delete marker.

pub fn name(&self) -> Option<&str>

The bucket name.

pub fn prefix(&self) -> Option<&str>

Selects objects that start with the value supplied by this parameter.

pub fn delimiter(&self) -> Option<&str>

The delimiter grouping the included keys. A delimiter is a character that you specify to group keys. All keys that contain the same string between the prefix and the first occurrence of the delimiter are grouped under a single result element in CommonPrefixes. These groups are counted as one result against the max-keys limitation. These keys are not returned elsewhere in the response.

pub fn max_keys(&self) -> i32

Specifies the maximum number of objects to return.

pub fn common_prefixes(&self) -> Option<&[CommonPrefix]>

All of the keys rolled up into a common prefix count as a single return when calculating the number of returns.

pub fn encoding_type(&self) -> Option<&EncodingType>

Encoding type used by Amazon S3 to encode object key names in the XML response.

If you specify encoding-type request parameter, Amazon S3 includes this element in the response, and returns encoded key name values in the following response elements:

KeyMarker, NextKeyMarker, Prefix, Key, and Delimiter.

impl ListObjectVersionsOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ListObjectVersionsOutput