Struct aws_sdk_s3::output::ListObjectsOutput

#[non_exhaustive]
pub struct ListObjectsOutput {
    pub is_truncated: bool,
    pub marker: Option<String>,
    pub next_marker: Option<String>,
    pub contents: Option<Vec<Object>>,
    pub name: Option<String>,
    pub prefix: Option<String>,
    pub delimiter: Option<String>,
    pub max_keys: i32,
    pub common_prefixes: Option<Vec<CommonPrefix>>,
    pub encoding_type: Option<EncodingType>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

is_truncated: bool

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria.

marker: Option<String>

Indicates where in the bucket listing begins. Marker is included in the response if it was sent with the request.

next_marker: Option<String>

When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMarker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

contents: Option<Vec<Object>>

Metadata about each object returned.

name: Option<String>

The bucket name.

prefix: Option<String>

Keys that begin with the indicated prefix.

delimiter: Option<String>

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

max_keys: i32

The maximum number of keys returned in the response body.

common_prefixes: Option<Vec<CommonPrefix>>

All of the keys (up to 1,000) rolled up in a common prefix count as a single return when calculating the number of returns.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by the delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/) as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

encoding_type: Option<EncodingType>

Encoding type used by Amazon S3 to encode object keys in the response.

Implementations

impl ListObjectsOutput

pub fn is_truncated(&self) -> bool

A flag that indicates whether Amazon S3 returned all of the results that satisfied the search criteria.

pub fn marker(&self) -> Option<&str>

Indicates where in the bucket listing begins. Marker is included in the response if it was sent with the request.

pub fn next_marker(&self) -> Option<&str>

When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMarker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

pub fn contents(&self) -> Option<&[Object]>

Metadata about each object returned.

pub fn name(&self) -> Option<&str>

The bucket name.

pub fn prefix(&self) -> Option<&str>

Keys that begin with the indicated prefix.

pub fn delimiter(&self) -> Option<&str>

Causes keys that contain the same string between the prefix and the first occurrence of the delimiter to be rolled up into a single result element in the CommonPrefixes collection. These rolled-up keys are not returned elsewhere in the response. Each rolled-up result counts as only one return against the MaxKeys value.

pub fn max_keys(&self) -> i32

The maximum number of keys returned in the response body.

pub fn common_prefixes(&self) -> Option<&[CommonPrefix]>

All of the keys (up to 1,000) rolled up in a common prefix count as a single return when calculating the number of returns.

A response can contain CommonPrefixes only if you specify a delimiter.

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by the delimiter.

CommonPrefixes lists keys that act like subdirectories in the directory specified by Prefix.

For example, if the prefix is notes/ and the delimiter is a slash (/) as in notes/summer/july, the common prefix is notes/summer/. All of the keys that roll up into a common prefix count as a single return when calculating the number of returns.

pub fn encoding_type(&self) -> Option<&EncodingType>

Encoding type used by Amazon S3 to encode object keys in the response.

impl ListObjectsOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ListObjectsOutput