Struct aws_sdk_s3::output::ListPartsOutput

#[non_exhaustive]
pub struct ListPartsOutput {

    pub abort_date: Option<DateTime>,
    pub abort_rule_id: Option<String>,
    pub bucket: Option<String>,
    pub key: Option<String>,
    pub upload_id: Option<String>,
    pub part_number_marker: Option<String>,
    pub next_part_number_marker: Option<String>,
    pub max_parts: i32,
    pub is_truncated: bool,
    pub parts: Option<Vec<Part>>,
    pub initiator: Option<Initiator>,
    pub owner: Option<Owner>,
    pub storage_class: Option<StorageClass>,
    pub request_charged: Option<RequestCharged>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

abort_date: Option<DateTime>

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, then the response includes this header indicating when the initiated multipart upload will become eligible for abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response will also include the x-amz-abort-rule-id header that will provide the ID of the lifecycle configuration rule that defines this action.

abort_rule_id: Option<String>

This header is returned along with the x-amz-abort-date header. It identifies applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

bucket: Option<String>

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

key: Option<String>

Object key for which the multipart upload was initiated.

upload_id: Option<String>

Upload ID identifying the multipart upload whose parts are being listed.

part_number_marker: Option<String>

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

next_part_number_marker: Option<String>

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

max_parts: i32

Maximum number of parts that were allowed in the response.

is_truncated: bool

Indicates whether the returned list of parts is truncated. A true value indicates that the list was truncated. A list can be truncated if the number of parts exceeds the limit returned in the MaxParts element.

parts: Option<Vec<Part>>

Container for elements related to a particular part. A response can contain zero or more Part elements.

initiator: Option<Initiator>

Container element that identifies who initiated the multipart upload. If the initiator is an Amazon Web Services account, this element provides the same information as the Owner element. If the initiator is an IAM User, this element provides the user ARN and display name.

owner: Option<Owner>

Container element that identifies the object owner, after the object is created. If multipart upload is initiated by an IAM user, this element provides the parent account ID and display name.

storage_class: Option<StorageClass>

Class of storage (STANDARD or REDUCED_REDUNDANCY) used to store the uploaded object.

request_charged: Option<RequestCharged>

If present, indicates that the requester was successfully charged for the request.

Implementations

impl ListPartsOutput

pub fn abort_date(&self) -> Option<&DateTime>

If the bucket has a lifecycle rule configured with an action to abort incomplete multipart uploads and the prefix in the lifecycle rule matches the object name in the request, then the response includes this header indicating when the initiated multipart upload will become eligible for abort operation. For more information, see Aborting Incomplete Multipart Uploads Using a Bucket Lifecycle Policy.

The response will also include the x-amz-abort-rule-id header that will provide the ID of the lifecycle configuration rule that defines this action.

pub fn abort_rule_id(&self) -> Option<&str>

This header is returned along with the x-amz-abort-date header. It identifies applicable lifecycle configuration rule that defines the action to abort incomplete multipart uploads.

pub fn bucket(&self) -> Option<&str>

The name of the bucket to which the multipart upload was initiated. Does not return the access point ARN or access point alias if used.

pub fn key(&self) -> Option<&str>

Object key for which the multipart upload was initiated.

pub fn upload_id(&self) -> Option<&str>

Upload ID identifying the multipart upload whose parts are being listed.

pub fn part_number_marker(&self) -> Option<&str>

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

pub fn next_part_number_marker(&self) -> Option<&str>

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

pub fn max_parts(&self) -> i32

Maximum number of parts that were allowed in the response.

pub fn is_truncated(&self) -> bool

Indicates whether the returned list of parts is truncated. A true value indicates that the list was truncated. A list can be truncated if the number of parts exceeds the limit returned in the MaxParts element.

pub fn parts(&self) -> Option<&[Part]>

Container for elements related to a particular part. A response can contain zero or more Part elements.

pub fn initiator(&self) -> Option<&Initiator>

Container element that identifies who initiated the multipart upload. If the initiator is an Amazon Web Services account, this element provides the same information as the Owner element. If the initiator is an IAM User, this element provides the user ARN and display name.

pub fn owner(&self) -> Option<&Owner>

Container element that identifies the object owner, after the object is created. If multipart upload is initiated by an IAM user, this element provides the parent account ID and display name.

pub fn storage_class(&self) -> Option<&StorageClass>

Class of storage (STANDARD or REDUCED_REDUNDANCY) used to store the uploaded object.

pub fn request_charged(&self) -> Option<&RequestCharged>

If present, indicates that the requester was successfully charged for the request.

impl ListPartsOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ListPartsOutput