aws_sdk_s3::types::builders

Struct CreateBucketConfigurationBuilder

#[non_exhaustive]
pub struct CreateBucketConfigurationBuilder { /* private fields */ }

A builder for CreateBucketConfiguration.

Implementations

impl CreateBucketConfigurationBuilder

pub fn location_constraint(self, input: BucketLocationConstraint) -> Self

Specifies the Region where the bucket will be created. You might choose a Region to optimize latency, minimize costs, or address regulatory requirements. For example, if you reside in Europe, you will probably find it advantageous to create buckets in the Europe (Ireland) Region.

If you don't specify a Region, the bucket is created in the US East (N. Virginia) Region (us-east-1) by default. Configurations using the value EU will create a bucket in eu-west-1.

For a list of the valid values for all of the Amazon Web Services Regions, see Regions and Endpoints.

This functionality is not supported for directory buckets.

pub fn set_location_constraint( self, input: Option<BucketLocationConstraint>, ) -> Self

Specifies the Region where the bucket will be created. You might choose a Region to optimize latency, minimize costs, or address regulatory requirements. For example, if you reside in Europe, you will probably find it advantageous to create buckets in the Europe (Ireland) Region.

If you don't specify a Region, the bucket is created in the US East (N. Virginia) Region (us-east-1) by default. Configurations using the value EU will create a bucket in eu-west-1.

For a list of the valid values for all of the Amazon Web Services Regions, see Regions and Endpoints.

This functionality is not supported for directory buckets.

pub fn get_location_constraint(&self) -> &Option<BucketLocationConstraint>

Specifies the Region where the bucket will be created. You might choose a Region to optimize latency, minimize costs, or address regulatory requirements. For example, if you reside in Europe, you will probably find it advantageous to create buckets in the Europe (Ireland) Region.

If you don't specify a Region, the bucket is created in the US East (N. Virginia) Region (us-east-1) by default. Configurations using the value EU will create a bucket in eu-west-1.

For a list of the valid values for all of the Amazon Web Services Regions, see Regions and Endpoints.

This functionality is not supported for directory buckets.

pub fn location(self, input: LocationInfo) -> Self

Specifies the location where the bucket will be created.

Directory buckets - The location type is Availability Zone or Local Zone. To use the Local Zone location type, your account must be enabled for Local Zones. Otherwise, you get an HTTP 403 Forbidden error with the error code AccessDenied. To learn more, see Enable accounts for Local Zones in the Amazon S3 User Guide.

This functionality is only supported by directory buckets.

pub fn set_location(self, input: Option<LocationInfo>) -> Self

Specifies the location where the bucket will be created.

Directory buckets - The location type is Availability Zone or Local Zone. To use the Local Zone location type, your account must be enabled for Local Zones. Otherwise, you get an HTTP 403 Forbidden error with the error code AccessDenied. To learn more, see Enable accounts for Local Zones in the Amazon S3 User Guide.

This functionality is only supported by directory buckets.

pub fn get_location(&self) -> &Option<LocationInfo>

Specifies the location where the bucket will be created.

Directory buckets - The location type is Availability Zone or Local Zone. To use the Local Zone location type, your account must be enabled for Local Zones. Otherwise, you get an HTTP 403 Forbidden error with the error code AccessDenied. To learn more, see Enable accounts for Local Zones in the Amazon S3 User Guide.

This functionality is only supported by directory buckets.

pub fn bucket(self, input: BucketInfo) -> Self

Specifies the information about the bucket that will be created.

This functionality is only supported by directory buckets.

pub fn set_bucket(self, input: Option<BucketInfo>) -> Self

Specifies the information about the bucket that will be created.

This functionality is only supported by directory buckets.

pub fn get_bucket(&self) -> &Option<BucketInfo>

Specifies the information about the bucket that will be created.

This functionality is only supported by directory buckets.

pub fn tags(self, input: Tag) -> Self

Appends an item to tags.

To override the contents of this collection use set_tags.

An array of tags that you can apply to the bucket that you're creating. Tags are key-value pairs of metadata used to categorize and organize your buckets, track costs, and control access.

You must have the s3:TagResource permission to create a general purpose bucket with tags or the s3express:TagResource permission to create a directory bucket with tags.

When creating buckets with tags, note that tag-based conditions using aws:ResourceTag and s3:BucketTag condition keys are applicable only after ABAC is enabled on the bucket. To learn more, see Enabling ABAC in general purpose buckets.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

An array of tags that you can apply to the bucket that you're creating. Tags are key-value pairs of metadata used to categorize and organize your buckets, track costs, and control access.

You must have the s3:TagResource permission to create a general purpose bucket with tags or the s3express:TagResource permission to create a directory bucket with tags.

When creating buckets with tags, note that tag-based conditions using aws:ResourceTag and s3:BucketTag condition keys are applicable only after ABAC is enabled on the bucket. To learn more, see Enabling ABAC in general purpose buckets.

pub fn get_tags(&self) -> &Option<Vec<Tag>>

An array of tags that you can apply to the bucket that you're creating. Tags are key-value pairs of metadata used to categorize and organize your buckets, track costs, and control access.

You must have the s3:TagResource permission to create a general purpose bucket with tags or the s3express:TagResource permission to create a directory bucket with tags.

When creating buckets with tags, note that tag-based conditions using aws:ResourceTag and s3:BucketTag condition keys are applicable only after ABAC is enabled on the bucket. To learn more, see Enabling ABAC in general purpose buckets.

pub fn build(self) -> CreateBucketConfiguration

Consumes the builder and constructs a CreateBucketConfiguration.