aws_sdk_s3::types::builders

Struct InventoryConfigurationBuilder

#[non_exhaustive]
pub struct InventoryConfigurationBuilder { /* private fields */ }

A builder for InventoryConfiguration.

Implementations

impl InventoryConfigurationBuilder

pub fn destination(self, input: InventoryDestination) -> Self

Contains information about where to publish the inventory results.
This field is required.

pub fn set_destination(self, input: Option<InventoryDestination>) -> Self

Contains information about where to publish the inventory results.

pub fn get_destination(&self) -> &Option<InventoryDestination>

Contains information about where to publish the inventory results.

pub fn is_enabled(self, input: bool) -> Self

Specifies whether the inventory is enabled or disabled. If set to True, an inventory list is generated. If set to False, no inventory list is generated.
This field is required.

pub fn set_is_enabled(self, input: Option<bool>) -> Self

Specifies whether the inventory is enabled or disabled. If set to True, an inventory list is generated. If set to False, no inventory list is generated.

pub fn get_is_enabled(&self) -> &Option<bool>

Specifies whether the inventory is enabled or disabled. If set to True, an inventory list is generated. If set to False, no inventory list is generated.

pub fn filter(self, input: InventoryFilter) -> Self

Specifies an inventory filter. The inventory only includes objects that meet the filter's criteria.

pub fn set_filter(self, input: Option<InventoryFilter>) -> Self

Specifies an inventory filter. The inventory only includes objects that meet the filter's criteria.

pub fn get_filter(&self) -> &Option<InventoryFilter>

Specifies an inventory filter. The inventory only includes objects that meet the filter's criteria.

pub fn id(self, input: impl Into<String>) -> Self

The ID used to identify the inventory configuration.
This field is required.

pub fn set_id(self, input: Option<String>) -> Self

The ID used to identify the inventory configuration.

pub fn get_id(&self) -> &Option<String>

The ID used to identify the inventory configuration.

pub fn included_object_versions( self, input: InventoryIncludedObjectVersions, ) -> Self

Object versions to include in the inventory list. If set to All, the list includes all the object versions, which adds the version-related fields VersionId, IsLatest, and DeleteMarker to the list. If set to Current, the list does not contain these version-related fields.
This field is required.

pub fn set_included_object_versions( self, input: Option<InventoryIncludedObjectVersions>, ) -> Self

Object versions to include in the inventory list. If set to All, the list includes all the object versions, which adds the version-related fields VersionId, IsLatest, and DeleteMarker to the list. If set to Current, the list does not contain these version-related fields.

pub fn get_included_object_versions( &self, ) -> &Option<InventoryIncludedObjectVersions>

Object versions to include in the inventory list. If set to All, the list includes all the object versions, which adds the version-related fields VersionId, IsLatest, and DeleteMarker to the list. If set to Current, the list does not contain these version-related fields.

pub fn optional_fields(self, input: InventoryOptionalField) -> Self

Appends an item to optional_fields.

To override the contents of this collection use set_optional_fields.

Contains the optional fields that are included in the inventory results.

pub fn set_optional_fields( self, input: Option<Vec<InventoryOptionalField>>, ) -> Self

Contains the optional fields that are included in the inventory results.

pub fn get_optional_fields(&self) -> &Option<Vec<InventoryOptionalField>>

Contains the optional fields that are included in the inventory results.

pub fn schedule(self, input: InventorySchedule) -> Self

Specifies the schedule for generating inventory results.
This field is required.

pub fn set_schedule(self, input: Option<InventorySchedule>) -> Self

Specifies the schedule for generating inventory results.

pub fn get_schedule(&self) -> &Option<InventorySchedule>

Specifies the schedule for generating inventory results.

pub fn build(self) -> Result<InventoryConfiguration, BuildError>

Consumes the builder and constructs a InventoryConfiguration. This method will fail if any of the following fields are not set:

is_enabled

id

included_object_versions