Enum aws_sdk_s3::types::SdkError

pub enum SdkError<E, R = Response> {
    ConstructionFailure(Box<dyn Error + Send + Sync + 'static, Global>),
    TimeoutError(Box<dyn Error + Send + Sync + 'static, Global>),
    DispatchFailure(ConnectorError),
    ResponseError {
        err: Box<dyn Error + Send + Sync + 'static, Global>,
        raw: R,
    },
    ServiceError {
        err: E,
        raw: R,
    },
}

Failed SDK Result

Variants

ConstructionFailure(Box<dyn Error + Send + Sync + 'static, Global>)

The request failed during construction. It was not dispatched over the network.

TimeoutError(Box<dyn Error + Send + Sync + 'static, Global>)

The request failed due to a timeout. The request MAY have been sent and received.

DispatchFailure(ConnectorError)

The request failed during dispatch. An HTTP response was not received. The request MAY have been sent.

ResponseError

Fields

err: Box<dyn Error + Send + Sync + 'static, Global>

Error encountered while parsing the response

raw: R

Raw response that was available

A response was received but it was not parseable according the the protocol (for example the server hung up while the body was being read)

ServiceError

Fields

err: E

Modeled service error

raw: R

Raw response from the service

An error response was received from the service