aws_sdk_s3::types

Struct LifecycleRuleFilter

#[non_exhaustive]
pub struct LifecycleRuleFilter {
    pub prefix: Option<String>,
    pub tag: Option<Tag>,
    pub object_size_greater_than: Option<i64>,
    pub object_size_less_than: Option<i64>,
    pub and: Option<LifecycleRuleAndOperator>,
}

The Filter is used to identify objects that a Lifecycle Rule applies to. A Filter can have exactly one of Prefix, Tag, ObjectSizeGreaterThan, ObjectSizeLessThan, or And specified. If the Filter element is left empty, the Lifecycle Rule applies to all objects in the bucket.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

prefix: Option<String>

Prefix identifying one or more objects to which the rule applies.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

tag: Option<Tag>

This tag must exist in the object's tag set in order for the rule to apply.

This parameter applies to general purpose buckets only. It is not supported for directory bucket lifecycle configurations.

object_size_greater_than: Option<i64>

Minimum object size to which the rule applies.

object_size_less_than: Option<i64>

Maximum object size to which the rule applies.

and: Option<LifecycleRuleAndOperator>

This is used in a Lifecycle Rule Filter to apply a logical AND to two or more predicates. The Lifecycle Rule will apply to any object matching all of the predicates configured inside the And operator.

Implementations

impl LifecycleRuleFilter

pub fn prefix(&self) -> Option<&str>

Prefix identifying one or more objects to which the rule applies.

Replacement must be made for object keys containing special characters (such as carriage returns) when using XML requests. For more information, see XML related object key constraints.

pub fn tag(&self) -> Option<&Tag>

This tag must exist in the object's tag set in order for the rule to apply.

This parameter applies to general purpose buckets only. It is not supported for directory bucket lifecycle configurations.

pub fn object_size_greater_than(&self) -> Option<i64>

Minimum object size to which the rule applies.

pub fn object_size_less_than(&self) -> Option<i64>

Maximum object size to which the rule applies.

pub fn and(&self) -> Option<&LifecycleRuleAndOperator>

This is used in a Lifecycle Rule Filter to apply a logical AND to two or more predicates. The Lifecycle Rule will apply to any object matching all of the predicates configured inside the And operator.

impl LifecycleRuleFilter

pub fn builder() -> LifecycleRuleFilterBuilder

Creates a new builder-style object to manufacture LifecycleRuleFilter.