aws_sdk_s3::waiters::bucket_exists

Type Alias WaitUntilBucketExistsError

pub type WaitUntilBucketExistsError = WaiterError<HeadBucketOutput, HeadBucketError>;

Error type for the bucket_exists waiter.

Aliased Type

pub enum WaitUntilBucketExistsError {
    ConstructionFailure(ConstructionFailure),
    ExceededMaxWait(ExceededMaxWait),
    FailureState(FailureState<HeadBucketOutput, HeadBucketError>),
    OperationFailed(OperationFailed<HeadBucketError>),
}

Variants

ConstructionFailure(ConstructionFailure)

An error occurred during waiter initialization.

This can happen if the input/config is invalid.

ExceededMaxWait(ExceededMaxWait)

The maximum wait time was exceeded without completion.

FailureState(FailureState<HeadBucketOutput, HeadBucketError>)

Waiting ended in a failure state.

A failed waiter state can occur on a successful response from the server if, for example, that response indicates that the thing being waited for won’t succeed/finish.

A failure state error will only occur for successful or modeled error responses. Unmodeled error responses will never make it into this error case.

OperationFailed(OperationFailed<HeadBucketError>)

A polling operation failed while waiting.

This error will only occur for unmodeled errors. Modeled errors can potentially be handled by the waiter logic, and will therefore end up in [WaiterError::FailureState].

Note: If retry is configured, this means that the operation failed after retrying the configured number of attempts.