Crate aws_sdk_s3

Getting Started

Examples are available for many services and operations, check out the usage examples.

The SDK provides one crate per AWS service. You must add Tokio as a dependency within your Rust project to execute asynchronous code. To add aws-sdk-s3 to your project, add the following to your Cargo.toml file:

[dependencies]
aws-config = { version = "1.1.7", features = ["behavior-version-latest"] }
aws-sdk-s3 = "1.121.0"
tokio = { version = "1", features = ["full"] }

Then in code, a client can be created with the following:

use aws_sdk_s3 as s3;

#[::tokio::main]
async fn main() -> Result<(), s3::Error> {
    let config = aws_config::load_from_env().await;
    let client = aws_sdk_s3::Client::new(&config);

    // ... make some calls with the client

    Ok(())
}

See the client documentation for information on what calls can be made, and the inputs and outputs for each of those calls.

Using the SDK

Until the SDK is released, we will be adding information about using the SDK to the Developer Guide. Feel free to suggest additional sections for the guide by opening an issue and describing what you are trying to do.

Getting Help

GitHub discussions - For ideas, RFCs & general questions

GitHub issues - For bug reports & feature requests

Generated Docs (latest version)

Usage examples

Crate Organization

The entry point for most customers will be Client, which exposes one method for each API offered by Amazon Simple Storage Service. The return value of each of these methods is a “fluent builder”, where the different inputs for that API are added by builder-style function call chaining, followed by calling send() to get a Future that will result in either a successful output or a SdkError.

Some of these API inputs may be structs or enums to provide more complex structured information. These structs and enums live in types. There are some simpler types for representing data such as date times or binary blobs that live in primitives.

All types required to configure a client via the Config struct live in config.

The operation module has a submodule for every API, and in each submodule is the input, output, and error type for that API, as well as builders to construct each of those.

There is a top-level Error type that encompasses all the errors that the client can return. Any other error type can be converted to this Error type via the From trait.

The other modules within this crate are not required for normal usage.

Modules

client

Client for calling Amazon Simple Storage Service.

config

Configuration for Amazon Simple Storage Service.

error

Common errors and error handling utilities.

meta

Information about this crate.

operation

All operations that this crate can perform.

presigning

Presigned request types and configuration.

primitives

Primitives such as Blob or DateTime used by other types.

types

Data structures used by operation inputs/outputs.

waiters

Supporting types for waiters.

Structs

Client

Client for Amazon Simple Storage Service

Config

Configuration for a aws_sdk_s3 service client.

Enums

Error

All possible error types for this service.