Struct aws_sdk_s3::client::fluent_builders::CreateBucket

pub struct CreateBucket<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Fluent builder constructing a request to CreateBucket.

Creates a new S3 bucket. To create a bucket, you must register with Amazon S3 and have a valid Amazon Web Services Access Key ID to authenticate requests. Anonymous requests are never allowed to create buckets. By creating the bucket, you become the bucket owner.

Not every string is an acceptable bucket name. For information about bucket naming restrictions, see Bucket naming rules.

If you want to create an Amazon S3 on Outposts bucket, see Create Bucket.

By default, the bucket is created in the US East (N. Virginia) Region. You can optionally specify a Region in the request body. You might choose a Region to optimize latency, minimize costs, or address regulatory requirements. For example, if you reside in Europe, you will probably find it advantageous to create buckets in the Europe (Ireland) Region. For more information, see Accessing a bucket.

If you send your create bucket request to the s3.amazonaws.com endpoint, the request goes to the us-east-1 Region. Accordingly, the signature calculations in Signature Version 4 must use us-east-1 as the Region, even if the location constraint in the request specifies another Region where the bucket is to be created. If you create a bucket in a Region other than US East (N. Virginia), your application must be able to handle 307 redirect. For more information, see Virtual hosting of buckets.

When creating a bucket using this operation, you can optionally specify the accounts or groups that should be granted specific permissions on the bucket. There are two ways to grant the appropriate permissions using the request headers.

Specify a canned ACL using the x-amz-acl request header. Amazon S3 supports a set of predefined ACLs, known as canned ACLs. Each canned ACL has a predefined set of grantees and permissions. For more information, see Canned ACL.

Specify access permissions explicitly using the x-amz-grant-read, x-amz-grant-write, x-amz-grant-read-acp, x-amz-grant-write-acp, and x-amz-grant-full-control headers. These headers map to the set of permissions Amazon S3 supports in an ACL. For more information, see Access control list (ACL) overview.

You specify each grantee as a type=value pair, where the type is one of the following:

id – if the value specified is the canonical user ID of an Amazon Web Services account

uri – if you are granting permissions to a predefined group

emailAddress – if the value specified is the email address of an Amazon Web Services account

Using email addresses to specify a grantee is only supported in the following Amazon Web Services Regions:

US East (N. Virginia)

US West (N. California)

US West (Oregon)

Asia Pacific (Singapore)

Asia Pacific (Sydney)

Asia Pacific (Tokyo)

Europe (Ireland)

South America (São Paulo)

For a list of all the Amazon S3 supported Regions and endpoints, see Regions and Endpoints in the Amazon Web Services General Reference.

For example, the following x-amz-grant-read header grants the Amazon Web Services accounts identified by account IDs permissions to read object data and its metadata:

x-amz-grant-read: id="11112222333", id="444455556666"

You can use either a canned ACL or specify access permissions explicitly. You cannot do both.

Permissions

If your CreateBucket request specifies ACL permissions and the ACL is public-read, public-read-write, authenticated-read, or if you specify access permissions explicitly through any other ACL, both s3:CreateBucket and s3:PutBucketAcl permissions are needed. If the ACL the CreateBucket request is private, only s3:CreateBucket permission is needed.

If ObjectLockEnabledForBucket is set to true in your CreateBucket request, s3:PutBucketObjectLockConfiguration and s3:PutBucketVersioning permissions are required.

The following operations are related to CreateBucket:

PutObject

DeleteBucket

Implementations

impl<C, M, R> CreateBucket<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateBucketOutput, SdkError<CreateBucketError>> where
R::Policy: SmithyRetryPolicy<CreateBucketInputOperationOutputAlias, CreateBucketOutput, CreateBucketError, CreateBucketInputOperationRetryAlias>,

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn acl(self, inp: BucketCannedAcl) -> Self

The canned ACL to apply to the bucket.

pub fn set_acl(self, input: Option<BucketCannedAcl>) -> Self

The canned ACL to apply to the bucket.

pub fn bucket(self, inp: impl Into<String>) -> Self

The name of the bucket to create.

pub fn set_bucket(self, input: Option<String>) -> Self

The name of the bucket to create.

pub fn create_bucket_configuration(self, inp: CreateBucketConfiguration) -> Self

The configuration information for the bucket.

pub fn set_create_bucket_configuration(
self,
input: Option<CreateBucketConfiguration>
) -> Self

The configuration information for the bucket.

pub fn grant_full_control(self, inp: impl Into<String>) -> Self

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

pub fn set_grant_full_control(self, input: Option<String>) -> Self

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

pub fn grant_read(self, inp: impl Into<String>) -> Self

Allows grantee to list the objects in the bucket.

pub fn set_grant_read(self, input: Option<String>) -> Self

Allows grantee to list the objects in the bucket.

pub fn grant_read_acp(self, inp: impl Into<String>) -> Self

Allows grantee to read the bucket ACL.

pub fn set_grant_read_acp(self, input: Option<String>) -> Self

Allows grantee to read the bucket ACL.

pub fn grant_write(self, inp: impl Into<String>) -> Self

Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

pub fn set_grant_write(self, input: Option<String>) -> Self

Allows grantee to create new objects in the bucket.

For the bucket and object owners of existing objects, also allows deletions and overwrites of those objects.

pub fn grant_write_acp(self, inp: impl Into<String>) -> Self

Allows grantee to write the ACL for the applicable bucket.

pub fn set_grant_write_acp(self, input: Option<String>) -> Self

Allows grantee to write the ACL for the applicable bucket.

pub fn object_lock_enabled_for_bucket(self, inp: bool) -> Self

Specifies whether you want S3 Object Lock to be enabled for the new bucket.

pub fn set_object_lock_enabled_for_bucket(self, input: Option<bool>) -> Self

Specifies whether you want S3 Object Lock to be enabled for the new bucket.