Struct aws_sdk_s3::client::fluent_builders::SelectObjectContent

pub struct SelectObjectContent<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Fluent builder constructing a request to SelectObjectContent.

This action filters the contents of an Amazon S3 object based on a simple structured query language (SQL) statement. In the request, along with the SQL expression, you must also specify a data serialization format (JSON, CSV, or Apache Parquet) of the object. Amazon S3 uses this format to parse object data into records, and returns only records that match the specified SQL expression. You must also specify the data serialization format for the response.

This action is not supported by Amazon S3 on Outposts.

For more information about Amazon S3 Select, see Selecting Content from Objects in the Amazon S3 User Guide.

For more information about using SQL with Amazon S3 Select, see SQL Reference for Amazon S3 Select and S3 Glacier Select in the Amazon S3 User Guide.

Permissions

You must have s3:GetObject permission for this operation. Amazon S3 Select does not support anonymous access. For more information about permissions, see Specifying Permissions in a Policy in the Amazon S3 User Guide.

Object Data Formats

You can use Amazon S3 Select to query objects that have the following format properties:

CSV, JSON, and Parquet - Objects must be in CSV, JSON, or Parquet format.

UTF-8 - UTF-8 is the only encoding type Amazon S3 Select supports.

GZIP or BZIP2 - CSV and JSON files can be compressed using GZIP or BZIP2. GZIP and BZIP2 are the only compression formats that Amazon S3 Select supports for CSV and JSON files. Amazon S3 Select supports columnar compression for Parquet using GZIP or Snappy. Amazon S3 Select does not support whole-object compression for Parquet objects.

Server-side encryption - Amazon S3 Select supports querying objects that are protected with server-side encryption.

For objects that are encrypted with customer-provided encryption keys (SSE-C), you must use HTTPS, and you must use the headers that are documented in the GetObject. For more information about SSE-C, see Server-Side Encryption (Using Customer-Provided Encryption Keys) in the Amazon S3 User Guide.

For objects that are encrypted with Amazon S3 managed encryption keys (SSE-S3) and Amazon Web Services KMS keys (SSE-KMS), server-side encryption is handled transparently, so you don't need to specify anything. For more information about server-side encryption, including SSE-S3 and SSE-KMS, see Protecting Data Using Server-Side Encryption in the Amazon S3 User Guide.

Working with the Response Body

Given the response size is unknown, Amazon S3 Select streams the response as a series of messages and includes a Transfer-Encoding header with chunked as its value in the response. For more information, see Appendix: SelectObjectContent Response.

GetObject Support

The SelectObjectContent action does not support the following GetObject functionality. For more information, see GetObject.

Range: Although you can specify a scan range for an Amazon S3 Select request (see SelectObjectContentRequest - ScanRange in the request parameters), you cannot specify the range of bytes of an object to return.

GLACIER, DEEP_ARCHIVE and REDUCED_REDUNDANCY storage classes: You cannot specify the GLACIER, DEEP_ARCHIVE, or REDUCED_REDUNDANCY storage classes. For more information, about storage classes see Storage Classes in the Amazon S3 User Guide.

Special Errors

For a list of special errors for this operation, see List of SELECT Object Content Error Codes

Related Resources

GetObject

GetBucketLifecycleConfiguration

PutBucketLifecycleConfiguration

Implementations

impl<C, M, R> SelectObjectContent<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<SelectObjectContentOutput, SdkError<SelectObjectContentError>> where
R::Policy: SmithyRetryPolicy<SelectObjectContentInputOperationOutputAlias, SelectObjectContentOutput, SelectObjectContentError, SelectObjectContentInputOperationRetryAlias>,

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn bucket(self, inp: impl Into<String>) -> Self

The S3 bucket.

pub fn set_bucket(self, input: Option<String>) -> Self

The S3 bucket.

pub fn key(self, inp: impl Into<String>) -> Self

The object key.

pub fn set_key(self, input: Option<String>) -> Self

The object key.

pub fn sse_customer_algorithm(self, inp: impl Into<String>) -> Self

The SSE Algorithm used to encrypt the object. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

The SSE Algorithm used to encrypt the object. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn sse_customer_key(self, inp: impl Into<String>) -> Self

The SSE Customer Key. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn set_sse_customer_key(self, input: Option<String>) -> Self

The SSE Customer Key. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn sse_customer_key_md5(self, inp: impl Into<String>) -> Self

The SSE Customer Key MD5. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

The SSE Customer Key MD5. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn expression(self, inp: impl Into<String>) -> Self

The expression that is used to query the object.

pub fn set_expression(self, input: Option<String>) -> Self

The expression that is used to query the object.

pub fn expression_type(self, inp: ExpressionType) -> Self

The type of the provided expression (for example, SQL).

pub fn set_expression_type(self, input: Option<ExpressionType>) -> Self

The type of the provided expression (for example, SQL).

pub fn request_progress(self, inp: RequestProgress) -> Self

Specifies if periodic request progress information should be enabled.

pub fn set_request_progress(self, input: Option<RequestProgress>) -> Self

Specifies if periodic request progress information should be enabled.

pub fn input_serialization(self, inp: InputSerialization) -> Self

Describes the format of the data in the object that is being queried.

pub fn set_input_serialization(self, input: Option<InputSerialization>) -> Self

Describes the format of the data in the object that is being queried.

pub fn output_serialization(self, inp: OutputSerialization) -> Self

Describes the format of the data that you want Amazon S3 to return in response.

pub fn set_output_serialization(
self,
input: Option<OutputSerialization>
) -> Self

Describes the format of the data that you want Amazon S3 to return in response.

pub fn scan_range(self, inp: ScanRange) -> Self

Specifies the byte range of the object to get the records from. A record is processed when its first byte is contained by the range. This parameter is optional, but when specified, it must not be empty. See RFC 2616, Section 14.35.1 about how to specify the start and end of the range.

ScanRangemay be used in the following ways:

50100 - process only the records starting between the bytes 50 and 100 (inclusive, counting from zero)

50 - process only the records starting after the byte 50

50 - process only the records within the last 50 bytes of the file.

pub fn set_scan_range(self, input: Option<ScanRange>) -> Self

Specifies the byte range of the object to get the records from. A record is processed when its first byte is contained by the range. This parameter is optional, but when specified, it must not be empty. See RFC 2616, Section 14.35.1 about how to specify the start and end of the range.

ScanRangemay be used in the following ways:

50100 - process only the records starting between the bytes 50 and 100 (inclusive, counting from zero)

50 - process only the records starting after the byte 50

50 - process only the records within the last 50 bytes of the file.

pub fn expected_bucket_owner(self, inp: impl Into<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, input: Option<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.