Struct aws_sdk_s3::client::fluent_builders::SelectObjectContent

pub struct SelectObjectContent<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> SelectObjectContent<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<SelectObjectContentOutput, SdkError<SelectObjectContentError>> where
R::Policy: SmithyRetryPolicy<SelectObjectContentInputOperationOutputAlias, SelectObjectContentOutput, SelectObjectContentError, SelectObjectContentInputOperationRetryAlias>,

pub fn bucket(self, inp: impl Into<String>) -> Self

The S3 bucket.

pub fn set_bucket(self, input: Option<String>) -> Self

pub fn key(self, inp: impl Into<String>) -> Self

The object key.

pub fn set_key(self, input: Option<String>) -> Self

pub fn sse_customer_algorithm(self, inp: impl Into<String>) -> Self

The SSE Algorithm used to encrypt the object. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

pub fn sse_customer_key(self, inp: impl Into<String>) -> Self

The SSE Customer Key. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn set_sse_customer_key(self, input: Option<String>) -> Self

pub fn sse_customer_key_md5(self, inp: impl Into<String>) -> Self

The SSE Customer Key MD5. For more information, see Server-Side Encryption (Using Customer-Provided Encryption Keys.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

pub fn expression(self, inp: impl Into<String>) -> Self

The expression that is used to query the object.

pub fn set_expression(self, input: Option<String>) -> Self

pub fn expression_type(self, inp: ExpressionType) -> Self

The type of the provided expression (for example, SQL).

pub fn set_expression_type(self, input: Option<ExpressionType>) -> Self

pub fn request_progress(self, inp: RequestProgress) -> Self

Specifies if periodic request progress information should be enabled.

pub fn set_request_progress(self, input: Option<RequestProgress>) -> Self

pub fn input_serialization(self, inp: InputSerialization) -> Self

Describes the format of the data in the object that is being queried.

pub fn set_input_serialization(self, input: Option<InputSerialization>) -> Self

pub fn output_serialization(self, inp: OutputSerialization) -> Self

Describes the format of the data that you want Amazon S3 to return in response.

pub fn set_output_serialization(
self,
input: Option<OutputSerialization>
) -> Self

pub fn scan_range(self, inp: ScanRange) -> Self

Specifies the byte range of the object to get the records from. A record is processed when its first byte is contained by the range. This parameter is optional, but when specified, it must not be empty. See RFC 2616, Section 14.35.1 about how to specify the start and end of the range.

ScanRangemay be used in the following ways:

50100 - process only the records starting between the bytes 50 and 100 (inclusive, counting from zero)

50 - process only the records starting after the byte 50

50 - process only the records within the last 50 bytes of the file.

pub fn set_scan_range(self, input: Option<ScanRange>) -> Self

pub fn expected_bucket_owner(self, inp: impl Into<String>) -> Self

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

pub fn set_expected_bucket_owner(self, input: Option<String>) -> Self