Enum aws_sdk_s3::Error

#[non_exhaustive]
pub enum Error {
    BucketAlreadyExists(BucketAlreadyExists),
    BucketAlreadyOwnedByYou(BucketAlreadyOwnedByYou),
    InvalidObjectState(InvalidObjectState),
    NoSuchBucket(NoSuchBucket),
    NoSuchKey(NoSuchKey),
    NoSuchUpload(NoSuchUpload),
    NotFound(NotFound),
    ObjectAlreadyInActiveTierError(ObjectAlreadyInActiveTierError),
    ObjectNotInActiveTierError(ObjectNotInActiveTierError),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),
}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BucketAlreadyExists(BucketAlreadyExists)

The requested bucket name is not available. The bucket namespace is shared by all users of the system. Select a different name and try again.

Tuple Fields of BucketAlreadyExists

0: BucketAlreadyExists

BucketAlreadyOwnedByYou(BucketAlreadyOwnedByYou)

The bucket you tried to create already exists, and you own it. Amazon S3 returns this error in all Amazon Web Services Regions except in the North Virginia Region. For legacy compatibility, if you re-create an existing bucket that you already own in the North Virginia Region, Amazon S3 returns 200 OK and resets the bucket access control lists (ACLs).

Tuple Fields of BucketAlreadyOwnedByYou

0: BucketAlreadyOwnedByYou

InvalidObjectState(InvalidObjectState)

Object is archived and inaccessible until restored.

Tuple Fields of InvalidObjectState

0: InvalidObjectState

NoSuchBucket(NoSuchBucket)

The specified bucket does not exist.

Tuple Fields of NoSuchBucket

0: NoSuchBucket

NoSuchKey(NoSuchKey)

The specified key does not exist.

Tuple Fields of NoSuchKey

0: NoSuchKey

NoSuchUpload(NoSuchUpload)

The specified multipart upload does not exist.

Tuple Fields of NoSuchUpload

0: NoSuchUpload

NotFound(NotFound)

The specified content does not exist.

Tuple Fields of NotFound

0: NotFound

ObjectAlreadyInActiveTierError(ObjectAlreadyInActiveTierError)

This action is not allowed against this storage tier.

Tuple Fields of ObjectAlreadyInActiveTierError

0: ObjectAlreadyInActiveTierError

ObjectNotInActiveTierError(ObjectNotInActiveTierError)

The source object of the COPY action is not in the active tier and is only stored in Amazon S3 Glacier.

Tuple Fields of ObjectNotInActiveTierError

0: ObjectNotInActiveTierError

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>