Enum aws_sdk_s3::Error

#[non_exhaustive]
pub enum Error {
    BucketAlreadyExists(BucketAlreadyExists),
    BucketAlreadyOwnedByYou(BucketAlreadyOwnedByYou),
    InvalidObjectState(InvalidObjectState),
    NoSuchBucket(NoSuchBucket),
    NoSuchKey(NoSuchKey),
    NoSuchUpload(NoSuchUpload),
    NotFound(NotFound),
    ObjectAlreadyInActiveTierError(ObjectAlreadyInActiveTierError),
    ObjectNotInActiveTierError(ObjectNotInActiveTierError),
    Unhandled(Unhandled),
}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BucketAlreadyExists(BucketAlreadyExists)

The requested bucket name is not available. The bucket namespace is shared by all users of the system. Select a different name and try again.

BucketAlreadyOwnedByYou(BucketAlreadyOwnedByYou)

The bucket you tried to create already exists, and you own it. Amazon S3 returns this error in all Amazon Web Services Regions except in the North Virginia Region. For legacy compatibility, if you re-create an existing bucket that you already own in the North Virginia Region, Amazon S3 returns 200 OK and resets the bucket access control lists (ACLs).

InvalidObjectState(InvalidObjectState)

Object is archived and inaccessible until restored.

NoSuchBucket(NoSuchBucket)

The specified bucket does not exist.

NoSuchKey(NoSuchKey)

The specified key does not exist.

NoSuchUpload(NoSuchUpload)

The specified multipart upload does not exist.

NotFound(NotFound)

The specified content does not exist.

ObjectAlreadyInActiveTierError(ObjectAlreadyInActiveTierError)

This action is not allowed against this storage tier.

ObjectNotInActiveTierError(ObjectNotInActiveTierError)

The source object of the COPY action is not in the active tier and is only stored in Amazon S3 Glacier.

Unhandled(Unhandled)

An unexpected error occurred (e.g., invalid JSON returned by the service or an unknown error code).

When logging an error from the SDK, it is recommended that you either wrap the error in DisplayErrorContext, use another error reporter library that visits the error’s cause/source chain, or call Error::source for more details about the underlying cause.