Module aws_sdk_s3::error

Modules

bucket_already_exists
See BucketAlreadyExists

bucket_already_owned_by_you
See BucketAlreadyOwnedByYou

invalid_object_state
See InvalidObjectState

no_such_bucket
See NoSuchBucket

no_such_key
See NoSuchKey

no_such_upload
See NoSuchUpload

not_found
See NotFound

object_already_in_active_tier_error
See ObjectAlreadyInActiveTierError

object_not_in_active_tier_error
See ObjectNotInActiveTierError

Structs

AbortMultipartUploadError

BucketAlreadyExists
The requested bucket name is not available. The bucket namespace is shared by all users

BucketAlreadyOwnedByYou
The bucket you tried to create already exists, and you own it. Amazon S3 returns this error

CompleteMultipartUploadError

CopyObjectError

CreateBucketError

CreateMultipartUploadError

DeleteBucketAnalyticsConfigurationError

DeleteBucketCorsError

DeleteBucketEncryptionError

DeleteBucketError

DeleteBucketIntelligentTieringConfigurationError

DeleteBucketInventoryConfigurationError

DeleteBucketLifecycleError

DeleteBucketMetricsConfigurationError

DeleteBucketOwnershipControlsError

DeleteBucketPolicyError

DeleteBucketReplicationError

DeleteBucketTaggingError

DeleteBucketWebsiteError

DeleteObjectError

DeleteObjectTaggingError

DeleteObjectsError

DeletePublicAccessBlockError

GetBucketAccelerateConfigurationError

GetBucketAclError

GetBucketAnalyticsConfigurationError

GetBucketCorsError

GetBucketEncryptionError

GetBucketIntelligentTieringConfigurationError

GetBucketInventoryConfigurationError

GetBucketLifecycleConfigurationError

GetBucketLocationError

GetBucketLoggingError

GetBucketMetricsConfigurationError

GetBucketNotificationConfigurationError

GetBucketOwnershipControlsError

GetBucketPolicyError

GetBucketPolicyStatusError

GetBucketReplicationError

GetBucketRequestPaymentError

GetBucketTaggingError

GetBucketVersioningError

GetBucketWebsiteError

GetObjectAclError

GetObjectError

GetObjectLegalHoldError

GetObjectLockConfigurationError

GetObjectRetentionError

GetObjectTaggingError

GetObjectTorrentError

GetPublicAccessBlockError

HeadBucketError

HeadObjectError

InvalidObjectState
Object is archived and inaccessible until restored.

ListBucketAnalyticsConfigurationsError

ListBucketIntelligentTieringConfigurationsError

ListBucketInventoryConfigurationsError

ListBucketMetricsConfigurationsError

ListBucketsError

ListMultipartUploadsError

ListObjectVersionsError

ListObjectsError

ListObjectsV2Error

ListPartsError

NoSuchBucket
The specified bucket does not exist.

NoSuchKey
The specified key does not exist.

NoSuchUpload
The specified multipart upload does not exist.

NotFound
The specified content does not exist.

ObjectAlreadyInActiveTierError
This action is not allowed against this storage tier.

ObjectNotInActiveTierError
The source object of the COPY action is not in the active tier and is only stored in

PutBucketAccelerateConfigurationError

PutBucketAclError

PutBucketAnalyticsConfigurationError

PutBucketCorsError

PutBucketEncryptionError

PutBucketIntelligentTieringConfigurationError

PutBucketInventoryConfigurationError

PutBucketLifecycleConfigurationError

PutBucketLoggingError

PutBucketMetricsConfigurationError

PutBucketNotificationConfigurationError

PutBucketOwnershipControlsError

PutBucketPolicyError

PutBucketReplicationError

PutBucketRequestPaymentError

PutBucketTaggingError

PutBucketVersioningError

PutBucketWebsiteError

PutObjectAclError

PutObjectError

PutObjectLegalHoldError

PutObjectLockConfigurationError

PutObjectRetentionError

PutObjectTaggingError

PutPublicAccessBlockError

RestoreObjectError

UploadPartCopyError

UploadPartError

WriteGetObjectResponseError

Enums

AbortMultipartUploadErrorKind

CompleteMultipartUploadErrorKind

CopyObjectErrorKind

CreateBucketErrorKind

CreateMultipartUploadErrorKind

DeleteBucketAnalyticsConfigurationErrorKind

DeleteBucketCorsErrorKind

DeleteBucketEncryptionErrorKind

DeleteBucketErrorKind

DeleteBucketIntelligentTieringConfigurationErrorKind

DeleteBucketInventoryConfigurationErrorKind

DeleteBucketLifecycleErrorKind

DeleteBucketMetricsConfigurationErrorKind

DeleteBucketOwnershipControlsErrorKind

DeleteBucketPolicyErrorKind

DeleteBucketReplicationErrorKind

DeleteBucketTaggingErrorKind

DeleteBucketWebsiteErrorKind

DeleteObjectErrorKind

DeleteObjectTaggingErrorKind

DeleteObjectsErrorKind

DeletePublicAccessBlockErrorKind

GetBucketAccelerateConfigurationErrorKind

GetBucketAclErrorKind

GetBucketAnalyticsConfigurationErrorKind

GetBucketCorsErrorKind

GetBucketEncryptionErrorKind

GetBucketIntelligentTieringConfigurationErrorKind

GetBucketInventoryConfigurationErrorKind

GetBucketLifecycleConfigurationErrorKind

GetBucketLocationErrorKind

GetBucketLoggingErrorKind

GetBucketMetricsConfigurationErrorKind

GetBucketNotificationConfigurationErrorKind

GetBucketOwnershipControlsErrorKind

GetBucketPolicyErrorKind

GetBucketPolicyStatusErrorKind

GetBucketReplicationErrorKind

GetBucketRequestPaymentErrorKind

GetBucketTaggingErrorKind

GetBucketVersioningErrorKind

GetBucketWebsiteErrorKind

GetObjectAclErrorKind

GetObjectErrorKind

GetObjectLegalHoldErrorKind

GetObjectLockConfigurationErrorKind

GetObjectRetentionErrorKind

GetObjectTaggingErrorKind

GetObjectTorrentErrorKind

GetPublicAccessBlockErrorKind

HeadBucketErrorKind

HeadObjectErrorKind

ListBucketAnalyticsConfigurationsErrorKind

ListBucketIntelligentTieringConfigurationsErrorKind

ListBucketInventoryConfigurationsErrorKind

ListBucketMetricsConfigurationsErrorKind

ListBucketsErrorKind

ListMultipartUploadsErrorKind

ListObjectVersionsErrorKind

ListObjectsErrorKind

ListObjectsV2ErrorKind

ListPartsErrorKind

PutBucketAccelerateConfigurationErrorKind

PutBucketAclErrorKind

PutBucketAnalyticsConfigurationErrorKind

PutBucketCorsErrorKind

PutBucketEncryptionErrorKind

PutBucketIntelligentTieringConfigurationErrorKind

PutBucketInventoryConfigurationErrorKind

PutBucketLifecycleConfigurationErrorKind

PutBucketLoggingErrorKind

PutBucketMetricsConfigurationErrorKind

PutBucketNotificationConfigurationErrorKind

PutBucketOwnershipControlsErrorKind

PutBucketPolicyErrorKind

PutBucketReplicationErrorKind

PutBucketRequestPaymentErrorKind

PutBucketTaggingErrorKind

PutBucketVersioningErrorKind

PutBucketWebsiteErrorKind

PutObjectAclErrorKind

PutObjectErrorKind

PutObjectLegalHoldErrorKind

PutObjectLockConfigurationErrorKind

PutObjectRetentionErrorKind

PutObjectTaggingErrorKind

PutPublicAccessBlockErrorKind

RestoreObjectErrorKind

UploadPartCopyErrorKind

UploadPartErrorKind

WriteGetObjectResponseErrorKind