Module aws_sdk_s3::error

Errors that can occur when calling the service.

Modules

bucket_already_exists

See BucketAlreadyExists

bucket_already_owned_by_you

See BucketAlreadyOwnedByYou

invalid_object_state

See InvalidObjectState

no_such_bucket

See NoSuchBucket

no_such_key

See NoSuchKey

no_such_upload

See NoSuchUpload

not_found

See NotFound

object_already_in_active_tier_error

See ObjectAlreadyInActiveTierError

object_not_in_active_tier_error

See ObjectNotInActiveTierError

Structs

AbortMultipartUploadError

Error type for the AbortMultipartUpload operation.

BucketAlreadyExists

The requested bucket name is not available. The bucket namespace is shared by all users of the system. Select a different name and try again.

BucketAlreadyOwnedByYou

The bucket you tried to create already exists, and you own it. Amazon S3 returns this error in all Amazon Web Services Regions except in the North Virginia Region. For legacy compatibility, if you re-create an existing bucket that you already own in the North Virginia Region, Amazon S3 returns 200 OK and resets the bucket access control lists (ACLs).

CompleteMultipartUploadError

Error type for the CompleteMultipartUpload operation.

CopyObjectError

Error type for the CopyObject operation.

CreateBucketError

Error type for the CreateBucket operation.

CreateMultipartUploadError

Error type for the CreateMultipartUpload operation.

DeleteBucketAnalyticsConfigurationError

Error type for the DeleteBucketAnalyticsConfiguration operation.

DeleteBucketCorsError

Error type for the DeleteBucketCors operation.

DeleteBucketEncryptionError

Error type for the DeleteBucketEncryption operation.

DeleteBucketError

Error type for the DeleteBucket operation.

DeleteBucketIntelligentTieringConfigurationError

Error type for the DeleteBucketIntelligentTieringConfiguration operation.

DeleteBucketInventoryConfigurationError

Error type for the DeleteBucketInventoryConfiguration operation.

DeleteBucketLifecycleError

Error type for the DeleteBucketLifecycle operation.

DeleteBucketMetricsConfigurationError

Error type for the DeleteBucketMetricsConfiguration operation.

DeleteBucketOwnershipControlsError

Error type for the DeleteBucketOwnershipControls operation.

DeleteBucketPolicyError

Error type for the DeleteBucketPolicy operation.

DeleteBucketReplicationError

Error type for the DeleteBucketReplication operation.

DeleteBucketTaggingError

Error type for the DeleteBucketTagging operation.

DeleteBucketWebsiteError

Error type for the DeleteBucketWebsite operation.

DeleteObjectError

Error type for the DeleteObject operation.

DeleteObjectTaggingError

Error type for the DeleteObjectTagging operation.

DeleteObjectsError

Error type for the DeleteObjects operation.

DeletePublicAccessBlockError

Error type for the DeletePublicAccessBlock operation.

GetBucketAccelerateConfigurationError

Error type for the GetBucketAccelerateConfiguration operation.

GetBucketAclError

Error type for the GetBucketAcl operation.

GetBucketAnalyticsConfigurationError

Error type for the GetBucketAnalyticsConfiguration operation.

GetBucketCorsError

Error type for the GetBucketCors operation.

GetBucketEncryptionError

Error type for the GetBucketEncryption operation.

GetBucketIntelligentTieringConfigurationError

Error type for the GetBucketIntelligentTieringConfiguration operation.

GetBucketInventoryConfigurationError

Error type for the GetBucketInventoryConfiguration operation.

GetBucketLifecycleConfigurationError

Error type for the GetBucketLifecycleConfiguration operation.

GetBucketLocationError

Error type for the GetBucketLocation operation.

GetBucketLoggingError

Error type for the GetBucketLogging operation.

GetBucketMetricsConfigurationError

Error type for the GetBucketMetricsConfiguration operation.

GetBucketNotificationConfigurationError

Error type for the GetBucketNotificationConfiguration operation.

GetBucketOwnershipControlsError

Error type for the GetBucketOwnershipControls operation.

GetBucketPolicyError

Error type for the GetBucketPolicy operation.

GetBucketPolicyStatusError

Error type for the GetBucketPolicyStatus operation.

GetBucketReplicationError

Error type for the GetBucketReplication operation.

GetBucketRequestPaymentError

Error type for the GetBucketRequestPayment operation.

GetBucketTaggingError

Error type for the GetBucketTagging operation.

GetBucketVersioningError

Error type for the GetBucketVersioning operation.

GetBucketWebsiteError

Error type for the GetBucketWebsite operation.

GetObjectAclError

Error type for the GetObjectAcl operation.

GetObjectError

Error type for the GetObject operation.

GetObjectLegalHoldError

Error type for the GetObjectLegalHold operation.

GetObjectLockConfigurationError

Error type for the GetObjectLockConfiguration operation.

GetObjectRetentionError

Error type for the GetObjectRetention operation.

GetObjectTaggingError

Error type for the GetObjectTagging operation.

GetObjectTorrentError

Error type for the GetObjectTorrent operation.

GetPublicAccessBlockError

Error type for the GetPublicAccessBlock operation.

HeadBucketError

Error type for the HeadBucket operation.

HeadObjectError

Error type for the HeadObject operation.

InvalidObjectState

Object is archived and inaccessible until restored.

ListBucketAnalyticsConfigurationsError

Error type for the ListBucketAnalyticsConfigurations operation.

ListBucketIntelligentTieringConfigurationsError

Error type for the ListBucketIntelligentTieringConfigurations operation.

ListBucketInventoryConfigurationsError

Error type for the ListBucketInventoryConfigurations operation.

ListBucketMetricsConfigurationsError

Error type for the ListBucketMetricsConfigurations operation.

ListBucketsError

Error type for the ListBuckets operation.

ListMultipartUploadsError

Error type for the ListMultipartUploads operation.

ListObjectVersionsError

Error type for the ListObjectVersions operation.

ListObjectsError

Error type for the ListObjects operation.

ListObjectsV2Error

Error type for the ListObjectsV2 operation.

ListPartsError

Error type for the ListParts operation.

NoSuchBucket

The specified bucket does not exist.

NoSuchKey

The specified key does not exist.

NoSuchUpload

The specified multipart upload does not exist.

NotFound

The specified content does not exist.

ObjectAlreadyInActiveTierError

This action is not allowed against this storage tier.

ObjectNotInActiveTierError

The source object of the COPY action is not in the active tier and is only stored in Amazon S3 Glacier.

PutBucketAccelerateConfigurationError

Error type for the PutBucketAccelerateConfiguration operation.

PutBucketAclError

Error type for the PutBucketAcl operation.

PutBucketAnalyticsConfigurationError

Error type for the PutBucketAnalyticsConfiguration operation.

PutBucketCorsError

Error type for the PutBucketCors operation.

PutBucketEncryptionError

Error type for the PutBucketEncryption operation.

PutBucketIntelligentTieringConfigurationError

Error type for the PutBucketIntelligentTieringConfiguration operation.

PutBucketInventoryConfigurationError

Error type for the PutBucketInventoryConfiguration operation.

PutBucketLifecycleConfigurationError

Error type for the PutBucketLifecycleConfiguration operation.

PutBucketLoggingError

Error type for the PutBucketLogging operation.

PutBucketMetricsConfigurationError

Error type for the PutBucketMetricsConfiguration operation.

PutBucketNotificationConfigurationError

Error type for the PutBucketNotificationConfiguration operation.

PutBucketOwnershipControlsError

Error type for the PutBucketOwnershipControls operation.

PutBucketPolicyError

Error type for the PutBucketPolicy operation.

PutBucketReplicationError

Error type for the PutBucketReplication operation.

PutBucketRequestPaymentError

Error type for the PutBucketRequestPayment operation.

PutBucketTaggingError

Error type for the PutBucketTagging operation.

PutBucketVersioningError

Error type for the PutBucketVersioning operation.

PutBucketWebsiteError

Error type for the PutBucketWebsite operation.

PutObjectAclError

Error type for the PutObjectAcl operation.

PutObjectError

Error type for the PutObject operation.

PutObjectLegalHoldError

Error type for the PutObjectLegalHold operation.

PutObjectLockConfigurationError

Error type for the PutObjectLockConfiguration operation.

PutObjectRetentionError

Error type for the PutObjectRetention operation.

PutObjectTaggingError

Error type for the PutObjectTagging operation.

PutPublicAccessBlockError

Error type for the PutPublicAccessBlock operation.

RestoreObjectError

Error type for the RestoreObject operation.

SelectObjectContentError

Error type for the SelectObjectContent operation.

UploadPartCopyError

Error type for the UploadPartCopy operation.

UploadPartError

Error type for the UploadPart operation.

WriteGetObjectResponseError

Error type for the WriteGetObjectResponse operation.

Enums

AbortMultipartUploadErrorKind

Types of errors that can occur for the AbortMultipartUpload operation.

CompleteMultipartUploadErrorKind

Types of errors that can occur for the CompleteMultipartUpload operation.

CopyObjectErrorKind

Types of errors that can occur for the CopyObject operation.

CreateBucketErrorKind

Types of errors that can occur for the CreateBucket operation.

CreateMultipartUploadErrorKind

Types of errors that can occur for the CreateMultipartUpload operation.

DeleteBucketAnalyticsConfigurationErrorKind

Types of errors that can occur for the DeleteBucketAnalyticsConfiguration operation.

DeleteBucketCorsErrorKind

Types of errors that can occur for the DeleteBucketCors operation.

DeleteBucketEncryptionErrorKind

Types of errors that can occur for the DeleteBucketEncryption operation.

DeleteBucketErrorKind

Types of errors that can occur for the DeleteBucket operation.

DeleteBucketIntelligentTieringConfigurationErrorKind

Types of errors that can occur for the DeleteBucketIntelligentTieringConfiguration operation.

DeleteBucketInventoryConfigurationErrorKind

Types of errors that can occur for the DeleteBucketInventoryConfiguration operation.

DeleteBucketLifecycleErrorKind

Types of errors that can occur for the DeleteBucketLifecycle operation.

DeleteBucketMetricsConfigurationErrorKind

Types of errors that can occur for the DeleteBucketMetricsConfiguration operation.

DeleteBucketOwnershipControlsErrorKind

Types of errors that can occur for the DeleteBucketOwnershipControls operation.

DeleteBucketPolicyErrorKind

Types of errors that can occur for the DeleteBucketPolicy operation.

DeleteBucketReplicationErrorKind

Types of errors that can occur for the DeleteBucketReplication operation.

DeleteBucketTaggingErrorKind

Types of errors that can occur for the DeleteBucketTagging operation.

DeleteBucketWebsiteErrorKind

Types of errors that can occur for the DeleteBucketWebsite operation.

DeleteObjectErrorKind

Types of errors that can occur for the DeleteObject operation.

DeleteObjectTaggingErrorKind

Types of errors that can occur for the DeleteObjectTagging operation.

DeleteObjectsErrorKind

Types of errors that can occur for the DeleteObjects operation.

DeletePublicAccessBlockErrorKind

Types of errors that can occur for the DeletePublicAccessBlock operation.

GetBucketAccelerateConfigurationErrorKind

Types of errors that can occur for the GetBucketAccelerateConfiguration operation.

GetBucketAclErrorKind

Types of errors that can occur for the GetBucketAcl operation.

GetBucketAnalyticsConfigurationErrorKind

Types of errors that can occur for the GetBucketAnalyticsConfiguration operation.

GetBucketCorsErrorKind

Types of errors that can occur for the GetBucketCors operation.

GetBucketEncryptionErrorKind

Types of errors that can occur for the GetBucketEncryption operation.

GetBucketIntelligentTieringConfigurationErrorKind

Types of errors that can occur for the GetBucketIntelligentTieringConfiguration operation.

GetBucketInventoryConfigurationErrorKind

Types of errors that can occur for the GetBucketInventoryConfiguration operation.

GetBucketLifecycleConfigurationErrorKind

Types of errors that can occur for the GetBucketLifecycleConfiguration operation.

GetBucketLocationErrorKind

Types of errors that can occur for the GetBucketLocation operation.

GetBucketLoggingErrorKind

Types of errors that can occur for the GetBucketLogging operation.

GetBucketMetricsConfigurationErrorKind

Types of errors that can occur for the GetBucketMetricsConfiguration operation.

GetBucketNotificationConfigurationErrorKind

Types of errors that can occur for the GetBucketNotificationConfiguration operation.

GetBucketOwnershipControlsErrorKind

Types of errors that can occur for the GetBucketOwnershipControls operation.

GetBucketPolicyErrorKind

Types of errors that can occur for the GetBucketPolicy operation.

GetBucketPolicyStatusErrorKind

Types of errors that can occur for the GetBucketPolicyStatus operation.

GetBucketReplicationErrorKind

Types of errors that can occur for the GetBucketReplication operation.

GetBucketRequestPaymentErrorKind

Types of errors that can occur for the GetBucketRequestPayment operation.

GetBucketTaggingErrorKind

Types of errors that can occur for the GetBucketTagging operation.

GetBucketVersioningErrorKind

Types of errors that can occur for the GetBucketVersioning operation.

GetBucketWebsiteErrorKind

Types of errors that can occur for the GetBucketWebsite operation.

GetObjectAclErrorKind

Types of errors that can occur for the GetObjectAcl operation.

GetObjectErrorKind

Types of errors that can occur for the GetObject operation.

GetObjectLegalHoldErrorKind

Types of errors that can occur for the GetObjectLegalHold operation.

GetObjectLockConfigurationErrorKind

Types of errors that can occur for the GetObjectLockConfiguration operation.

GetObjectRetentionErrorKind

Types of errors that can occur for the GetObjectRetention operation.

GetObjectTaggingErrorKind

Types of errors that can occur for the GetObjectTagging operation.

GetObjectTorrentErrorKind

Types of errors that can occur for the GetObjectTorrent operation.

GetPublicAccessBlockErrorKind

Types of errors that can occur for the GetPublicAccessBlock operation.

HeadBucketErrorKind

Types of errors that can occur for the HeadBucket operation.

HeadObjectErrorKind

Types of errors that can occur for the HeadObject operation.

ListBucketAnalyticsConfigurationsErrorKind

Types of errors that can occur for the ListBucketAnalyticsConfigurations operation.

ListBucketIntelligentTieringConfigurationsErrorKind

Types of errors that can occur for the ListBucketIntelligentTieringConfigurations operation.

ListBucketInventoryConfigurationsErrorKind

Types of errors that can occur for the ListBucketInventoryConfigurations operation.

ListBucketMetricsConfigurationsErrorKind

Types of errors that can occur for the ListBucketMetricsConfigurations operation.

ListBucketsErrorKind

Types of errors that can occur for the ListBuckets operation.

ListMultipartUploadsErrorKind

Types of errors that can occur for the ListMultipartUploads operation.

ListObjectVersionsErrorKind

Types of errors that can occur for the ListObjectVersions operation.

ListObjectsErrorKind

Types of errors that can occur for the ListObjects operation.

ListObjectsV2ErrorKind

Types of errors that can occur for the ListObjectsV2 operation.

ListPartsErrorKind

Types of errors that can occur for the ListParts operation.

PutBucketAccelerateConfigurationErrorKind

Types of errors that can occur for the PutBucketAccelerateConfiguration operation.

PutBucketAclErrorKind

Types of errors that can occur for the PutBucketAcl operation.

PutBucketAnalyticsConfigurationErrorKind

Types of errors that can occur for the PutBucketAnalyticsConfiguration operation.

PutBucketCorsErrorKind

Types of errors that can occur for the PutBucketCors operation.

PutBucketEncryptionErrorKind

Types of errors that can occur for the PutBucketEncryption operation.

PutBucketIntelligentTieringConfigurationErrorKind

Types of errors that can occur for the PutBucketIntelligentTieringConfiguration operation.

PutBucketInventoryConfigurationErrorKind

Types of errors that can occur for the PutBucketInventoryConfiguration operation.

PutBucketLifecycleConfigurationErrorKind

Types of errors that can occur for the PutBucketLifecycleConfiguration operation.

PutBucketLoggingErrorKind

Types of errors that can occur for the PutBucketLogging operation.

PutBucketMetricsConfigurationErrorKind

Types of errors that can occur for the PutBucketMetricsConfiguration operation.

PutBucketNotificationConfigurationErrorKind

Types of errors that can occur for the PutBucketNotificationConfiguration operation.

PutBucketOwnershipControlsErrorKind

Types of errors that can occur for the PutBucketOwnershipControls operation.

PutBucketPolicyErrorKind

Types of errors that can occur for the PutBucketPolicy operation.

PutBucketReplicationErrorKind

Types of errors that can occur for the PutBucketReplication operation.

PutBucketRequestPaymentErrorKind

Types of errors that can occur for the PutBucketRequestPayment operation.

PutBucketTaggingErrorKind

Types of errors that can occur for the PutBucketTagging operation.

PutBucketVersioningErrorKind

Types of errors that can occur for the PutBucketVersioning operation.

PutBucketWebsiteErrorKind

Types of errors that can occur for the PutBucketWebsite operation.

PutObjectAclErrorKind

Types of errors that can occur for the PutObjectAcl operation.

PutObjectErrorKind

Types of errors that can occur for the PutObject operation.

PutObjectLegalHoldErrorKind

Types of errors that can occur for the PutObjectLegalHold operation.

PutObjectLockConfigurationErrorKind

Types of errors that can occur for the PutObjectLockConfiguration operation.

PutObjectRetentionErrorKind

Types of errors that can occur for the PutObjectRetention operation.

PutObjectTaggingErrorKind

Types of errors that can occur for the PutObjectTagging operation.

PutPublicAccessBlockErrorKind

Types of errors that can occur for the PutPublicAccessBlock operation.

RestoreObjectErrorKind

Types of errors that can occur for the RestoreObject operation.

SelectObjectContentErrorKind

Types of errors that can occur for the SelectObjectContent operation.

UploadPartCopyErrorKind

Types of errors that can occur for the UploadPartCopy operation.

UploadPartErrorKind

Types of errors that can occur for the UploadPart operation.

WriteGetObjectResponseErrorKind

Types of errors that can occur for the WriteGetObjectResponse operation.