Module aws_sdk_s3::input

Input structures for operations.

Modules

abort_multipart_upload_input

See AbortMultipartUploadInput

complete_multipart_upload_input

See CompleteMultipartUploadInput

copy_object_input

See CopyObjectInput

create_bucket_input

See CreateBucketInput

create_multipart_upload_input

See CreateMultipartUploadInput

delete_bucket_analytics_configuration_input

See DeleteBucketAnalyticsConfigurationInput

delete_bucket_cors_input

See DeleteBucketCorsInput

delete_bucket_encryption_input

See DeleteBucketEncryptionInput

delete_bucket_input

See DeleteBucketInput

delete_bucket_intelligent_tiering_configuration_input

See DeleteBucketIntelligentTieringConfigurationInput

delete_bucket_inventory_configuration_input

See DeleteBucketInventoryConfigurationInput

delete_bucket_lifecycle_input

See DeleteBucketLifecycleInput

delete_bucket_metrics_configuration_input

See DeleteBucketMetricsConfigurationInput

delete_bucket_ownership_controls_input

See DeleteBucketOwnershipControlsInput

delete_bucket_policy_input

See DeleteBucketPolicyInput

delete_bucket_replication_input

See DeleteBucketReplicationInput

delete_bucket_tagging_input

See DeleteBucketTaggingInput

delete_bucket_website_input

See DeleteBucketWebsiteInput

delete_object_input

See DeleteObjectInput

delete_object_tagging_input

See DeleteObjectTaggingInput

delete_objects_input

See DeleteObjectsInput

delete_public_access_block_input

See DeletePublicAccessBlockInput

get_bucket_accelerate_configuration_input

See GetBucketAccelerateConfigurationInput

get_bucket_acl_input

See GetBucketAclInput

get_bucket_analytics_configuration_input

See GetBucketAnalyticsConfigurationInput

get_bucket_cors_input

See GetBucketCorsInput

get_bucket_encryption_input

See GetBucketEncryptionInput

get_bucket_intelligent_tiering_configuration_input

See GetBucketIntelligentTieringConfigurationInput

get_bucket_inventory_configuration_input

See GetBucketInventoryConfigurationInput

get_bucket_lifecycle_configuration_input

See GetBucketLifecycleConfigurationInput

get_bucket_location_input

See GetBucketLocationInput

get_bucket_logging_input

See GetBucketLoggingInput

get_bucket_metrics_configuration_input

See GetBucketMetricsConfigurationInput

get_bucket_notification_configuration_input

See GetBucketNotificationConfigurationInput

get_bucket_ownership_controls_input

See GetBucketOwnershipControlsInput

get_bucket_policy_input

See GetBucketPolicyInput

get_bucket_policy_status_input

See GetBucketPolicyStatusInput

get_bucket_replication_input

See GetBucketReplicationInput

get_bucket_request_payment_input

See GetBucketRequestPaymentInput

get_bucket_tagging_input

See GetBucketTaggingInput

get_bucket_versioning_input

See GetBucketVersioningInput

get_bucket_website_input

See GetBucketWebsiteInput

get_object_acl_input

See GetObjectAclInput

get_object_input

See GetObjectInput

get_object_legal_hold_input

See GetObjectLegalHoldInput

get_object_lock_configuration_input

See GetObjectLockConfigurationInput

get_object_retention_input

See GetObjectRetentionInput

get_object_tagging_input

See GetObjectTaggingInput

get_object_torrent_input

See GetObjectTorrentInput

get_public_access_block_input

See GetPublicAccessBlockInput

head_bucket_input

See HeadBucketInput

head_object_input

See HeadObjectInput

list_bucket_analytics_configurations_input

See ListBucketAnalyticsConfigurationsInput

list_bucket_intelligent_tiering_configurations_input

See ListBucketIntelligentTieringConfigurationsInput

list_bucket_inventory_configurations_input

See ListBucketInventoryConfigurationsInput

list_bucket_metrics_configurations_input

See ListBucketMetricsConfigurationsInput

list_buckets_input

See ListBucketsInput

list_multipart_uploads_input

See ListMultipartUploadsInput

list_object_versions_input

See ListObjectVersionsInput

list_objects_input

See ListObjectsInput

list_objects_v2_input

See ListObjectsV2Input

list_parts_input

See ListPartsInput

put_bucket_accelerate_configuration_input

See PutBucketAccelerateConfigurationInput

put_bucket_acl_input

See PutBucketAclInput

put_bucket_analytics_configuration_input

See PutBucketAnalyticsConfigurationInput

put_bucket_cors_input

See PutBucketCorsInput

put_bucket_encryption_input

See PutBucketEncryptionInput

put_bucket_intelligent_tiering_configuration_input

See PutBucketIntelligentTieringConfigurationInput

put_bucket_inventory_configuration_input

See PutBucketInventoryConfigurationInput

put_bucket_lifecycle_configuration_input

See PutBucketLifecycleConfigurationInput

put_bucket_logging_input

See PutBucketLoggingInput

put_bucket_metrics_configuration_input

See PutBucketMetricsConfigurationInput

put_bucket_notification_configuration_input

See PutBucketNotificationConfigurationInput

put_bucket_ownership_controls_input

See PutBucketOwnershipControlsInput

put_bucket_policy_input

See PutBucketPolicyInput

put_bucket_replication_input

See PutBucketReplicationInput

put_bucket_request_payment_input

See PutBucketRequestPaymentInput

put_bucket_tagging_input

See PutBucketTaggingInput

put_bucket_versioning_input

See PutBucketVersioningInput

put_bucket_website_input

See PutBucketWebsiteInput

put_object_acl_input

See PutObjectAclInput

put_object_input

See PutObjectInput

put_object_legal_hold_input

See PutObjectLegalHoldInput

put_object_lock_configuration_input

See PutObjectLockConfigurationInput

put_object_retention_input

See PutObjectRetentionInput

put_object_tagging_input

See PutObjectTaggingInput

put_public_access_block_input

See PutPublicAccessBlockInput

restore_object_input

See RestoreObjectInput

select_object_content_input

See SelectObjectContentInput

upload_part_copy_input

See UploadPartCopyInput

upload_part_input

See UploadPartInput

write_get_object_response_input

See WriteGetObjectResponseInput

Structs

AbortMultipartUploadInput

CompleteMultipartUploadInput

CopyObjectInput

CreateBucketInput

CreateMultipartUploadInput

DeleteBucketAnalyticsConfigurationInput

DeleteBucketCorsInput

DeleteBucketEncryptionInput

DeleteBucketInput

DeleteBucketIntelligentTieringConfigurationInput

DeleteBucketInventoryConfigurationInput

DeleteBucketLifecycleInput

DeleteBucketMetricsConfigurationInput

DeleteBucketOwnershipControlsInput

DeleteBucketPolicyInput

DeleteBucketReplicationInput

DeleteBucketTaggingInput

DeleteBucketWebsiteInput

DeleteObjectInput

DeleteObjectTaggingInput

DeleteObjectsInput

DeletePublicAccessBlockInput

GetBucketAccelerateConfigurationInput

GetBucketAclInput

GetBucketAnalyticsConfigurationInput

GetBucketCorsInput

GetBucketEncryptionInput

GetBucketIntelligentTieringConfigurationInput

GetBucketInventoryConfigurationInput

GetBucketLifecycleConfigurationInput

GetBucketLocationInput

GetBucketLoggingInput

GetBucketMetricsConfigurationInput

GetBucketNotificationConfigurationInput

GetBucketOwnershipControlsInput

GetBucketPolicyInput

GetBucketPolicyStatusInput

GetBucketReplicationInput

GetBucketRequestPaymentInput

GetBucketTaggingInput

GetBucketVersioningInput

GetBucketWebsiteInput

GetObjectAclInput

GetObjectInput

GetObjectLegalHoldInput

GetObjectLockConfigurationInput

GetObjectRetentionInput

GetObjectTaggingInput

GetObjectTorrentInput

GetPublicAccessBlockInput

HeadBucketInput

HeadObjectInput

ListBucketAnalyticsConfigurationsInput

ListBucketIntelligentTieringConfigurationsInput

ListBucketInventoryConfigurationsInput

ListBucketMetricsConfigurationsInput

ListBucketsInput

ListMultipartUploadsInput

ListObjectVersionsInput

ListObjectsInput

ListObjectsV2Input

ListPartsInput

PutBucketAccelerateConfigurationInput

PutBucketAclInput

PutBucketAnalyticsConfigurationInput

PutBucketCorsInput

PutBucketEncryptionInput

PutBucketIntelligentTieringConfigurationInput

PutBucketInventoryConfigurationInput

PutBucketLifecycleConfigurationInput

PutBucketLoggingInput

PutBucketMetricsConfigurationInput

PutBucketNotificationConfigurationInput

PutBucketOwnershipControlsInput

PutBucketPolicyInput

PutBucketReplicationInput

PutBucketRequestPaymentInput

PutBucketTaggingInput

PutBucketVersioningInput

PutBucketWebsiteInput

PutObjectAclInput

PutObjectInput

PutObjectLegalHoldInput

PutObjectLockConfigurationInput

PutObjectRetentionInput

PutObjectTaggingInput

PutPublicAccessBlockInput

RestoreObjectInput

SelectObjectContentInput

Request to filter the contents of an Amazon S3 object based on a simple Structured Query Language (SQL) statement. In the request, along with the SQL expression, you must specify a data serialization format (JSON or CSV) of the object. Amazon S3 uses this to parse object data into records. It returns only records that match the specified SQL expression. You must also specify the data serialization format for the response. For more information, see S3Select API Documentation.

UploadPartCopyInput

UploadPartInput

WriteGetObjectResponseInput