Struct aws_sdk_s3::input::CreateMultipartUploadInput

#[non_exhaustive]
pub struct CreateMultipartUploadInput {

    pub acl: Option<ObjectCannedAcl>,
    pub bucket: Option<String>,
    pub cache_control: Option<String>,
    pub content_disposition: Option<String>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_type: Option<String>,
    pub expires: Option<DateTime>,
    pub grant_full_control: Option<String>,
    pub grant_read: Option<String>,
    pub grant_read_acp: Option<String>,
    pub grant_write_acp: Option<String>,
    pub key: Option<String>,
    pub metadata: Option<HashMap<String, String>>,
    pub server_side_encryption: Option<ServerSideEncryption>,
    pub storage_class: Option<StorageClass>,
    pub website_redirect_location: Option<String>,
    pub sse_customer_algorithm: Option<String>,
    pub sse_customer_key: Option<String>,
    pub sse_customer_key_md5: Option<String>,
    pub ssekms_key_id: Option<String>,
    pub ssekms_encryption_context: Option<String>,
    pub bucket_key_enabled: bool,
    pub request_payer: Option<RequestPayer>,
    pub tagging: Option<String>,
    pub object_lock_mode: Option<ObjectLockMode>,
    pub object_lock_retain_until_date: Option<DateTime>,
    pub object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>,
    pub expected_bucket_owner: Option<String>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

acl: Option<ObjectCannedAcl>

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

bucket: Option<String>

The name of the bucket to which to initiate the upload

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

cache_control: Option<String>

Specifies caching behavior along the request/reply chain.

content_disposition: Option<String>

Specifies presentational information for the object.

content_encoding: Option<String>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

content_language: Option<String>

The language the content is in.

content_type: Option<String>

A standard MIME type describing the format of the object data.

expires: Option<DateTime>

The date and time at which the object is no longer cacheable.

grant_full_control: Option<String>

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

grant_read: Option<String>

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

grant_read_acp: Option<String>

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

grant_write_acp: Option<String>

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

key: Option<String>

Object key for which the multipart upload is to be initiated.

metadata: Option<HashMap<String, String>>

A map of metadata to store with the object in S3.

server_side_encryption: Option<ServerSideEncryption>

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

storage_class: Option<StorageClass>

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

website_redirect_location: Option<String>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

sse_customer_algorithm: Option<String>

Specifies the algorithm to use to when encrypting the object (for example, AES256).

sse_customer_key: Option<String>

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

sse_customer_key_md5: Option<String>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

ssekms_key_id: Option<String>

Specifies the ID of the symmetric customer managed key to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

ssekms_encryption_context: Option<String>

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

bucket_key_enabled: bool

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with an object action doesn’t affect bucket-level settings for S3 Bucket Key.

request_payer: Option<RequestPayer>

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

tagging: Option<String>

The tag-set for the object. The tag-set must be encoded as URL Query parameters.

object_lock_mode: Option<ObjectLockMode>

Specifies the Object Lock mode that you want to apply to the uploaded object.

object_lock_retain_until_date: Option<DateTime>

Specifies the date and time when you want the Object Lock to expire.

object_lock_legal_hold_status: Option<ObjectLockLegalHoldStatus>

Specifies whether you want to apply a Legal Hold to the uploaded object.

expected_bucket_owner: Option<String>

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

Implementations

impl CreateMultipartUploadInput

pub async fn make_operation( &self, _config: &Config ) -> Result<Operation<CreateMultipartUpload, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CreateMultipartUpload>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CreateMultipartUploadInput

impl CreateMultipartUploadInput

pub fn acl(&self) -> Option<&ObjectCannedAcl>

The canned ACL to apply to the object.

This action is not supported by Amazon S3 on Outposts.

pub fn bucket(&self) -> Option<&str>

The name of the bucket to which to initiate the upload

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

pub fn cache_control(&self) -> Option<&str>

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(&self) -> Option<&str>

Specifies presentational information for the object.

pub fn content_encoding(&self) -> Option<&str>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(&self) -> Option<&str>

The language the content is in.

pub fn content_type(&self) -> Option<&str>

A standard MIME type describing the format of the object data.

pub fn expires(&self) -> Option<&DateTime>

The date and time at which the object is no longer cacheable.

pub fn grant_full_control(&self) -> Option<&str>

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

pub fn grant_read(&self) -> Option<&str>

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

pub fn grant_read_acp(&self) -> Option<&str>

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

pub fn grant_write_acp(&self) -> Option<&str>

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

pub fn key(&self) -> Option<&str>

Object key for which the multipart upload is to be initiated.

pub fn metadata(&self) -> Option<&HashMap<String, String>>

A map of metadata to store with the object in S3.

pub fn server_side_encryption(&self) -> Option<&ServerSideEncryption>

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn storage_class(&self) -> Option<&StorageClass>

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

pub fn website_redirect_location(&self) -> Option<&str>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

pub fn sse_customer_algorithm(&self) -> Option<&str>

Specifies the algorithm to use to when encrypting the object (for example, AES256).

pub fn sse_customer_key(&self) -> Option<&str>

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

pub fn sse_customer_key_md5(&self) -> Option<&str>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

pub fn ssekms_key_id(&self) -> Option<&str>

Specifies the ID of the symmetric customer managed key to use for object encryption. All GET and PUT requests for an object protected by Amazon Web Services KMS will fail if not made via SSL or using SigV4. For information about configuring using any of the officially supported Amazon Web Services SDKs and Amazon Web Services CLI, see Specifying the Signature Version in Request Authentication in the Amazon S3 User Guide.

pub fn ssekms_encryption_context(&self) -> Option<&str>

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

pub fn bucket_key_enabled(&self) -> bool

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with an object action doesn’t affect bucket-level settings for S3 Bucket Key.

pub fn request_payer(&self) -> Option<&RequestPayer>

Confirms that the requester knows that they will be charged for the request. Bucket owners need not specify this parameter in their requests. For information about downloading objects from requester pays buckets, see Downloading Objects in Requestor Pays Buckets in the Amazon S3 User Guide.

pub fn tagging(&self) -> Option<&str>

The tag-set for the object. The tag-set must be encoded as URL Query parameters.

pub fn object_lock_mode(&self) -> Option<&ObjectLockMode>

Specifies the Object Lock mode that you want to apply to the uploaded object.

pub fn object_lock_retain_until_date(&self) -> Option<&DateTime>

Specifies the date and time when you want the Object Lock to expire.

pub fn object_lock_legal_hold_status( &self ) -> Option<&ObjectLockLegalHoldStatus>

Specifies whether you want to apply a Legal Hold to the uploaded object.

pub fn expected_bucket_owner(&self) -> Option<&str>

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.