Struct aws_sdk_s3::output::head_object_output::Builder

#[non_exhaustive]
pub struct Builder { /* private fields */ }

A builder for HeadObjectOutput

Implementations

impl Builder

pub fn delete_marker(self, input: bool) -> Self

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

pub fn set_delete_marker(self, input: Option<bool>) -> Self

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

pub fn accept_ranges(self, input: impl Into<String>) -> Self

Indicates that a range of bytes was specified.

pub fn set_accept_ranges(self, input: Option<String>) -> Self

Indicates that a range of bytes was specified.

pub fn expiration(self, input: impl Into<String>) -> Self

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

pub fn set_expiration(self, input: Option<String>) -> Self

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL encoded.

pub fn restore(self, input: impl Into<String>) -> Self

If the object is an archived object (an object whose storage class is GLACIER), the response includes this header if either the archive restoration is in progress (see RestoreObject or an archive copy is already restored.

If an archive copy is already restored, the header value indicates when Amazon S3 is scheduled to delete the object copy. For example:

x-amz-restore: ongoing-request="false", expiry-date="Fri, 21 Dec 2012 00:00:00 GMT"

If the object restoration is in progress, the header returns the value ongoing-request="true".

For more information about archiving objects, see Transitioning Objects: General Considerations.

pub fn set_restore(self, input: Option<String>) -> Self

If the object is an archived object (an object whose storage class is GLACIER), the response includes this header if either the archive restoration is in progress (see RestoreObject or an archive copy is already restored.

If an archive copy is already restored, the header value indicates when Amazon S3 is scheduled to delete the object copy. For example:

x-amz-restore: ongoing-request="false", expiry-date="Fri, 21 Dec 2012 00:00:00 GMT"

If the object restoration is in progress, the header returns the value ongoing-request="true".

For more information about archiving objects, see Transitioning Objects: General Considerations.

pub fn archive_status(self, input: ArchiveStatus) -> Self

The archive state of the head object.

pub fn set_archive_status(self, input: Option<ArchiveStatus>) -> Self

The archive state of the head object.

pub fn last_modified(self, input: DateTime) -> Self

Creation date of the object.

pub fn set_last_modified(self, input: Option<DateTime>) -> Self

Creation date of the object.

pub fn content_length(self, input: i64) -> Self

Size of the body in bytes.

pub fn set_content_length(self, input: Option<i64>) -> Self

Size of the body in bytes.

pub fn e_tag(self, input: impl Into<String>) -> Self

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn set_e_tag(self, input: Option<String>) -> Self

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn missing_meta(self, input: i32) -> Self

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn set_missing_meta(self, input: Option<i32>) -> Self

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn version_id(self, input: impl Into<String>) -> Self

Version of the object.

pub fn set_version_id(self, input: Option<String>) -> Self

Version of the object.

pub fn cache_control(self, input: impl Into<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn set_cache_control(self, input: Option<String>) -> Self

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(self, input: impl Into<String>) -> Self

Specifies presentational information for the object.

pub fn set_content_disposition(self, input: Option<String>) -> Self

Specifies presentational information for the object.

pub fn content_encoding(self, input: impl Into<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn set_content_encoding(self, input: Option<String>) -> Self

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(self, input: impl Into<String>) -> Self

The language the content is in.

pub fn set_content_language(self, input: Option<String>) -> Self

The language the content is in.

pub fn content_type(self, input: impl Into<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn set_content_type(self, input: Option<String>) -> Self

A standard MIME type describing the format of the object data.

pub fn expires(self, input: DateTime) -> Self

The date and time at which the object is no longer cacheable.

pub fn set_expires(self, input: Option<DateTime>) -> Self

The date and time at which the object is no longer cacheable.

pub fn website_redirect_location(self, input: impl Into<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

pub fn set_website_redirect_location(self, input: Option<String>) -> Self

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

pub fn server_side_encryption(self, input: ServerSideEncryption) -> Self

If the object is stored using server-side encryption either with an Amazon Web Services KMS key or an Amazon S3-managed encryption key, the response includes this header with the value of the server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn set_server_side_encryption(
self,
input: Option<ServerSideEncryption>
) -> Self

If the object is stored using server-side encryption either with an Amazon Web Services KMS key or an Amazon S3-managed encryption key, the response includes this header with the value of the server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn metadata(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to metadata.

To override the contents of this collection use set_metadata.

A map of metadata to store with the object in S3.

pub fn set_metadata(self, input: Option<HashMap<String, String>>) -> Self

A map of metadata to store with the object in S3.

pub fn sse_customer_algorithm(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

pub fn set_sse_customer_algorithm(self, input: Option<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

pub fn sse_customer_key_md5(self, input: impl Into<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

pub fn set_sse_customer_key_md5(self, input: Option<String>) -> Self

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

pub fn ssekms_key_id(self, input: impl Into<String>) -> Self

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

pub fn set_ssekms_key_id(self, input: Option<String>) -> Self

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

pub fn bucket_key_enabled(self, input: bool) -> Self

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn set_bucket_key_enabled(self, input: Option<bool>) -> Self

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn storage_class(self, input: StorageClass) -> Self

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

pub fn set_storage_class(self, input: Option<StorageClass>) -> Self

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

For more information, see Storage Classes.

pub fn request_charged(self, input: RequestCharged) -> Self

If present, indicates that the requester was successfully charged for the request.

pub fn set_request_charged(self, input: Option<RequestCharged>) -> Self

If present, indicates that the requester was successfully charged for the request.

pub fn replication_status(self, input: ReplicationStatus) -> Self

Amazon S3 can return this header if your request involves a bucket that is either a source or a destination in a replication rule.

In replication, you have a source bucket on which you configure replication and destination bucket or buckets where Amazon S3 stores object replicas. When you request an object (GetObject) or object metadata (HeadObject) from these buckets, Amazon S3 will return the x-amz-replication-status header in the response as follows:

If requesting an object from the source bucket — Amazon S3 will return the x-amz-replication-status header if the object in your request is eligible for replication.

For example, suppose that in your replication configuration, you specify object prefix TaxDocs requesting Amazon S3 to replicate objects with key prefix TaxDocs. Any objects you upload with this key name prefix, for example TaxDocs/document1.pdf, are eligible for replication. For any object request with this key name prefix, Amazon S3 will return the x-amz-replication-status header with value PENDING, COMPLETED or FAILED indicating object replication status.

If requesting an object from a destination bucket — Amazon S3 will return the x-amz-replication-status header with value REPLICA if the object in your request is a replica that Amazon S3 created and there is no replica modification replication in progress.

When replicating objects to multiple destination buckets the x-amz-replication-status header acts differently. The header of the source object will only return a value of COMPLETED when replication is successful to all destinations. The header will remain at value PENDING until replication has completed for all destinations. If one or more destinations fails replication the header will return FAILED.

For more information, see Replication.

pub fn set_replication_status(self, input: Option<ReplicationStatus>) -> Self

Amazon S3 can return this header if your request involves a bucket that is either a source or a destination in a replication rule.

In replication, you have a source bucket on which you configure replication and destination bucket or buckets where Amazon S3 stores object replicas. When you request an object (GetObject) or object metadata (HeadObject) from these buckets, Amazon S3 will return the x-amz-replication-status header in the response as follows:

If requesting an object from the source bucket — Amazon S3 will return the x-amz-replication-status header if the object in your request is eligible for replication.

For example, suppose that in your replication configuration, you specify object prefix TaxDocs requesting Amazon S3 to replicate objects with key prefix TaxDocs. Any objects you upload with this key name prefix, for example TaxDocs/document1.pdf, are eligible for replication. For any object request with this key name prefix, Amazon S3 will return the x-amz-replication-status header with value PENDING, COMPLETED or FAILED indicating object replication status.

If requesting an object from a destination bucket — Amazon S3 will return the x-amz-replication-status header with value REPLICA if the object in your request is a replica that Amazon S3 created and there is no replica modification replication in progress.

When replicating objects to multiple destination buckets the x-amz-replication-status header acts differently. The header of the source object will only return a value of COMPLETED when replication is successful to all destinations. The header will remain at value PENDING until replication has completed for all destinations. If one or more destinations fails replication the header will return FAILED.

For more information, see Replication.

pub fn parts_count(self, input: i32) -> Self

The count of parts this object has.

pub fn set_parts_count(self, input: Option<i32>) -> Self

The count of parts this object has.

pub fn object_lock_mode(self, input: ObjectLockMode) -> Self

The Object Lock mode, if any, that's in effect for this object. This header is only returned if the requester has the s3:GetObjectRetention permission. For more information about S3 Object Lock, see Object Lock.

pub fn set_object_lock_mode(self, input: Option<ObjectLockMode>) -> Self

The Object Lock mode, if any, that's in effect for this object. This header is only returned if the requester has the s3:GetObjectRetention permission. For more information about S3 Object Lock, see Object Lock.

pub fn object_lock_retain_until_date(self, input: DateTime) -> Self

The date and time when the Object Lock retention period expires. This header is only returned if the requester has the s3:GetObjectRetention permission.

pub fn set_object_lock_retain_until_date(self, input: Option<DateTime>) -> Self

The date and time when the Object Lock retention period expires. This header is only returned if the requester has the s3:GetObjectRetention permission.

pub fn object_lock_legal_hold_status(
self,
input: ObjectLockLegalHoldStatus
) -> Self

Specifies whether a legal hold is in effect for this object. This header is only returned if the requester has the s3:GetObjectLegalHold permission. This header is not returned if the specified version of this object has never had a legal hold applied. For more information about S3 Object Lock, see Object Lock.

pub fn set_object_lock_legal_hold_status(
self,
input: Option<ObjectLockLegalHoldStatus>
) -> Self

Specifies whether a legal hold is in effect for this object. This header is only returned if the requester has the s3:GetObjectLegalHold permission. This header is not returned if the specified version of this object has never had a legal hold applied. For more information about S3 Object Lock, see Object Lock.

pub fn build(self) -> HeadObjectOutput

Consumes the builder and constructs a HeadObjectOutput