Struct aws_sdk_s3::output::GetObjectOutput

#[non_exhaustive]
pub struct GetObjectOutput {
    pub body: ByteStream,
    /* private fields */
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

body: ByteStream

Object data.

Implementations

impl GetObjectOutput

pub fn body(&self) -> &ByteStream

Object data.

pub fn delete_marker(&self) -> bool

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

pub fn accept_ranges(&self) -> Option<&str>

Indicates that a range of bytes was specified.

pub fn expiration(&self) -> Option<&str>

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key-value pairs providing object expiration information. The value of the rule-id is URL-encoded.

pub fn restore(&self) -> Option<&str>

Provides information about object restoration action and expiration time of the restored object copy.

pub fn last_modified(&self) -> Option<&DateTime>

Creation date of the object.

pub fn content_length(&self) -> i64

Size of the body in bytes.

pub fn e_tag(&self) -> Option<&str>

An entity tag (ETag) is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

pub fn checksum_crc32(&self) -> Option<&str>

The base64-encoded, 32-bit CRC32 checksum of the object. This will only be present if it was uploaded with the object. With multipart uploads, this may not be a checksum value of the object. For more information about how checksums are calculated with multipart uploads, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_crc32_c(&self) -> Option<&str>

The base64-encoded, 32-bit CRC32C checksum of the object. This will only be present if it was uploaded with the object. With multipart uploads, this may not be a checksum value of the object. For more information about how checksums are calculated with multipart uploads, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_sha1(&self) -> Option<&str>

The base64-encoded, 160-bit SHA-1 digest of the object. This will only be present if it was uploaded with the object. With multipart uploads, this may not be a checksum value of the object. For more information about how checksums are calculated with multipart uploads, see Checking object integrity in the Amazon S3 User Guide.

pub fn checksum_sha256(&self) -> Option<&str>

The base64-encoded, 256-bit SHA-256 digest of the object. This will only be present if it was uploaded with the object. With multipart uploads, this may not be a checksum value of the object. For more information about how checksums are calculated with multipart uploads, see Checking object integrity in the Amazon S3 User Guide.

pub fn missing_meta(&self) -> i32

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

pub fn version_id(&self) -> Option<&str>

Version of the object.

pub fn cache_control(&self) -> Option<&str>

Specifies caching behavior along the request/reply chain.

pub fn content_disposition(&self) -> Option<&str>

Specifies presentational information for the object.

pub fn content_encoding(&self) -> Option<&str>

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

pub fn content_language(&self) -> Option<&str>

The language the content is in.

pub fn content_range(&self) -> Option<&str>

The portion of the object returned in the response.

pub fn content_type(&self) -> Option<&str>

A standard MIME type describing the format of the object data.

pub fn expires(&self) -> Option<&DateTime>

The date and time at which the object is no longer cacheable.

pub fn website_redirect_location(&self) -> Option<&str>

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

pub fn server_side_encryption(&self) -> Option<&ServerSideEncryption>

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

pub fn metadata(&self) -> Option<&HashMap<String, String>>

A map of metadata to store with the object in S3.

pub fn sse_customer_algorithm(&self) -> Option<&str>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

pub fn sse_customer_key_md5(&self) -> Option<&str>

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

pub fn ssekms_key_id(&self) -> Option<&str>

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

pub fn bucket_key_enabled(&self) -> bool

Indicates whether the object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

pub fn storage_class(&self) -> Option<&StorageClass>

Provides storage class information of the object. Amazon S3 returns this header for all objects except for S3 Standard storage class objects.

pub fn request_charged(&self) -> Option<&RequestCharged>

If present, indicates that the requester was successfully charged for the request.

pub fn replication_status(&self) -> Option<&ReplicationStatus>

Amazon S3 can return this if your request involves a bucket that is either a source or destination in a replication rule.

pub fn parts_count(&self) -> i32

The count of parts this object has. This value is only returned if you specify partNumber in your request and the object was uploaded as a multipart upload.

pub fn tag_count(&self) -> i32

The number of tags, if any, on the object.

pub fn object_lock_mode(&self) -> Option<&ObjectLockMode>

The Object Lock mode currently in place for this object.

pub fn object_lock_retain_until_date(&self) -> Option<&DateTime>

The date and time when this object's Object Lock will expire.

pub fn object_lock_legal_hold_status( &self ) -> Option<&ObjectLockLegalHoldStatus>

Indicates whether this object has an active legal hold. This field is only returned if you have permission to view an object's legal hold status.

impl GetObjectOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture GetObjectOutput.